% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_point_forecasts.R
\name{abs_error}
\alias{abs_error}
\title{Absolute Error}
\usage{
abs_error(observed, predicted)
}
\arguments{
\item{observed}{A vector with the observed values of size n}

\item{predicted}{numeric vector with predictions, corresponding to the
quantiles in a second vector, \code{quantiles}.}
}
\value{
vector with the absolute error
}
\description{
Calculate absolute error as

\deqn{
  \textrm{abs}(\textrm{observed} - \textrm{median\_prediction})
}{
  abs(observed - prediction)
}
}
\examples{
observed <- rnorm(30, mean = 1:30)
predicted_values <- rnorm(30, mean = 1:30)
abs_error(observed, predicted_values)
}
\seealso{
\code{\link[=ae_median_sample]{ae_median_sample()}}, \code{\link[=ae_median_quantile]{ae_median_quantile()}}
}
\keyword{metric}
