% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coverage.R
\name{add_coverage_raw_data}
\alias{add_coverage_raw_data}
\alias{add_coverage_raw_data.default}
\alias{add_coverage_raw_data.scoringutils_quantile}
\alias{add_coverage_raw_data.scoringutils_sample}
\title{Add Coverage of Central Prediction Intervals to Forecasts}
\usage{
add_coverage_raw_data(data, by = NULL, ranges = c(50, 90))

\method{add_coverage_raw_data}{default}(data, by = NULL, ranges = c(50, 90))

\method{add_coverage_raw_data}{scoringutils_quantile}(data, by = NULL, ranges = c(50, 90))

\method{add_coverage_raw_data}{scoringutils_sample}(data, by = NULL, ranges = c(50, 90))
}
\arguments{
\item{data}{A data.frame or data.table with predicted and observed values.}

\item{by}{character vector with column names to add the coverage for.}

\item{ranges}{numeric vector of the ranges of the central prediction intervals
for which coverage values shall be added. Ranges should be given as
percentages. For example, \code{ranges = c(50, 90)}
will add coverage values for the 50\% and 90\% central prediction intervals (
corresponding to the 0.05, 0.25, 0.75 and 0.95 quantiles of the predictive
distribution).}
}
\value{
a data.table with with columns added for the
coverage of the central prediction intervals. While the overall data.table
is still unsummarised, note that for the coverage columns some level of
summary is present according to the value specified in \code{by}.
}
\description{
Adds a column with the coverage of central prediction intervals
to a data.table with forecasts either in a quantile or in a sample-based
format (following the input requirements of \code{score()}).

Coverage for a given interval range is defined as the proportion of
observations that fall within the corresponding central prediction intervals.
Central prediction intervals are symmetric around the median and and formed
by two quantiles that denote the lower and upper bound. For example, the 50\%
central prediction interval is the interval between the 0.25 and 0.75
quantiles of the predictive distribution.

The coverage values that are added are computed according to the values
specified in \code{by}. If, for example, \code{by = "model"}, then there will be one
coverage value for every model. If \code{by = c("model", "target_type")}, then
there will be one coverage value for every combination of model and target
type.

\code{add_coverage_raw_data.default()} validates the input data,
checks the forecast type and calls \code{add_coverage_raw_data()} again to
dispatch to the appropriate method.
}
\examples{
library(magrittr) # pipe operator
example_quantile \%>\%
  add_coverage_raw_data(by = c("model", "target_type"))
}
\keyword{scoring}
