% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_scores.R
\name{add_pairwise_comparison}
\alias{add_pairwise_comparison}
\title{Add pairwise comparisons}
\usage{
add_pairwise_comparison(
  scores,
  by = "model",
  metric = intersect(c("wis", "crps", "brier_score"), names(scores)),
  baseline = NULL
)
}
\arguments{
\item{scores}{MORE INFO HERE.}

\item{by}{character vector with column names to summarise scores by. Default
is "model", meaning that there will be one relative skill score per model.}

\item{metric}{character with the name of the metric for which
a relative skill shall be computed.}

\item{baseline}{character string with the name of a model. If a baseline is
given, then a scaled relative skill with respect to the baseline will be
returned. By default (\code{NULL}), relative skill will not be scaled with
respect to a baseline model.}
}
\description{
Adds a columns with relative skills computed by running
pairwise comparisons on the scores.

a column called
'model' must be present in the input data. For more information on
the computation of relative skill, see \code{\link[=pairwise_comparison]{pairwise_comparison()}}.
Relative skill will be calculated for the aggregation level specified in
\code{by}.
WRITE MORE INFO HERE.
}
\keyword{keyword}
\keyword{scoring}
