% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_forecasts_helper.R
\name{add_quantiles}
\alias{add_quantiles}
\title{Add Quantiles to Predictions When Summarising}
\usage{
add_quantiles(dt, quantiles, summarise_by)
}
\arguments{
\item{dt}{the data.table operated on}

\item{quantiles}{the desired quantiles}

\item{varnames}{names of the variables for which to calculate quantiles}

\item{by}{grouping variable in \code{\link[=score]{score()}}}
}
\value{
\code{data.table} with quantiles added
}
\description{
Helper function used within score
}
\keyword{internal}
