% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{apply_rules}
\alias{apply_rules}
\title{Apply A List Of Functions To A Data Table Of Forecasts}
\usage{
apply_rules(data, metrics, ...)
}
\arguments{
\item{data}{A forecast object (a validated data.table with predicted and
observed values, see \code{\link[=as_forecast]{as_forecast()}})}

\item{metrics}{A named list of scoring functions. Names will be used as
column names in the output. See \code{\link[=rules_point]{rules_point()}}, \code{\link[=rules_binary]{rules_binary()}},
\code{\link[=rules_quantile]{rules_quantile()}}, and \code{\link[=rules_sample]{rules_sample()}} for more information on the
default metrics used.}

\item{...}{additional arguments}
}
\value{
A data table with the forecasts and the calculated metrics
}
\description{
This helper function applies scoring rules (stored as a list of
functions) to a data table of forecasts. \code{apply_rules} is used within
\code{score()} to apply all scoring rules to the data.
Scoring rules are wrapped in \code{\link[=run_safely]{run_safely()}} to catch errors and to make
sure that only arguments are passed to the scoring rule that are actually
accepted by it.
}
\keyword{internal}
