% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{arg_present_in_dots}
\alias{arg_present_in_dots}
\title{Check if an argument is passed in to (...)}
\usage{
arg_present_in_dots(argname, ...)
}
\arguments{
\item{argname}{the name of the argument to check}
}
\value{
TRUE or FALSE
}
\description{
Helper function that checks whether an argument was passed to
(...) in a function call.
}
\keyword{internal}
