% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{as_forecast}
\alias{as_forecast}
\alias{as_forecast.default}
\title{Create a \code{forecast} Object}
\usage{
as_forecast(data, ...)

\method{as_forecast}{default}(
  data,
  forecast_unit = NULL,
  observed = NULL,
  predicted = NULL,
  model = NULL,
  quantile_level = NULL,
  sample_id = NULL,
  ...
)
}
\arguments{
\item{data}{A data.frame or data.table with predicted and observed values.}

\item{...}{additional arguments}

\item{forecast_unit}{(optional) Name of the columns in \code{data} (after
any renaming of columns done by \code{as_forecast()}) that denote the unit of a
single forecast. See \code{\link[=get_forecast_unit]{get_forecast_unit()}} for details.
If \code{NULL} (the default), all columns that are not required columns are
assumed to form the unit of a single forecast. If specified, all columns
that are not part of the forecast unit (or required columns) will be removed.}

\item{observed}{(optional) Name of the column in \code{data} that contains the
observed values. This column will be renamed to "observed".}

\item{predicted}{(optional) Name of the column in \code{data} that contains the
predicted values. This column will be renamed to "predicted".}

\item{model}{(optional) Name of the column in \code{data} that contains the names
of the models/forecasters that generated the predicted values.
This column will be renamed to "model".}

\item{quantile_level}{(optional) Name of the column in \code{data} that contains
the quantile level of the predicted values. This column will be renamed to
"quantile_level". Only applicable to quantile-based forecasts.}

\item{sample_id}{(optional) Name of the column in \code{data} that contains the
sample id. This column will be renamed to "sample_id". Only applicable to
sample-based forecasts.}
}
\value{
Depending on the forecast type, an object of class
\code{forecast_binary}, \code{forecast_point}, \code{forecast_sample} or
\code{forecast_quantile}.
}
\description{
Convert a data.frame or similar of forecasts into an object of
class \verb{forecast_*} and validate it.

\code{as_forecast()}
\itemize{
\item allows users to specify the current names of the columns that correspond
to the columns required by \code{scoringutils} (\code{observed}, \code{predicted},
\code{model}, as well \code{quantile_level} for quantile-based forecasts and
\code{sample_id} for sample-based forecasts). \code{as_forecast()} renames the
existing columns.
\item allows users to specify the unit of a single forecast. It removes all
columns that are neither part of the forecast unit nor a required column
(see \code{\link[=set_forecast_unit]{set_forecast_unit()}} for details)
\item Determines the forecast type (binary, point, sample-based or
quantile-based) from the input data (using the function
\code{\link[=get_forecast_type]{get_forecast_type()}}.
\item Constructs a forecast object of the appropriate class
(\code{forecast_binary}, \code{forecast_point}, \code{forecast_sample}, or
\code{forecast_quantile}, using the function \code{\link[=new_forecast]{new_forecast()}}).
\item Calls \code{\link[=validate_forecast]{validate_forecast()}} on the newly created forecast object to
validate it
}
}
\section{Forecast types and input format}{
Various different forecast types / forecast formats are supported. At the
moment, those are
\itemize{
\item point forecasts
\item binary forecasts ("soft binary classification")
\item Probabilistic forecasts in a quantile-based format (a forecast is
represented as a set of predictive quantiles)
\item Probabilistic forecasts in a sample-based format (a forecast is represented
as a set of predictive samples)
}

Forecast types are determined based on the columns present in the input data.

\emph{Point forecasts} require a column \code{observed} of type numeric and a column
\code{predicted} of type numeric.

\emph{Binary forecasts} require a column \code{observed} of type factor with exactly
two levels and a column \code{predicted} of type numeric with probabilities,
corresponding to the probability that \code{observed} is equal to the second
factor level. See details \link[=brier_score]{here} for more information.

\emph{Quantile-based forecasts} require a column \code{observed} of type numeric,
a column \code{predicted} of type numeric, and a column \code{quantile_level} of type
numeric with quantile-levels (between 0 and 1).

\emph{Sample-based forecasts} require a column \code{observed} of type numeric,
a column \code{predicted} of type numeric, and a column \code{sample_id} of type
numeric with sample indices.

For more information see the vignettes and the example data
(\link{example_quantile}, \link{example_continuous}, \link{example_integer},
\code{\link[=example_point]{example_point()}}, and \link{example_binary}).
}

\examples{
as_forecast(example_binary)
as_forecast(
  example_quantile,
  forecast_unit = c("model", "target_type", "target_end_date",
                    "horizon", "location")
)
}
\keyword{check-forecasts}
