% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-forecast-quantile.R
\name{as_forecast_point.forecast_quantile}
\alias{as_forecast_point.forecast_quantile}
\title{Convert a \code{forecast_quantile} to a \code{forecast_point}}
\usage{
\method{as_forecast_point}{forecast_quantile}(data, ...)
}
\arguments{
\item{data}{A \code{forecast_quantile} object (as created by
\code{\link[=as_forecast_quantile]{as_forecast_quantile()}}).}

\item{...}{Unused.}
}
\value{
A \code{forecast} object of class \code{forecast_point}.
}
\description{
When converting a \code{forecast_quantile} object into a \code{forecast_point} object,
the 0.5 quantile is extracted and returned as the point forecast.
}
\seealso{
\code{\link[=as_forecast_point]{as_forecast_point()}}

Other functions to create forecast objects: 
\code{\link{as_forecast_binary}()},
\code{\link{as_forecast_multivariate_sample}()},
\code{\link{as_forecast_nominal}()},
\code{\link{as_forecast_ordinal}()},
\code{\link{as_forecast_point}()},
\code{\link{as_forecast_quantile}()},
\code{\link{as_forecast_quantile.forecast_sample}()},
\code{\link{as_forecast_sample}()}
}
\concept{functions to create forecast objects}
\keyword{as_forecast}
