% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-forecast-sample.R
\name{as_forecast_quantile.forecast_sample}
\alias{as_forecast_quantile.forecast_sample}
\title{Convert a \code{forecast_sample} to a \code{forecast_quantile}}
\usage{
\method{as_forecast_quantile}{forecast_sample}(
  data,
  probs = c(0.05, 0.25, 0.5, 0.75, 0.95),
  type = 7,
  ...
)
}
\arguments{
\item{data}{A \code{forecast_sample} object (as created by
\code{\link[=as_forecast_sample]{as_forecast_sample()}}).}

\item{probs}{A numeric vector of quantile levels for which
quantiles will be computed. Corresponds to the \code{probs} argument in
\code{\link[=quantile]{quantile()}}.}

\item{type}{Type argument passed down to the quantile function. For more
information, see \code{\link[=quantile]{quantile()}}.}

\item{...}{Unused.}
}
\value{
A \code{forecast} object of class \code{forecast_quantile}.
}
\description{
When creating a \code{forecast_quantile} object from a \code{forecast_sample} object,
the quantiles are estimated by computing empirical quantiles from the samples
via \code{\link[=quantile]{quantile()}}. Note that empirical quantiles are a biased estimator for
the true quantiles in particular in the tails of the distribution and
when the number of available samples is low.
}
\seealso{
\code{\link[=as_forecast_quantile]{as_forecast_quantile()}}

Other functions to create forecast objects: 
\code{\link{as_forecast_binary}()},
\code{\link{as_forecast_multivariate_sample}()},
\code{\link{as_forecast_nominal}()},
\code{\link{as_forecast_ordinal}()},
\code{\link{as_forecast_point}()},
\code{\link{as_forecast_point.forecast_quantile}()},
\code{\link{as_forecast_quantile}()},
\code{\link{as_forecast_sample}()}
}
\concept{functions to create forecast objects}
