% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_point.R
\name{as_point.forecast_quantile}
\alias{as_point.forecast_quantile}
\title{Convert a quantile forecast to a point forecast}
\usage{
\method{as_point}{forecast_quantile}(forecast, quantile_level = 0.5, ...)
}
\arguments{
\item{forecast}{The \code{forecast_quantile} object.}

\item{quantile_level}{The desired quantile level of the current forecast
that should become the point forecast. Defaults to 0.5 (median).}

\item{...}{Additional arguments passing inherited from the default method but
unused.}
}
\value{
A \code{forecast_point} object.
}
\description{
This function takes a quantile forecast and converts it to a point forecast
by selecting the forecast corresponding to the specified quantile level.
}
\examples{
as_point(as_forecast(example_quantile))
}
\keyword{check-forecasts}
