% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_point.R
\name{as_point.forecast_sample}
\alias{as_point.forecast_sample}
\title{Convert a sample based forecast to a point forecast}
\usage{
\method{as_point}{forecast_sample}(forecast, fun = median, ...)
}
\arguments{
\item{forecast}{The \code{forecast_sample} object to be converted to a point
forecast.}

\item{fun}{A custom function to summarize the forecast. Defaults to \code{median}.}

\item{...}{Additional arguments passed to \code{fun}..}
}
\value{
A \code{forecast_point} object.
}
\description{
This function converts a forecast object to a point forecast by
summarising the forecast using a custom function (for example \code{mean}).
}
\examples{
sample_forecast <- as_forecast(example_sample_continuous)

as_point(sample_forecast)
as_point(sample_forecast, fun = mean)
}
\keyword{check-forecasts}
