% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_point.R
\name{as_point}
\alias{as_point}
\alias{as_point.default}
\title{Convert a forecast object to a point forecast}
\usage{
as_point(forecast, ...)

\method{as_point}{default}(forecast, ...)
}
\arguments{
\item{forecast}{An object of class \verb{forecast_\{type\}},
representing a forecast.}

\item{...}{Additional arguments to be passed to the specific method.}
}
\value{
The function returns a point forecast object, which is a specific
type of forecast object that represents a single value prediction.
}
\description{
The \code{as_point} function is used to convert a forecast object to a point
forecast. It is a generic function that dispatches the conversion to the
appropriate method based on the class of the input forecast object.
}
\examples{
as_point(as_forecast(example_quantile))
as_point(as_forecast(example_sample_continuous))
}
\keyword{check-forecasts}
