% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_quantile.R
\name{as_quantile}
\alias{as_quantile}
\alias{as_quantile.default}
\title{Convert a forecast object to a quantile object}
\usage{
as_quantile(forecast, ...)

\method{as_quantile}{default}(forecast, ...)
}
\arguments{
\item{forecast}{An object of class \verb{forecast_\{type\}},
representing a forecast.}

\item{...}{Additional arguments to be passed to the specific method.}
}
\value{
A \code{forecast_quantile} object.
}
\description{
This function is used to convert a forecast object to a quantile object.
It dispatches to the appropriate method based on the class of the forecast
object.
}
\examples{
as_quantile(as_forecast(example_sample_continuous))

}
\keyword{check-forecasts}
