% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{as_scores}
\alias{as_scores}
\title{Create An Object Of Class \code{scores} From Data}
\usage{
as_scores(scores, score_names)
}
\arguments{
\item{scores}{A data.table or similar with scores as produced by \code{\link[=score]{score()}}}

\item{score_names}{A character vector with the names of the scores
(i.e. the names of the scoring rules used for scoring)}
}
\value{
An object of class \code{scores}
}
\description{
This convenience function wraps \code{\link[=new_scores]{new_scores()}} and validates
the \code{scores} object.
}
\keyword{internal}
