% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-input-helpers.R
\name{assert_equal_length}
\alias{assert_equal_length}
\title{Check Length}
\usage{
assert_equal_length(..., one_allowed = TRUE, call_levels_up = 2)
}
\arguments{
\item{...}{The variables to check}

\item{one_allowed}{logical, allow arguments of length one that can be
recycled}

\item{call_levels_up}{How many levels to go up when including the function
call in the error message. This is useful when calling \code{assert_equal_length()}
within another checking function.}
}
\value{
Returns NULL invisibly if the assertion was successful and throws an
error otherwise.
}
\description{
Check whether variables all have the same length
}
\keyword{internal}
