% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-forecast.R
\name{assert_forecast_generic}
\alias{assert_forecast_generic}
\title{Validation common to all forecast types}
\usage{
assert_forecast_generic(data, verbose = TRUE)
}
\arguments{
\item{data}{A data.table with forecasts and observed values that should
be validated.}

\item{verbose}{Logical. If \code{FALSE} (default is \code{TRUE}), no messages and
warnings will be created.}
}
\value{
returns the input
}
\description{
The function runs input checks that apply to all input data, regardless of
forecast type. The function
\itemize{
\item asserts that the forecast is a data.table which has columns \code{observed} and
\code{predicted}
\item checks the forecast type and forecast unit
\item checks there are no duplicate forecasts
\item checks that observed values are constant within each forecast unit
\item if appropriate, checks the number of samples / quantiles is the same
for all forecasts.
}
}
\keyword{internal_input_check}
