% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-inputs-scoring-functions.R
\name{assert_input_point}
\alias{assert_input_point}
\title{Assert that inputs are correct for point forecast}
\usage{
assert_input_point(observed, predicted)
}
\arguments{
\item{observed}{Input to be checked. Should be a numeric vector with the
observed values of size n}

\item{predicted}{Input to be checked. Should be a numeric vector with the
predicted values of size n}
}
\value{
Returns NULL invisibly if the check was successful and throws an
error otherwise.
}
\description{
Helper function to assert whether the input is suitable for
scoring.
}
\keyword{check-inputs}
