% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-sample-multivariate.R
\name{assert_input_sample_multiv}
\alias{assert_input_sample_multiv}
\title{Assert that inputs are correct for sample-based forecast}
\usage{
assert_input_sample_multiv(observed, predicted, grouping_id)
}
\arguments{
\item{observed}{Input to be checked. Should be a numeric vector with the
observed values of size n.}

\item{predicted}{Input to be checked. Should be a numeric nxN matrix of
predictive samples, n (number of rows) being the number of data points and
N (number of columns) the number of samples per forecast.
If \code{observed} is just a single number, then predicted values can just be a
vector of size N.}

\item{grouping_id}{Numeric vector of length n with ids indicating the
grouping of predicted values. Conceptually, each row of the \code{predicted}
matrix \emph{could} be seen as a separate (univariate) forecast.
The grouping id then groups several of those forecasts together, treating
them as a single multivariate forecast.}
}
\value{
Returns NULL invisibly if the assertion was successful and throws an
error otherwise.
}
\description{
Function assesses whether the inputs correspond to the requirements for
scoring sample-based forecasts.
}
\keyword{internal_input_check}
