% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_attributes}
\alias{assign_attributes}
\title{Assign attributes to an object from a named list}
\usage{
assign_attributes(object, attribute_list)
}
\arguments{
\item{object}{An object to assign attributes to}

\item{attribute_list}{A named list of attributes to assign to the object.}
}
\value{
The object with new attributes according to the contents of
\code{attribute_list}
}
\description{
Every list item will be made an attribute of the object.
}
\keyword{internal}
