% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-input-helpers.R
\name{ensure_model_column}
\alias{ensure_model_column}
\title{Assure that Data Has a \code{model} Column}
\usage{
ensure_model_column(data)
}
\arguments{
\item{data}{A data.frame or data.table with predicted and observed values.}
}
\value{
The data.table with a column called \code{model}
}
\description{
Check whether the data.table has a column called \code{model}.
If not, a column called \code{model} is added with the value \verb{Unspecified model}.
}
\keyword{internal_input_check}
