% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_forecasts.R
\name{available_forecasts}
\alias{available_forecasts}
\title{Count Number of Available Forecasts}
\usage{
available_forecasts(data, by = NULL, collapse = c("quantile", "sample_id"))
}
\arguments{
\item{data}{A data.frame or data.table with the following columns:
\itemize{
\item \code{observed} - observed values
\item \code{predicted} - predictions, predictive samples or predictive quantiles
\item \code{model} - name of the model or forecaster who made a prediction
}

Depending on the forecast type, one of the following columns may be required:
\itemize{
\item \code{sample_id} - index for the predictive samples in the 'predicted' column
\item \code{quantile}: quantile-level of the corresponding value in \code{predicted}
}

For more information see the vignettes and the example data
(\link{example_quantile}, \link{example_continuous},
\link{example_integer}, \code{\link[=example_point]{example_point()}}, and \link{example_binary}).}

\item{by}{character vector or \code{NULL} (the default) that denotes the
categories over which the number of forecasts should be counted.
By default (\code{by = NULL}) this will be the unit of a single forecast (i.e.
all available columns (apart from a few "protected" columns such as
'predicted' and 'observed') plus "quantile" or "sample_id" where present).}

\item{collapse}{character vector (default is \verb{c("quantile", "sample_id"})
with names of categories for which the number of rows should be collapsed to
one when counting. For example, a single forecast is usually represented by a
set of several quantiles or samples and collapsing these to one makes sure
that a single forecast only gets counted once. Setting \code{collapse = c()}
would mean that all quantiles / samples would be counted as individual
forecasts.}
}
\value{
A data.table with columns as specified in \code{by} and an additional
column "count" with the number of forecasts.
}
\description{
Given a data set with forecasts, count the number of available forecasts
for arbitrary grouping (e.g. the number of forecasts per model, or the
number of forecasts per model and location).
This is useful to determine whether there are any missing forecasts.
}
\examples{
data.table::setDTthreads(1) # only needed to avoid issues on CRAN

available_forecasts(example_quantile,
  by = c("model", "target_type")
)
}
\keyword{check-forecasts}
