% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{available_metrics}
\alias{available_metrics}
\title{Available metrics in scoringutils}
\usage{
available_metrics()
}
\value{
A vector with the name of all available metrics
}
\description{
Available metrics in scoringutils
}
\keyword{info}
