% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{available_rules}
\alias{available_rules}
\title{Available scoring rules in scoringutils}
\usage{
available_rules()
}
\value{
A vector with the name of all available scoring rules
}
\description{
Available scoring rules in scoringutils
}
\keyword{info}
