% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_probabilstic_forecasts.R
\name{bias}
\alias{bias}
\title{Determines bias of an incidence forecast from predictive Monte-Carlo
samples as the proportion of predictive samples greater than the data}
\usage{
bias(true_values, samples)
}
\arguments{
\item{true_values}{A vector with the true observed values of size n}

\item{samples}{nxN matrix of predictive samples, n (number of rows) being
the number of data points and N (number of columns) the
number of Monte Carlo samples}
}
\value{
data frame with bias by date
}
\description{
Determines bias of an incidence forecast from predictive Monte-Carlo
samples as the proportion of predictive samples greater than the data
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
