% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-quantile.R
\name{bias_quantile_single_vector}
\alias{bias_quantile_single_vector}
\title{Compute Bias for a Single Vector of Quantile Predictions}
\usage{
bias_quantile_single_vector(observed, predicted, quantile_level, na.rm)
}
\arguments{
\item{observed}{scalar with the observed value}

\item{predicted}{vector of length N corresponding to the number of quantiles
that holds predictions}

\item{quantile_level}{vector of corresponding size N with the quantile levels
for which predictions were made. If this does not contain the median (0.5)
then the median is imputed as being the mean of the two innermost quantiles.}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
scalar with the quantile bias for a single quantile prediction
}
\description{
Internal function to compute bias for a single observed value,
a vector of predicted values and a vector of quantiles.
}
\keyword{internal}
