% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binary_example_data}
\alias{binary_example_data}
\title{Binary Example Data}
\format{
A data.table with 60 rows and 4 variables:
\describe{
  \item{id}{unique identifier for true observed values}
  \item{true_values}{true observed values}
  \item{model}{name of the model that generated the forecasts}
  \item{predictions}{predicted probability that the corresponding true
  value will be 1}
}
}
\usage{
binary_example_data
}
\description{
A toy dataset for a probability forecast of a binary outcome variable
}
\keyword{datasets}
