% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_probabilistic_forecast_binary.R
\name{Brier_score}
\alias{Brier_score}
\title{Brier Score}
\usage{
Brier_score(true_values, predictions)
}
\arguments{
\item{true_values}{A vector with the true observed values of size n}

\item{predictions}{A vector with probability for true_value = 1}
}
\value{
list
}
\description{
Missing
}
\examples{
NULL

}
