% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-input-helpers.R
\name{check_attribute_conflict}
\alias{check_attribute_conflict}
\title{Check Whether There Is a Conflict Between Data and Attributes}
\usage{
check_attribute_conflict(object, attribute, expected)
}
\arguments{
\item{object}{The object to check}

\item{attribute}{The name of the attribute to check}

\item{expected}{The expected value of the attribute}
}
\value{
Returns TRUE if the check was successful and a string with an
error message otherwise
}
\description{
Check whether there is a conflict between a stored attribute and the
same value as inferred from the data. For example, this could be if
an attribute \code{forecast_unit} is stored, but is different from the
\code{forecast_unit} inferred from the data. The check is successful if
the stored and the inferred value are the same.
}
\keyword{internal_input_check}
