% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-input-helpers.R
\name{check_columns_present}
\alias{check_columns_present}
\title{Check column names are present in a data.frame}
\usage{
check_columns_present(data, columns)
}
\arguments{
\item{data}{A data.frame or similar to be checked}

\item{columns}{A character vector of column names to check}
}
\value{
Returns string with a message with the first issue encountered if
any of the column names are not in data, otherwise returns TRUE
}
\description{
Check column names are present in a data.frame
}
\keyword{check-inputs}
