% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-templates.R
\name{check_data_doc_template}
\alias{check_data_doc_template}
\title{Documentation template for various checks on a data.frame}
\usage{
check_data_doc_template(data, columns)
}
\arguments{
\item{data}{A data.frame or similar to be checked}

\item{columns}{A character vector of column names to check}
}
\description{
Documentation template for various checks on a data.frame
}
