% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-input-helpers.R
\name{check_has_attribute}
\alias{check_has_attribute}
\title{Check whether an attribute is present}
\usage{
check_has_attribute(object, attribute)
}
\arguments{
\item{object}{An object to be checked}

\item{attribute}{name of an attribute to be checked}
}
\value{
Returns TRUE if attribute is there and an error message as
a string otherwise
}
\description{
Checks whether an object has an attribute
}
\keyword{check-inputs}
