% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-check-helpers.R
\name{check_input_binary}
\alias{check_input_binary}
\title{Check input for binary forecast}
\usage{
check_input_binary(true_value, prediction)
}
\arguments{
\item{true_value}{Input to be checked. Should be a factor of length n with
exactly two levels, holding the observed values.
The highest factor level is assumed to be the reference level. This means
that \code{prediction} represents the probability that the observed value is equal
to the highest factor level.}

\item{prediction}{Input to be checked. Prediction should be a vector of
length n, holding probabilities. Values represent the probability that
the corresponding value in \code{true_value} will be equal to the highest
available factor level.}
}
\value{
Returns NULL invisibly if the check was successful and throws an
error otherwise.
}
\description{
Helper function to check whether the input is suitable for
scoring.
}
\keyword{internal}
