% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-check-helpers.R
\name{check_input_binary2}
\alias{check_input_binary2}
\title{Check input for binary forecast}
\usage{
check_input_binary2(true_value, prediction, class)
}
\arguments{
\item{true_value}{Input to be checked. Should be a factor with the true
observed values of size n. Factor levels denote the possible values
\code{true_values} can assume.}

\item{prediction}{Input to be checked. Prediction should be a vector of
probabilities. Values should denote the probability that the corresponding
value in \code{true_value} will be equal to the corresponding value denoted
in \code{class}.}

\item{class}{Input to be checked. Should be a vector of size n that denote
the outcome for which a prediction was made. Values in \code{prediction} then
denote the probability that the \code{true_value} assumes the value in \code{class}.
one of several outcomes.}
}
\value{
Returns NULL invisibly if the check was successful and throws an
error otherwise.
}
\description{
Helper function to check whether the input is suitable for
scoring.
}
\keyword{internal}
