% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-inputs-scoring-functions.R
\name{check_input_nominal}
\alias{check_input_nominal}
\title{Check That Inputs Are Correct For Categorical Forecasts}
\usage{
check_input_nominal(observed, predicted, predicted_label)
}
\arguments{
\item{observed}{XXX}

\item{predicted}{XXX}

\item{predicted_label}{XXX}
}
\value{
Returns TRUE if the check was successful and a string with an
error message otherwise.
}
\description{
Function assesses whether the inputs correspond to the
requirements for scoring nominal forecasts.
}
\keyword{internal_input_check}
