% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-point.R
\name{check_input_point}
\alias{check_input_point}
\title{Check that inputs are correct for point forecast}
\usage{
check_input_point(observed, predicted)
}
\arguments{
\item{observed}{Input to be checked. Should be a numeric vector with the
observed values of size n.}

\item{predicted}{Input to be checked. Should be a numeric vector with the
predicted values of size n.}
}
\value{
Returns TRUE if the check was successful and a string with an
error message otherwise.
}
\description{
Function assesses whether the inputs correspond to the
requirements for scoring point forecasts.
}
\keyword{internal_input_check}
