% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-check-helpers.R
\name{check_input_quantile}
\alias{check_input_quantile}
\title{Check input for sample-based forecast}
\usage{
check_input_quantile(observed, predicted, quantile)
}
\arguments{
\item{observed}{Input to be checked. Should be a vector with the true
observed values of size n}

\item{predicted}{Input to be checked. Should be nxN matrix of predictive
quantiles, n (number of rows) being the number of data points and N
(number of columns) the number of quantiles per forecast.
If \code{observed} is just a single number, then predicted can just be a
vector of size N.}

\item{quantile}{Input to be checked. Should be a vector of size N that
denotes the quantile levels corresponding to the columns of the prediction
matrix.}
}
\value{
Returns NULL invisibly if the check was successful and throws an
error otherwise.
}
\description{
Helper function to check whether the input is suitable for
scoring.
}
\keyword{internal}
