% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-check-helpers.R
\name{check_input_sample}
\alias{check_input_sample}
\title{Check input for sample-based forecast}
\usage{
check_input_sample(true_value, prediction)
}
\arguments{
\item{true_value}{Input to be checked. Should be a vector with the true
observed values of size n}

\item{predictions}{Input to be checked. Should be nxN matrix of predictive
samples, n (number of rows) being the number of data points and N (number of
columns) the number of samples per forecast.
If \code{true_value} is just a single number, then predictions can just be a
vector of size N.}
}
\value{
Returns NULL invisibly if the check was successful and throws an
error otherwise.
}
\description{
Helper function to check whether the input is suitable for
scoring.
}
\keyword{internal}
