% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-check-helpers.R
\name{check_metrics}
\alias{check_metrics}
\title{Check whether the desired metrics are available in scoringutils}
\usage{
check_metrics(metrics)
}
\arguments{
\item{metrics}{character vector with desired metrics}
}
\value{
A character vector with metrics that can be used for downstream
computation
}
\description{
Helper function to check whether desired metrics are
available. If the input is \code{NULL}, all metrics will be returned.
}
\keyword{internal}
