% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-non-monotonic-forecasts.R
\name{check_monotonicity}
\alias{check_monotonicity}
\title{Check that predictions are monotonically non-decreasing with quantile level}
\usage{
check_monotonicity(data)
}
\arguments{
\item{data}{A data.frame (or similar) with predicted and observed values.
See the details section of for additional information
on the required input format.}
}
\value{
Returns TRUE if the check was successful and a string with an
error message otherwise.
}
\description{
Runs \code{\link[=get_non_monotonic_forecasts]{get_non_monotonic_forecasts()}} and returns a message if an issue is
encountered
}
\keyword{internal_input_check}
