% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-input-helpers.R
\name{check_no_NA_present}
\alias{check_no_NA_present}
\title{Check columns in data.frame don't have NA values}
\usage{
check_no_NA_present(data, columns)
}
\arguments{
\item{data}{A data.frame or similar to be checked}

\item{columns}{A character vector of column names to check}
}
\value{
Returns an string with a message if any of the column names
have NA values, otherwise returns TRUE
}
\description{
Check columns in data.frame don't have NA values
}
\keyword{internal}
