% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-input-helpers.R
\name{check_not_all_NA}
\alias{check_not_all_NA}
\title{Check Whether All Elements Are NA}
\usage{
check_not_all_NA(x)
}
\arguments{
\item{x}{vector to check}
}
\value{
Returns TRUE if the check was successful and a string with an
error message otherwise
}
\description{
Check Whether All Elements Are NA
}
