% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-input-helpers.R
\name{check_number_per_forecast}
\alias{check_number_per_forecast}
\title{Check that all forecasts have the same number of quantiles or samples}
\usage{
check_number_per_forecast(data, forecast_unit)
}
\arguments{
\item{data}{data.frame to check}

\item{forecast_unit}{Character vector denoting the unit of a single forecast.}
}
\value{
Returns an string with a message if any forecasts have differing
numbers of samples or quantiles, otherwise returns TRUE
}
\description{
Check that all forecasts have the same number of quantiles or samples
}
\keyword{internal}
