% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-forecast.R
\name{check_observed_constant}
\alias{check_observed_constant}
\title{Check that observed values are constant within each forecast unit}
\usage{
check_observed_constant(data, forecast_unit)
}
\arguments{
\item{data}{A data.frame or similar to be checked}

\item{forecast_unit}{Character vector denoting the unit of a single forecast.}
}
\value{
Returns TRUE if the check was successful and a string with an
error message otherwise.
}
\description{
Helper function that checks that all rows belonging to the same forecast
(as defined by the forecast unit) share a single observed value. Rows
where \code{observed} is \code{NA} are ignored.
If the observed values are constant within each forecast unit, the
function returns \code{TRUE} and a string with an error message otherwise.
}
\keyword{internal_input_check}
