% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_helper.R
\name{check_score_params}
\alias{check_score_params}
\title{Check input parameters for \code{\link[=score]{score()}}}
\usage{
check_score_params(data, forecast_unit, metrics, summarise_by)
}
\arguments{
\item{data}{A data.frame or data.table with the predictions and observations.
The following columns need to be present:
\itemize{
\item \code{true_value} - the true observed values
\item \code{prediction} - predictions or predictive samples for one
true value. (You only don't need to provide a prediction column if
you want to score quantile forecasts in a wide range format.)}
For integer and continuous forecasts a \code{sample} column is needed:
\itemize{
\item \code{sample} - an index to identify the predictive samples in the
prediction column generated by one model for one true value. Only
necessary for continuous and integer forecasts, not for
binary predictions.}
For a quantile-format forecast you should provide a column called \code{quantile}:
\itemize{
\item \code{quantile}: quantile to which the prediction corresponds
}}

\item{metrics}{the metrics you want to have in the output. If \code{NULL} (the
default), all available metrics will be computed. For a list of available
metrics see \code{\link[=available_metrics]{available_metrics()}}}
}
\description{
A helper function to check the input parameters for
\code{\link[=score]{score()}}.
}
\keyword{internal}
