% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-check-helpers.R
\name{check_true_values}
\alias{check_true_values}
\title{Check Observed Value Input For Lower-level Scoring Functions}
\usage{
check_true_values(true_values, type = c("continuous", "integer", "binary"))
}
\arguments{
\item{true_values}{A vector with the true observed values of size n}

\item{type}{character, one of "continuous" (default), "integer" or "binary" that
defines the type of the forecast}
}
\description{
Helper function to check inputs for lower-level score functions.
}
\keyword{internal}
