% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-input-helpers.R
\name{check_try}
\alias{check_try}
\title{Helper function to convert assert statements into checks}
\usage{
check_try(expr)
}
\arguments{
\item{expr}{an expression to be evaluated}
}
\value{
Returns TRUE if expression was executed successfully, otherwise
returns a string with the resulting error message
}
\description{
Tries to execute an expression. Internally, this is used to
see whether assertions fail when checking inputs
}
\keyword{internal}
