% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{clean_forecast}
\alias{clean_forecast}
\title{Clean forecast object}
\usage{
clean_forecast(forecast, copy = FALSE, na.omit = FALSE)
}
\arguments{
\item{forecast}{A forecast object (a validated data.table with predicted and
observed values, see \code{\link[=as_forecast]{as_forecast()}})}

\item{copy}{Logical, default is \code{FALSE}. If \code{TRUE}, a copy of the input data
is created.}

\item{na.omit}{Logical, default is \code{FALSE}. If \code{TRUE}, rows with missing
values are removed.}
}
\description{
The function makes it possible to silently validate an object. In addition,
it can return a copy of the data and remove rows with missing values.
}
\keyword{internal}
