% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{coord_cartesian_panels}
\alias{coord_cartesian_panels}
\title{Cartesian coordinates per facet-panel}
\usage{
coord_cartesian_panels(
  panel_limits,
  expand = TRUE,
  default = FALSE,
  clip = "on"
)
}
\arguments{
\item{panel_limits}{'data.frame' with faceting variables and
limiting variables, see 'Details'}

\item{expand, default, clip}{as defined/used in
\code{\link[ggplot2:coord_cartesian]{ggplot2::coord_cartesian()}}}
}
\description{
This function mimics the behavior of \code{\link[ggplot2:coord_cartesian]{ggplot2::coord_cartesian()}},
while supporting per-panel limits when faceted. It has been taken verbatim
from https://gist.github.com/r2evans/6057f7995c117bb787495dc14a228d5d
}
\details{
A 'panel_limits' data frame may contain:
\itemize{
\item zero or more faceting variables, all of which must be found
within the grob's 'layout' (i.e., defined by
\code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}} or \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}};
\item zero or more of 'xmin', 'xmax', 'ymin', and 'ymax', where missing
columns and 'NA' values within columns will default to ggplot2's
normal min/max determination;
\item each panel in the plot must match no more than one row in
'panel_limits';
\item each row may match more than one panel, such as when some
faceting variables are not included (in 'panel_limits');
\item if no faceting variables are included, then 'panel_limits' must
be at most one row (in which case it effectively falls back to
\code{\link[ggplot2:coord_cartesian]{ggplot2::coord_cartesian()}} behavior).
}

It is an error if:
\itemize{
\item a panel is matched by more than one row (no matches is okay);
\item a faceting variable in 'panel_limits' is not found within the
faceted layout.
}
}
\section{Thanks}{

\itemize{
\item burchill (github) and the original version;
https://gist.github.com/burchill/d780d3e8663ad15bcbda7869394a348a
\item Z.Lin (stackoverflow) for helping me through some of the
initial errors; https://stackoverflow.com/a/63556918
\item teunbrand (github and stackoverflow), possible future extension
of the non-list-index version; https://github.com/teunbrand/ggh4x
}
}

\examples{
\dontrun{

library(dplyr)
library(tidyr)
library(ggplot2)

testdata <- tibble(
  x = rep(1:100, 2),
  y = rep(sin(seq(0,2*pi,length.out=100)), 2)
) \%>\%
  mutate(y1 = y - 0.3, y2 = y + 0.3) \%>\%
  tidyr::crossing(
    tidyr::expand_grid(facet1 = c("aa", "bb"), facet2 = c("11", "22"))
  )

gg <- ggplot(testdata, aes(x, y)) +
  geom_ribbon(aes(ymin = y1, ymax = y2), fill = "#ff8888aa") +
  geom_path(color = "red", size = 1) +
  facet_wrap(facet1 + facet2 ~ ., scales = "free")
gg

# single-panel change,
gg + coord_cartesian_panels(
  panel_limits = tribble::tribble(
    ~facet1, ~facet2, ~ymin, ~ymax
  , "aa"   , "22"   , -0.75,   0.5
  )
)
}

}
