% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{correlation}
\alias{correlation}
\title{Correlation Between Metrics}
\usage{
correlation(scores, metrics = NULL)
}
\arguments{
\item{scores}{A data.table of scores as produced by \code{\link[=score]{score()}}.}

\item{metrics}{A character vector with the metrics to show. If set to
\code{NULL} (default), all metrics present in \code{scores} will
be shown}
}
\value{
A data.table with correlations for the different metrics. In addition,
the output is of class \code{scoringutils_correlation} and can e.g. be visualised
using \code{\link[=plot]{plot()}} (which dispatches \code{\link[=plot.scoringutils_correlation]{plot.scoringutils_correlation()}}.
}
\description{
Calculate the correlation between different metrics for a data.frame of
scores as produced by \code{\link[=score]{score()}}. You can visualise results
}
\examples{
scores <- score(example_quantile)
corrs <- correlation(scores)
plot(corrs)
}
\seealso{
\code{\link[=plot.scoringutils_correlation]{plot.scoringutils_correlation()}}
}
\keyword{scoring}
