% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-sample.R
\name{crps_sample}
\alias{crps_sample}
\title{(Continuous) ranked probability score}
\usage{
crps_sample(observed, predicted, ...)
}
\arguments{
\item{observed}{A vector with observed values of size n}

\item{predicted}{nxN matrix of predictive samples, n (number of rows) being
the number of data points and N (number of columns) the number of Monte
Carlo samples. Alternatively, \code{predicted} can just be a vector of size n.}

\item{...}{Additional arguments passed to
\link[scoringRules:scores_sample_univ]{crps_sample()} from the scoringRules package.}
}
\value{
Vector with scores.
}
\description{
Wrapper around the \code{\link[scoringRules:scores_sample_univ]{crps_sample()}}
function from the
\pkg{scoringRules} package. Can be used for continuous as well as integer
valued forecasts
}
\examples{
observed <- rpois(30, lambda = 1:30)
predicted <- replicate(200, rpois(n = 30, lambda = 1:30))
crps_sample(observed, predicted)
}
\references{
Alexander Jordan, Fabian Krüger, Sebastian Lerch, Evaluating Probabilistic
Forecasts with scoringRules, \url{https://www.jstatsoft.org/article/view/v090i12}
}
\keyword{metric}
