% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-forecast-sample-multivariate.R
\name{define_grouping_cols}
\alias{define_grouping_cols}
\title{Return column names to define the grouping for multivariate forecasts}
\usage{
define_grouping_cols(data, across)
}
\arguments{
\item{data}{A data.frame (or similar) with predicted and observed values.
See the details section of for additional information
on the required input format.}

\item{across}{Character vector with the names of the columns over which to
group.}
}
\value{
A character vector with the names of the columns that define the grouping.
}
\description{
Return column names to define the grouping for multivariate forecasts
}
