% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-templates.R
\name{document_check_functions}
\alias{document_check_functions}
\title{Documentation template for check functions}
\value{
Returns TRUE if the check was successful and a string with an
error message otherwise
}
\description{
Documentation template for check functions
}
