% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-templates.R
\name{document_test_functions}
\alias{document_test_functions}
\title{Documentation template for test functions}
\value{
Returns TRUE if the check was successful and FALSE otherwise
}
\description{
Documentation template for test functions
}
\keyword{internal}
