% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_-functions.R
\name{.get_forecast_type}
\alias{.get_forecast_type}
\title{Get forecast type from forecast object
Internal function to obtain the forecast type as a string.}
\usage{
.get_forecast_type(forecast)
}
\arguments{
\item{forecast}{A forecast object (a validated data.table with predicted and
observed values, see \code{\link[=as_forecast]{as_forecast()}})}
}
\value{
Character vector of length one with the forecast type.
}
\description{
Get forecast type from forecast object
Internal function to obtain the forecast type as a string.
}
\keyword{internal_input_check}
