% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_probabilstic_forecasts_integer.R
\name{dss}
\alias{dss}
\title{Dawid-Sebastiani Score}
\usage{
dss(true_values, predictions)
}
\arguments{
\item{true_values}{A vector with the true observed values of size n}

\item{predictions}{nxN matrix of predictive samples, n (number of rows) being
the number of data points and N (number of columns) the
number of Monte Carlo samples}
}
\value{
data frame with DSS by date
}
\description{
Wrapper around the \code{\link[scoringRules]{dss_sample}} function from the
\code{scoringRules} package.
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
