% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_data.table}
\alias{ensure_data.table}
\title{Ensure That an Object is a Data Table}
\usage{
ensure_data.table(data)
}
\arguments{
\item{data}{An object to ensure is a data table}
}
\value{
A data table
}
\description{
This function ensures that an object is a data table.
If the object is not a data table, it is converted to one. If the object
is a data table, a copy of the object is returned.
}
\keyword{internal}
