% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-input-helpers.R
\name{ensure_model_column}
\alias{ensure_model_column}
\title{Assure that data has a \code{model} column}
\usage{
ensure_model_column(data)
}
\arguments{
\item{data}{A data.frame (or similar) with predicted and observed values.
See the details section of \code{\link[=as_forecast]{as_forecast()}} for additional information
on required input formats.}
}
\value{
The data.table with a column called \code{model}
}
\description{
Check whether the data.table has a column called \code{model}.
If not, a column called \code{model} is added with the value \verb{Unspecified model}.
}
\keyword{internal_input_check}
