% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_probabilstic_forecasts_integer.R
\name{error_handling_prob_int}
\alias{error_handling_prob_int}
\title{Error Handling Probabilistic Integer Forecasts}
\usage{
error_handling_prob_int(true_values, predictions)
}
\arguments{
\item{true_values}{A vector with the true observed values of size n}

\item{predictions}{nxN matrix of predictive samples, n (number of rows) being
the number of data points and N (number of columns) the
number of Monte Carlo samples}
}
\value{
updated predictions
}
\description{
Does internal error handling
}
\author{
Nikos Bosse \email{nikosbosse@gmail.com}
}
