% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_forecasts.R
\name{eval_forecasts}
\alias{eval_forecasts}
\title{Applies forecast assessments and scores to the forecasts}
\usage{
eval_forecasts(
  true_values,
  predictions,
  prediction_type = "probabilistic",
  outcome_type = "integer",
  metrics = c()
)
}
\arguments{
\item{true_values}{A vector with the true observed values of size n}

\item{predictions}{a list of nxN matrices of predictive samples, n (number of rows) being
the number of data points and N (number of columns) the
number of Monte Carlo samples}
}
\value{
list of data frames with the forecast scores
}
\description{
Applies forecast assessments and scores to the forecasts
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
