% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_forecasts.R
\name{eval_forecasts_point_bin}
\alias{eval_forecasts_point_bin}
\title{Applies forecast assessments and scores to point predictinos of
binary outcomes}
\usage{
eval_forecasts_point_bin(
  true_values,
  predictions,
  metrics = c(),
  output = "df"
)
}
\arguments{
\item{true_values}{A vector with the true observed values of size n}

\item{predictions}{a list of vectors of predicted values of size n that
correspond to the elements in true_values. This should be a vector containing
only zeros and ones. If you want to make predictions using a probability,
use \code{\link{eval_forecasts_prob_bin}}}

\item{metrics}{what metrics to include. currently not used}

\item{output}{"df" returns a data.frame, everything else returns a list.}
}
\value{
either a data.frame or list of data frames with the forecast scores
}
\description{
Applies forecast assessments and scores to point predictinos of
binary outcomes
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
