% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_forecasts.R
\name{eval_forecasts_prob_cont}
\alias{eval_forecasts_prob_cont}
\title{Applies forecast assessments and scores to probabilistic
forecasts of continuous outcomes}
\usage{
eval_forecasts_prob_cont(
  true_values,
  predictions,
  metrics = c(),
  output = "df"
)
}
\arguments{
\item{true_values}{A vector with the true observed values of size n}

\item{predictions}{a list of nxN matrices of predictive samples,
n (number of rows) being
the number of data points and N (number of columns) the
number of Monte Carlo samples}

\item{metrics}{what metrics to include. currently not used}

\item{output}{"df" returns a data.frame, everything else returns a list.}
}
\value{
either a data.frame or list of data frames with the forecast scores
}
\description{
Applies forecast assessments and scores to probabilistic
forecasts of continuous outcomes
}
