% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_continuous}
\alias{example_continuous}
\title{Continuous Forecast Example Data}
\format{
A data frame with 13,429 rows and 10 columns:
\describe{
\item{location}{the country for which a prediction was made}
\item{target_end_date}{the date for which a prediction was made}
\item{target_type}{the target to be predicted (cases or deaths)}
\item{true_value}{true observed values}
\item{location_name}{name of the country for which a prediction was made}
\item{forecast_date}{the date on which a prediction was made}
\item{model}{name of the model that generated the forecasts}
\item{horizon}{forecast horizon in weeks}
\item{prediction}{predicted value}
\item{sample}{id for the corresponding sample}
}
}
\source{
\url{https://github.com/covid19-forecast-hub-europe/covid19-forecast-hub-europe/commit/a42867b1ea152c57e25b04f9faa26cfd4bfd8fa6/}
}
\usage{
example_continuous
}
\description{
A data set with continuous predictions for COVID-19 cases and deaths
constructed from data submitted to the European Forecast Hub.
}
\details{
The data was created using the script create-example-data.R in the inst/
folder (or the top level folder in a compiled package).
}
\keyword{datasets}
