% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-forecast-point.R
\docType{data}
\name{example_point}
\alias{example_point}
\title{Point forecast example data}
\format{
An object of class \code{forecast_point} (see \code{\link[=as_forecast_point]{as_forecast_point()}})
with the following columns:
\describe{
\item{location}{the country for which a prediction was made}
\item{target_end_date}{the date for which a prediction was made}
\item{target_type}{the target to be predicted (cases or deaths)}
\item{observed}{observed values}
\item{location_name}{name of the country for which a prediction was made}
\item{forecast_date}{the date on which a prediction was made}
\item{predicted}{predicted value}
\item{model}{name of the model that generated the forecasts}
\item{horizon}{forecast horizon in weeks}
}
}
\source{
\url{https://github.com/european-modelling-hubs/covid19-forecast-hub-europe_archive/commit/a42867b1ea152c57e25b04f9faa26cfd4bfd8fa6/}
}
\usage{
example_point
}
\description{
A data set with predictions for COVID-19 cases and deaths submitted to the
European Forecast Hub. This data set is like the quantile example data, only
that the median has been replaced by a point forecast.
}
\details{
The data was created using the script create-example-data.R in the inst/
folder (or the top level folder in a compiled package).
}
\keyword{datasets}
