% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_quantile_forecasts_only}
\alias{example_quantile_forecasts_only}
\title{Quantile Example Data - Forecasts only}
\format{
A data frame with 7,581 rows and 9 columns:
\describe{
  \item{value_date}{the date for which a prediction was made}
  \item{value_type}{the target to be predicted (short form)}
  \item{geography}{the region for which a prediction was made}
  \item{model}{name of the model that generated the forecasts}
  \item{creation_date}{date on which the forecast was made}
  \item{quantile}{quantile of the corresponding prediction}
  \item{prediction}{quantile predictions}
  \item{value_desc}{long form description of the prediction target}
  \item{horizon}{forecast horizon in days}

}
}
\usage{
example_quantile_forecasts_only
}
\description{
A data set with predictions for different quantities relevant in the
2020 UK Covid-19 epidemic, but no true_values
}
\keyword{datasets}
