% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_range_long}
\alias{example_range_long}
\title{Range Forecast Example Data (Long Format)}
\format{
A data frame with:
\describe{
\item{location}{the country for which a prediction was made}
\item{target_end_date}{the date for which a prediction was made}
\item{target_type}{the target to be predicted (cases or deaths)}
\item{true_value}{true observed values}
\item{location_name}{name of the country for which a prediction was made}
\item{forecast_date}{the date on which a prediction was made}
\item{prediction}{predicted value}
\item{model}{name of the model that generated the forecasts}
\item{horizon}{forecast horizon in weeks}
\item{boundary}{indicate lower or upper bound of prediction interval}
\item{range}{range of the corresponding prediction interval}
}
}
\usage{
example_range_long
}
\description{
A data set with predictions for COVID-19 cases and deaths submitted to the European Forecast Hub.
}
\keyword{datasets}
