% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_range_semi_wide}
\alias{example_range_semi_wide}
\title{Range Forecast Example Data (Semi-Wide Format)}
\format{
A data frame with 5,419 rows and 12 columns:
\describe{
\item{location}{the country for which a prediction was made}
\item{target_end_date}{the date for which a prediction was made}
\item{target_type}{the target to be predicted (cases or deaths)}
\item{true_value}{true observed values}
\item{location_name}{name of the country for which a prediction was made}
\item{forecast_date}{the date on which a prediction was made}
\item{model}{name of the model that generated the forecasts}
\item{horizon}{forecast horizon in weeks}
\item{range}{range of the corresponding prediction interval}
\item{lower}{prediction for the lower bound of the corresponding interval}
\item{upper}{prediction for the upper bound of the corresponding interval}
}
}
\usage{
example_range_semi_wide
}
\description{
A data set with predictions for COVID-19 cases and deaths submitted to the European Forecast Hub.
}
\keyword{datasets}
