% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_range_wide}
\alias{example_range_wide}
\title{Range Forecast Example Data (Wide Format)}
\format{
A data frame with:
\describe{
\item{location}{the country for which a prediction was made}
\item{target_end_date}{the date for which a prediction was made}
\item{target_type}{the target to be predicted (cases or deaths)}
\item{true_value}{true observed values}
\item{location_name}{name of the country for which a prediction was made}
\item{forecast_date}{the date on which a prediction was made}
\item{model}{name of the model that generated the forecasts}
\item{horizon}{forecast horizon in weeks}
\item{lower_0}{prediction for the lower bound of the 0\% interval range (median)}
\item{lower_10}{prediction for the lower bound of the 10\% interval range}
\item{lower_20}{prediction for the lower bound of the 20\% interval range}
\item{lower_30}{prediction for the lower bound of the 30\% interval range}
\item{lower_40}{prediction for the lower bound of the 40\% interval range}
\item{lower_50}{prediction for the lower bound of the 50\% interval range}
\item{lower_60}{prediction for the lower bound of the 60\% interval range}
\item{lower_70}{prediction for the lower bound of the 70\% interval range}
\item{lower_80}{prediction for the lower bound of the 80\% interval range}
\item{lower_90}{prediction for the lower bound of the 90\% interval range}
\item{lower_95}{prediction for the lower bound of the 95\% interval range}
\item{lower_98}{prediction for the lower bound of the 98\% interval range}
\item{upper_0}{prediction for the upper bound of the 0\% interval range}
\item{upper_10}{prediction for the upper bound of the 1\% interval range}
\item{upper_20}{prediction for the upper bound of the 20\% interval range}
\item{upper_30}{prediction for the upper bound of the 30\% interval range}
\item{upper_40}{prediction for the upper bound of the 40\% interval range}
\item{upper_50}{prediction for the upper bound of the 50\% interval range}
\item{upper_60}{prediction for the upper bound of the 60\% interval range}
\item{upper_70}{prediction for the upper bound of the 70\% interval range}
\item{upper_80}{prediction for the upper bound of the 80\% interval range}
\item{upper_90}{prediction for the upper bound of the 90\% interval range}
\item{upper_95}{prediction for the upper bound of the 95\% interval range}
\item{upper_98}{prediction for the upper bound of the 98\% interval range}
}
}
\usage{
example_range_wide
}
\description{
A data set with predictions for COVID-19 cases and deaths submitted to the European Forecast Hub.
}
\keyword{datasets}
