% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-forecast-sample.R
\docType{data}
\name{example_sample_continuous}
\alias{example_sample_continuous}
\title{Continuous forecast example data}
\format{
An object of class \code{forecast_sample} (see \code{\link[=as_forecast_sample]{as_forecast_sample()}})
with the following columns:
\describe{
\item{location}{the country for which a prediction was made}
\item{target_end_date}{the date for which a prediction was made}
\item{target_type}{the target to be predicted (cases or deaths)}
\item{observed}{observed values}
\item{location_name}{name of the country for which a prediction was made}
\item{forecast_date}{the date on which a prediction was made}
\item{model}{name of the model that generated the forecasts}
\item{horizon}{forecast horizon in weeks}
\item{predicted}{predicted value}
\item{sample_id}{id for the corresponding sample}
}
}
\source{
\url{https://github.com/european-modelling-hubs/covid19-forecast-hub-europe_archive/commit/a42867b1ea152c57e25b04f9faa26cfd4bfd8fa6/}
}
\usage{
example_sample_continuous
}
\description{
A data set with continuous predictions for COVID-19 cases and deaths
constructed from data submitted to the European Forecast Hub.
}
\details{
The data was created using the script create-example-data.R in the inst/
folder (or the top level folder in a compiled package).
}
\keyword{datasets}
