% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_truth_data_only}
\alias{example_truth_data_only}
\title{Truth data only}
\format{
A data frame with 140 rows and 5 columns:
\describe{
  \item{value_date}{the date for which a prediction was made}
  \item{value_type}{the target to be predicted (short form)}
  \item{geography}{the region for which a prediction was made}
  \item{value_desc}{long form description of the prediction target}
  \item{true_value}{true observed values}

}
}
\usage{
example_truth_data_only
}
\description{
A data set with truth data for different quantities relevant in the
2020 UK Covid-19 epidemic, but no predictions
}
\keyword{datasets}
