% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_truth_only}
\alias{example_truth_only}
\title{Truth data only}
\format{
A data frame with 140 rows and 5 columns:
\describe{
\item{location}{the country for which a prediction was made}
\item{target_end_date}{the date for which a prediction was made}
\item{target_type}{the target to be predicted (cases or deaths)}
\item{true_value}{true observed values}
\item{location_name}{name of the country for which a prediction was made}
}
}
\usage{
example_truth_only
}
\description{
A data set with truth values for COVID-19 cases and deaths
submitted to the European Forecast Hub.
}
\keyword{datasets}
