% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_from_list}
\alias{extract_from_list}
\title{Extract Elements From a List of Lists}
\usage{
extract_from_list(list, what)
}
\arguments{
\item{list}{the list of lists}

\item{what}{character with the name of the element to extract from every
individual list element of `list`}
}
\value{
A list with the extracted element from every sublist
missing.
}
\description{
Extract corresponding elements from a list of lists.
}
