% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{filter_data}
\alias{filter_data}
\title{Filter Data Used for Plotting}
\usage{
filter_data(data = NULL, what = c("truth", "forecast", "both"), ...)
}
\arguments{
\item{data}{a data.frame that follows the same specifications outlined in
\code{\link[=score]{score()}}.}

\item{what}{character vector that determines which values should be turned
into \code{NA}. If \code{what = truth}, values in the column 'true_value' will be
turned into \code{NA}. If \code{what = forecast}, values in the column 'prediction'
will be turned into \code{NA}. If \code{what = both}, values in both column will be
turned into \code{NA}.}

\item{...}{logical statements used to filter the data}
}
\value{
A data.table
}
\description{
Filter data before it gets passed to \code{\link[=plot_predictions]{plot_predictions()}} using arbitrary
filter statements. In contrast to the usual behaviour of filtering functions,
rows do net get removed, but instead corresponding values of either
'true_value' or 'predction' will be replaced with \code{NA}. This allows to filter
truth and prediction data independently (either turning 'true_value' or
'prediction' or both into \code{NA}).
}
\examples{
filter_data (
    example_continuous,
    what = "truth",
    target_end_date <= "2021-07-22",
    target_end_date > "2021-05-01"
  )
}
