% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_function_args}
\alias{filter_function_args}
\title{Filter function arguments}
\usage{
filter_function_args(fun, args)
}
\arguments{
\item{fun}{A function to which arguments shall be passed}

\item{args}{A list of arguments that shall be passed to fun}
}
\value{
A list of function arguments (a subset of \code{args}) that \code{fun} can
accept.
}
\description{
This function compares a list of arguments with the arguments
that a function can accept. It only returns those arguments that can be
passed to the function.

The function is used in \code{\link[=score]{score()}} to handle additional arguments passed to
\code{\link[=score]{score()}} that get then passed along to the different scoring functions.
}
\keyword{internal}
