% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-duplicate-forecasts.R
\name{find_duplicates}
\alias{find_duplicates}
\title{Find duplicate rows in forecast data}
\usage{
find_duplicates(data, forecast_unit = NULL, counts = FALSE)
}
\arguments{
\item{data}{A data.frame (or similar) with predicted and observed values.
See the details section of for additional information
on the required input format.}

\item{forecast_unit}{(optional) Name of the columns in \code{data} (after
any renaming of columns) that denote the unit of a
single forecast. See \code{\link[=get_forecast_unit]{get_forecast_unit()}} for details.
If \code{NULL} (the default), all columns that are not required columns are
assumed to form the unit of a single forecast. If specified, all columns
that are not part of the forecast unit (or required columns) will be removed.}

\item{counts}{Should the output show the number of duplicates per
forecast unit instead of the individual duplicated rows?
Default is \code{FALSE}.}
}
\value{
A \code{data.table} of duplicate rows (or counts).
}
\description{
Find duplicate rows in forecast data
}
\keyword{internal}
