% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geom_mean_helper}
\alias{geom_mean_helper}
\title{Calculate Geometric Mean}
\usage{
geom_mean_helper(x)
}
\arguments{
\item{x}{numeric vector of values for which to calculate the geometric mean}
}
\value{
the geometric mean of the values in \code{x}
}
\description{
Calculate Geometric Mean
}
\keyword{internal}
