% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise-comparisons.R
\name{geometric_mean}
\alias{geometric_mean}
\title{Calculate Geometric Mean}
\usage{
geometric_mean(x)
}
\arguments{
\item{x}{numeric vector of values for which to calculate the geometric mean}
}
\value{
the geometric mean of the values in \code{x}. \code{NA} values are ignored.
}
\description{
Calculate Geometric Mean
}
\details{
Used in \code{\link[=get_pairwise_comparisons]{get_pairwise_comparisons()}}.
}
\keyword{internal}
