% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_-functions.R
\name{get_complete_forecasts}
\alias{get_complete_forecasts}
\title{Get Complete Forecasts}
\usage{
get_complete_forecasts(data)
}
\arguments{
\item{data}{A data.frame or data.table with predicted and observed values.}
}
\value{
A data.table with the same columns as the input, but
without rows where either \code{predicted} or \code{observed} is \code{NA}.
}
\description{
Helper function to remove rows from a data.frame where the
value in either one of the columns \code{predicted} or \code{observed} is \code{NA}.
}
\keyword{check-forecasts}
