% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{get_correlations}
\alias{get_correlations}
\title{Correlation Between Metrics}
\usage{
get_correlations(scores, metrics = NULL, digits = NULL)
}
\arguments{
\item{scores}{An object of class \code{scores} (a data.table with
scores and an additional attribute \code{metrics} as produced by \code{\link[=score]{score()}})}

\item{metrics}{A character vector with the metrics to show. If set to
\code{NULL} (default), all metrics present in \code{scores} will
be shown}

\item{digits}{A number indicating how many decimal places the result should
be rounded to. By default (\code{digits = NULL}) no rounding takes place.}
}
\value{
An object of class \code{scores} (a data.table with an additional
attribute \code{metrics} holding the names of the scores) with correlations
between different metrics
}
\description{
Calculate the correlation between different metrics for a data.frame of
scores as produced by \code{\link[=score]{score()}}.
}
\examples{
scores <- score(as_forecast(example_quantile))
get_correlations(scores, digits = 2)
}
\keyword{scoring}
