% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-forecast-type-score().R
\name{get_forecast_type}
\alias{get_forecast_type}
\title{Infer the type of a forecast based on a data.frame}
\usage{
get_forecast_type(data)
}
\arguments{
\item{data}{A data.frame or data.table with the predictions and observations.
For scoring using \code{\link[=score]{score()}}, the following columns need to be present:
\itemize{
\item \code{observed} - the observed values
\item \code{predicted} - predictions or predictive samples for one
observed value. (You only don't need to provide a prediction column if
you want to score quantile forecasts in a wide range format.)}
For scoring integer and continuous forecasts a \code{sample_id} column is needed:
\itemize{
\item \code{sample_id} - an index to identify the predictive samples in the
'predicted' column generated by one model for one observed value. Only
necessary for continuous and integer forecasts, not for
binary predictions.}
For scoring predictions in a quantile-format forecast you should provide
a column called \code{quantile}:
\itemize{
\item \code{quantile}: quantile to which the prediction corresponds
}

In addition a \code{model} column is suggested and if not present this will be
flagged and added to the input data with all forecasts assigned as an
"unspecified model").

You can check the format of your data using \code{\link[=check_forecasts]{check_forecasts()}} and there
are examples for each format (\link{example_quantile}, \link{example_continuous},
\link{example_integer}, and \link{example_binary}).}
}
\value{
Character vector of length one with either "binary", "quantile",
"sample" or "point".
}
\description{
Internal helper function to get the type of the forecast.
Options are "sample-based", "quantile-based", "binary" or "point" forecast.
The function runs additional checks to make sure the data satisfies
requirements and throws an informative error if any issues are found.
}
\keyword{internal}
