% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_-functions.R
\name{get_forecast_type_from_class}
\alias{get_forecast_type_from_class}
\title{Get Forecast Type From A Class Attribute}
\usage{
get_forecast_type_from_class(data)
}
\arguments{
\item{data}{An object of class \verb{forecast_*} object
as produced by \code{as_forecast()}}
}
\value{
Character vector of length one with either "binary", "quantile",
"sample" or "point".
}
\description{
Internal helper function to get the
forecast type from a class attribute.
}
\keyword{check-forecasts}
