% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-unit.R
\name{get_grouping}
\alias{get_grouping}
\title{Get grouping for a multivariate forecast}
\usage{
get_grouping(forecast)
}
\value{
A character vector with the names of the columns that define the grouping.
}
\description{
Helper function to get the grouping for a multivariate forecast.
}
