% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-forecast-multivariate-sample.R
\name{get_joint_across}
\alias{get_joint_across}
\title{Get joint_across for a multivariate forecast}
\usage{
get_joint_across(forecast)
}
\arguments{
\item{forecast}{A forecast object (a validated data.table with predicted and
observed values).}
}
\value{
A character vector with the names of the columns that the forecast is
joint across.
}
\description{
Recovers the \code{joint_across} columns from a multivariate forecast by
taking the set difference between the forecast unit and the grouping.
This is the inverse of \code{\link[=get_grouping]{get_grouping()}}.
}
\keyword{diagnose-inputs}
