% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_metrics}
\alias{get_metrics}
\title{Get metrics}
\usage{
get_metrics(x, ...)
}
\arguments{
\item{x}{A \code{forecast} or \code{scores} object.}

\item{...}{Additional arguments passed to the method.}
}
\description{
Generic function to to obtain default metrics available for scoring or metrics
that were used for scoring.
\itemize{
\item If called on a \code{forecast} object it returns a list of functions that can be
used for scoring.
\item If called on a \code{scores} object (see \code{\link[=score]{score()}}), it returns a character vector
with the names of the metrics that were used for scoring.
}

See the documentation for the actual methods in the \verb{See Also} section below
for more details. Alternatively call \verb{?get_metrics.<forecast_type>} or
\code{?get_metrics.scores}.
}
\seealso{
Other get_metrics functions:
\code{\link[=get_metrics.forecast_binary]{get_metrics.forecast_binary()}},
\code{\link[=get_metrics.forecast_multivariate_point]{get_metrics.forecast_multivariate_point()}},
\code{\link[=get_metrics.forecast_multivariate_sample]{get_metrics.forecast_multivariate_sample()}},
\code{\link[=get_metrics.forecast_nominal]{get_metrics.forecast_nominal()}},
\code{\link[=get_metrics.forecast_ordinal]{get_metrics.forecast_ordinal()}},
\code{\link[=get_metrics.forecast_point]{get_metrics.forecast_point()}},
\code{\link[=get_metrics.forecast_quantile]{get_metrics.forecast_quantile()}},
\code{\link[=get_metrics.forecast_sample]{get_metrics.forecast_sample()}},
\code{\link[=get_metrics.scores]{get_metrics.scores()}}
}
\concept{get_metrics functions}
\keyword{handle-metrics}
