% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-forecast-type.R
\name{get_observed_type}
\alias{get_observed_type}
\title{Get type of the observed values in a forecast object}
\usage{
get_observed_type(forecast)
}
\arguments{
\item{forecast}{A forecast object (a validated data.table with predicted and
observed values).}
}
\value{
Character vector of length one with either "classification",
"integer", or "continuous".
}
\description{
Extract the \code{observed} column from a forecast object and determine its type
using \code{\link[=get_vector_type]{get_vector_type()}}.
}
\examples{
get_observed_type(example_sample_continuous)
get_observed_type(example_binary)
}
