% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pit.R
\name{get_pit}
\alias{get_pit}
\title{Probability integral transformation (data.frame version)}
\usage{
get_pit(forecast, by, n_replicates = 100)
}
\arguments{
\item{forecast}{A forecast object (a validated data.table with predicted and
observed values, see \code{\link[=as_forecast]{as_forecast()}}).}

\item{by}{Character vector with the columns according to which the
PIT values shall be grouped. If you e.g. have the columns 'model' and
'location' in the input data and want to have a PIT histogram for
every model and location, specify \code{by = c("model", "location")}.}

\item{n_replicates}{The number of draws for the randomised PIT for
discrete predictions. Will be ignored if forecasts are continuous.}
}
\value{
A data.table with PIT values according to the grouping specified in
\code{by}.
}
\description{
Compute the Probability Integral Transformation (PIT) for
validated forecast objects.
}
\examples{
example <- as_forecast_sample(example_sample_continuous)
result <- get_pit(example, by = "model")
plot_pit(result)

# example with quantile data
example <- as_forecast_quantile(example_quantile)
result <- get_pit(example, by = "model")
plot_pit(result)
}
\references{
Sebastian Funk, Anton Camacho, Adam J. Kucharski, Rachel Lowe,
Rosalind M. Eggo, W. John Edmunds (2019) Assessing the performance of
real-time epidemic forecasts: A case study of Ebola in the Western Area
region of Sierra Leone, 2014-15, \doi{10.1371/journal.pcbi.1006785}
}
\keyword{scoring}
