% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_prediction_type}
\alias{get_prediction_type}
\title{Get prediction type of a forecast}
\usage{
get_prediction_type(data)
}
\arguments{
\item{data}{A data.frame or similar as would be used for \code{\link[=score]{score()}}}
}
\value{
Character vector of length one with either "quantile", "integer", or
"continuous".
}
\description{
Internal helper function to get the prediction type of a
forecast. That is inferred based on the properties of the values in the
\code{prediction} column.
}
\keyword{internal}
