% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-quantile-interval-range.R
\name{get_range_from_quantile}
\alias{get_range_from_quantile}
\title{Get interval range belonging to a quantile}
\usage{
get_range_from_quantile(quantile_level)
}
\arguments{
\item{quantile_level}{A numeric vector of quantile levels of size N.}
}
\value{
a numeric vector of interval ranges of size N
}
\description{
Every quantile can be thought of either as the lower or the
upper bound of a symmetric central prediction interval. This helper function
returns the range of the central prediction interval to which the quantile
belongs.

Due to numeric instability that sometimes occurred in the past, ranges are
rounded to 10 decimal places. This is not a problem for the vast majority of
use cases, but it is something to be aware of.
}
\keyword{internal}
