% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_-functions.R
\name{get_score_names}
\alias{get_score_names}
\title{Get metrics that were used for scoring}
\usage{
get_score_names(scores, error = FALSE)
}
\arguments{
\item{scores}{A data.table with an attribute \code{score_names}}

\item{error}{Throw an error if there is no attribute called \code{score_names}?
Default is FALSE.}
}
\value{
Character vector with the metrics that were used for scoring.
}
\description{
Internal helper function to get the metrics that were used
to score forecasts.
}
\keyword{check-forecasts}
