% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_-functions.R
\name{get_scoringutils_attributes}
\alias{get_scoringutils_attributes}
\title{Get a list of all attributes of a scoringutils object}
\usage{
get_scoringutils_attributes(object)
}
\arguments{
\item{object}{A object of class \code{scoringutils_}}
}
\value{
A named list with the attributes of that object.
}
\description{
Get a list of all attributes of a scoringutils object
}
\keyword{internal}
