% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-forecast-type.R
\name{get_type}
\alias{get_type}
\title{Get type of a vector or matrix of observed values or predictions}
\usage{
get_type(x)
}
\arguments{
\item{x}{Input the type should be determined for.}
}
\value{
Character vector of length one with either "classification",
"integer", or "continuous".
}
\description{
Internal helper function to get the type of a vector (usually
of observed or predicted values). The function checks whether the input is
a factor, or else whether it is integer (or can be coerced to integer) or
whether it's continuous.
}
\keyword{internal_input_check}
