% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_unit_of_forecast}
\alias{get_unit_of_forecast}
\title{Get unit of a single forecast}
\usage{
get_unit_of_forecast(data)
}
\arguments{
\item{data}{A data.frame or similar as would be used for \code{\link[=score]{score()}}}
}
\value{
A character vector with the column names that define the unit of
a single forecast
}
\description{
Helper function to get the unit of a single forecast, i.e.
the column names that define where a single forecast was made for
}
\keyword{internal}
