% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-forecast-type.R
\name{get_vector_type}
\alias{get_vector_type}
\title{Get type of a vector or matrix}
\usage{
get_vector_type(x)
}
\arguments{
\item{x}{Input the type should be determined for.}
}
\value{
Character vector of length one with either "classification",
"integer", or "continuous".
}
\description{
Determine the type of a vector or matrix of values. The function checks
whether the input is a factor (returns "classification"), or else whether it
is integer (or can be coerced to integer without loss, returns "integer") or
whether it's continuous (returns "continuous").
}
\examples{
get_vector_type(1:3)
get_vector_type(c(1.5, 2.3))
get_vector_type(factor(c("a", "b")))
}
