% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pit.R
\name{hist_PIT_quantile}
\alias{hist_PIT_quantile}
\title{PIT Histogram Quantile}
\usage{
hist_PIT_quantile(PIT_samples, num_bins = NULL, caption = NULL)
}
\arguments{
\item{PIT_samples}{A vector with the PIT values of size n}

\item{num_bins}{the number of bins in the PIT histogram.}

\item{caption}{provide a caption that gets passed to the plot
If not given, the square root of n will be used}
}
\value{
vector with the scoring values
}
\description{
Make a simple histogram of the probability integral transformed values to
visually check whether a uniform distribution seems likely.
}
