% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-missing-scores.R
\name{impute_model_score}
\alias{impute_model_score}
\title{Impute with a reference model's scores}
\usage{
impute_model_score(model)
}
\arguments{
\item{model}{Character string naming the reference model
whose scores should be used for imputation. The reference
model must have scores for all target combinations that
need imputing; otherwise an error is raised.}
}
\value{
A strategy function for \code{\link[=impute_missing_scores]{impute_missing_scores()}}.
}
\description{
Strategy for \code{\link[=impute_missing_scores]{impute_missing_scores()}} that fills missing
scores with the scores of a specified reference model for
the same target combination.
}
\examples{
\dontshow{
  data.table::setDTthreads(2)
}
scores <- example_quantile |>
  as_forecast_quantile() |>
  score()

impute_missing_scores(
  scores,
  strategy = impute_model_score("EuroCOVIDhub-baseline")
)
}
\seealso{
\code{\link[=impute_missing_scores]{impute_missing_scores()}}
}
\keyword{postprocess-scores}
