% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-missing-scores.R
\name{impute_na_score}
\alias{impute_na_score}
\title{Impute with NA values}
\usage{
impute_na_score()
}
\value{
A strategy function for \code{\link[=impute_missing_scores]{impute_missing_scores()}}.
}
\description{
Strategy for \code{\link[=impute_missing_scores]{impute_missing_scores()}} that fills each
missing metric with \code{NA_real_}.
}
\examples{
\dontshow{
  data.table::setDTthreads(2)
}
scores <- example_quantile |>
  as_forecast_quantile() |>
  score()

impute_missing_scores(scores, strategy = impute_na_score())
}
\seealso{
\code{\link[=impute_missing_scores]{impute_missing_scores()}}
}
\keyword{postprocess-scores}
