% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise-comparisons.R
\name{infer_rel_skill_rule}
\alias{infer_rel_skill_rule}
\title{Infer rule for pairwise comparisons}
\usage{
infer_rel_skill_rule(scores)
}
\arguments{
\item{scores}{A data.table of scores as produced by \code{\link[=score]{score()}}.}
}
\description{
Helper function to infer the rule for which pairwise comparisons shall
be made. The function simply checks the names of the available columns and
chooses the most widely used rule.
Used in \code{\link[=pairwise_comparison]{pairwise_comparison()}}.
}
\keyword{internal}
