% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_handling.R
\name{interval_long_to_quantile}
\alias{interval_long_to_quantile}
\title{Change data from an interval format to a quantile format}
\usage{
interval_long_to_quantile(data, keep_range_col = FALSE)
}
\arguments{
\item{data}{A data.frame following the specifications from
\code{\link[=score]{score()}}) for quantile forecasts.}

\item{keep_range_col}{Logical, default is \code{FALSE}. Keep the
interval_range and boundary columns after transformation?}
}
\value{
a data.table in a plain quantile format
}
\description{
Transform data from a format that uses interval ranges to denote quantiles
to a format that uses quantiles only.
}
\keyword{internal}
