% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_scoringutils}
\alias{is_scoringutils}
\title{Check whether object is of class scoringutils}
\usage{
is_scoringutils(data)
}
\arguments{
\item{data}{An object to be checked}
}
\value{
Logical, either TRUE or FALSE
}
\description{
Helper function to determine whether an object is of class
scoringutils.
}
\keyword{internal}
