% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_scoringutils_check}
\alias{is_scoringutils_check}
\title{Check whether object has been checked with check_forecasts()}
\usage{
is_scoringutils_check(data)
}
\arguments{
\item{data}{An object of class \code{scoringutils_check()} as produced by
\code{\link[=check_forecasts]{check_forecasts()}}.}
}
\value{
Logical, either TRUE or FALSE
}
\description{
Helper function to determine whether an object has been checked
by and passed \code{\link[=check_forecasts]{check_forecasts()}}.
}
\keyword{internal}
