% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-sample.R
\name{logs_sample}
\alias{logs_sample}
\title{Logarithmic score}
\usage{
logs_sample(observed, predicted, ...)
}
\arguments{
\item{observed}{A vector with observed values of size n}

\item{predicted}{nxN matrix of predictive samples, n (number of rows) being
the number of data points and N (number of columns) the number of Monte
Carlo samples. Alternatively, \code{predicted} can just be a vector of size n.}

\item{...}{additional arguments passed to
\link[scoringRules:scores_sample_univ]{logs_sample()} from the scoringRules package.}
}
\value{
vector with the scoring values
}
\description{
Wrapper around the \code{\link[scoringRules:scores_sample_univ]{logs_sample()}}
function from the
\pkg{scoringRules} package. Used to score continuous predictions.
While the Log Score is in theory also applicable
to integer forecasts, the problem lies in the implementation: The Log Score
needs a kernel density estimation, which is not well defined with
integer-valued Monte Carlo Samples. The Log Score can be used for specific
integer valued probability distributions. See the scoringRules package for
more details.
}
\examples{
observed <- rpois(30, lambda = 1:30)
predicted <- replicate(200, rpois(n = 30, lambda = 1:30))
logs_sample(observed, predicted)
}
\references{
Alexander Jordan, Fabian Krüger, Sebastian Lerch, Evaluating Probabilistic
Forecasts with scoringRules, \url{https://www.jstatsoft.org/article/view/v090i12}
}
\keyword{rule}
