% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_handling.R
\name{merge_pred_and_obs}
\alias{merge_pred_and_obs}
\title{Merge Forecast Data And Observations}
\usage{
merge_pred_and_obs(forecasts, observations, by = NULL)
}
\arguments{
\item{forecasts}{data.frame with the forecast data (as can be passed to
\code{\link{eval_forecasts}}).}

\item{observations}{data.frame with the observations}

\item{by}{character vector that denotes the columns by which to merge. Any
value that is not a column in observations will be removed.}
}
\value{
a data.frame with forecasts and observations
}
\description{
The function more or less provides a wrapper around \code{merge} that
aims to handle the merging well if additional columns are present
in one or both data sets. If in doubt, you should probably merge the
data sets manually.
}
