% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-scoring-rules.R
\name{metrics}
\alias{metrics}
\alias{metrics.default}
\alias{metrics.forecast_binary}
\alias{metrics.forecast_point}
\alias{metrics.forecast_sample}
\alias{metrics.forecast_quantile}
\title{Default metrics and scoring rules}
\usage{
metrics(forecast, select = NULL, exclude = NULL, ...)

\method{metrics}{default}(forecast, select = NULL, exclude = NULL, ...)

\method{metrics}{forecast_binary}(forecast, select = NULL, exclude = NULL, ...)

\method{metrics}{forecast_point}(forecast, select = NULL, exclude = NULL, ...)

\method{metrics}{forecast_sample}(forecast, select = NULL, exclude = NULL, ...)

\method{metrics}{forecast_quantile}(forecast, select = NULL, exclude = NULL, ...)
}
\arguments{
\item{select}{A character vector of scoring rules to select from the list. If
\code{select} is \code{NULL} (the default), all possible scoring rules are returned.}

\item{exclude}{A character vector of scoring rules to exclude from the list.
If \code{select} is not \code{NULL}, this argument is ignored.}
}
\value{
A list of scoring rules.
}
\description{
Helper function that returns a named list of default
scoring rules.

The default for the different forecast types are:

\strong{binary}
\itemize{
\item "brier_score" = \code{\link[=brier_score]{brier_score()}}
\item "log_score" = \code{\link[=logs_binary]{logs_binary()}}
}

\strong{point}
\itemize{
\item "ae_point" = \link[Metrics:ae]{ae()}
\item "se_point" = \link[Metrics:se]{se()}
\item "ape" = \link[Metrics:ape]{ape()}
}

\strong{sample}
\itemize{
\item "crps" = \code{\link[=crps_sample]{crps_sample()}}
\item "log_score" = \code{\link[=logs_sample]{logs_sample()}}
\item "dss" = \code{\link[=dss_sample]{dss_sample()}}
\item "mad" = \code{\link[=mad_sample]{mad_sample()}}
\item "bias" = \code{\link[=bias_sample]{bias_sample()}}
\item "ae_median" = \code{\link[=ae_median_sample]{ae_median_sample()}}
\item "se_mean" = \code{\link[=se_mean_sample]{se_mean_sample()}}
}

\strong{quantile}

The default scoring rules are:
\itemize{
\item "wis" = \link{wis}
\item "overprediction" = \code{\link[=overprediction]{overprediction()}}
\item "underprediction" = \code{\link[=underprediction]{underprediction()}}
\item "dispersion" = \code{\link[=dispersion]{dispersion()}}
\item "bias" = \code{\link[=bias_quantile]{bias_quantile()}}
\item "interval_coverage_50" = \code{\link[=interval_coverage]{interval_coverage()}}
\item "interval_coverage_90" = customise_metric(
interval_coverage, interval_range = 90
)
\item "interval_coverage_deviation" = \code{\link[=interval_coverage_deviation]{interval_coverage_deviation()}},
\item "ae_median" = \code{\link[=ae_median_quantile]{ae_median_quantile()}}
}

Note: The \code{interval_coverage_90} scoring rule is created by modifying
\code{\link[=interval_coverage]{interval_coverage()}}, making use of the function \code{\link[=customise_metric]{customise_metric()}}.
This construct allows the function to deal with arbitrary arguments in \code{...},
while making sure that only those that \code{\link[=interval_coverage]{interval_coverage()}} can
accept get passed on to it. \code{interval_range = 90} is set in the function
definition, as passing an argument \code{interval_range = 90} to \code{\link[=score]{score()}} would
mean it would also get passed to \code{interval_coverage_50}.
}
\examples{
metrics(as_forecast_binary(example_binary))
metrics(as_forecast_point(example_point), select = "ape")
metrics(as_forecast_sample(example_sample_continuous), exclude = "mad")
}
\keyword{metric}
