% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metrics_binary}
\alias{metrics_binary}
\title{Default metrics for binary forecasts.}
\format{
An object of class \code{list} of length 2.
}
\usage{
metrics_binary
}
\description{
A named list with functions:
\itemize{
\item "brier_score" = \code{\link[=brier_score]{brier_score()}}
\item "log_score" = \code{\link[=logs_binary]{logs_binary()}}
}
}
\keyword{info}
