% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-scoring-rules.R
\name{metrics_nominal}
\alias{metrics_nominal}
\title{Scoring rules for nominal forecasts}
\usage{
metrics_nominal(select = NULL, exclude = NULL)
}
\arguments{
\item{select}{A character vector of scoring rules to select from the list. If
\code{select} is \code{NULL} (the default), all possible scoring rules are returned.}

\item{exclude}{A character vector of scoring rules to exclude from the list.
If \code{select} is not \code{NULL}, this argument is ignored.}
}
\value{
A list of scoring functions.
}
\description{
Helper function that returns a named list of default
scoring rules suitable for nominal forecasts.

The default scoring rules are:
\itemize{
\item "log_score" = \code{\link[=logs_nominal]{logs_nominal()}}
}
}
\examples{
metrics_nominal()
metrics_nominal(select = "log_score")
}
\keyword{metric}
