% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metrics_point}
\alias{metrics_point}
\title{Default metrics for point forecasts.}
\format{
An object of class \code{list} of length 3.
}
\usage{
metrics_point
}
\description{
A named list with functions:
\itemize{
\item "ae_point" = \link[Metrics:ae]{ae()}
\item "se_point" = \link[Metrics:se]{se()}
\item "ape" = \link[Metrics:ape]{ape()}
}
}
\keyword{info}
