% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metrics_quantile}
\alias{metrics_quantile}
\title{Default metrics for quantile-based forecasts.}
\format{
An object of class \code{list} of length 8.
}
\usage{
metrics_quantile
}
\description{
A named list with functions:
\itemize{
\item "wis" = \code{\link[=wis]{wis()}}
\item "bias" = \code{\link[=bias_quantile]{bias_quantile()}}
\item "coverage_50" = \(...) {run_safely(..., range = 50, fun = interval_coverage_quantile)} #nolint
\item "coverage_90" = \(...) {run_safely(..., range = 90, fun = interval_coverage_quantile)} #nolint
}
}
\keyword{info}
