% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metrics_sample}
\alias{metrics_sample}
\title{Default metrics for sample-based forecasts.}
\format{
An object of class \code{list} of length 7.
}
\usage{
metrics_sample
}
\description{
A named list with functions:
\itemize{
\item "mad" = \code{\link[=mad_sample]{mad_sample()}}
\item "bias" = \code{\link[=bias_sample]{bias_sample()}}
\item "dss" = \code{\link[=dss_sample]{dss_sample()}}
\item "crps" = \code{\link[=crps_sample]{crps_sample()}}
\item "log_score" = \code{\link[=logs_sample]{logs_sample()}}
\item "mad" = \code{\link[=mad_sample]{mad_sample()}}
\item "ae_median" = \code{\link[=ae_median_sample]{ae_median_sample()}}
\item "se_mean" = \code{\link[=se_mean_sample]{se_mean_sample()}}
}
}
\keyword{info}
