% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{new_forecast}
\alias{new_forecast}
\title{Class constructor for \code{forecast} objects}
\usage{
new_forecast(data, classname)
}
\arguments{
\item{data}{A data.frame (or similar) with predicted and observed values.
See the details section of \code{\link[=as_forecast]{as_forecast()}} for additional information
on required input formats.}

\item{classname}{name of the class to be created}
}
\value{
An object of the class indicated by \code{classname}
}
\description{
Construct a class based on a data.frame or similar. The constructor
\itemize{
\item coerces the data into a data.table
\item makes sure that a column called \code{model} exists and if not creates one
\item assigns a class
}
}
\keyword{internal}
