% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{new_score}
\alias{new_score}
\title{Create An Object Of Class \code{score}}
\usage{
new_score(scores, score_names)
}
\description{
This function creates an object of class \code{score} based on a
data.table or similar.
}
