% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{new_scoringutils}
\alias{new_scoringutils}
\title{Class constructor for scoringutils objects}
\usage{
new_scoringutils(data, classname)
}
\arguments{
\item{data}{A data.frame or data.table with predicted and observed values.}

\item{classname}{name of the class to be created}
}
\value{
An object of the class indicated by \code{classname}
}
\description{
Construct a class based on a data.frame or similar. The constructor
\itemize{
\item coerces the data into a data.table
\item makes sure that a column called \code{model} exists and if not creates one
\item assigns a class
}
}
\keyword{validate}
