% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{permutation_test}
\alias{permutation_test}
\title{Simple permutation test}
\usage{
permutation_test(
  scores1,
  scores2,
  n_permutation = 999,
  one_sided = FALSE,
  comparison_mode = c("difference", "ratio")
)
}
\value{
p-value of the permutation test
}
\description{
#' The implementation of the permutation test follows the
function
\code{permutationTest} from the \code{surveillance} package by Michael Höhle,
Andrea Riebler and Michaela Paul.
}
\keyword{internal}
