% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_probabilstic_forecasts.R
\name{PIT}
\alias{PIT}
\title{PIT}
\usage{
PIT(true_values, samples, num_bins = NULL, n_replicates = 20)
}
\arguments{
\item{true_values}{A vector with the true observed values of size n}

\item{samples}{nxN matrix of predictive samples, n (number of rows) being
the number of data points and N (number of columns) the
number of Monte Carlo samples}

\item{num_bins}{the number of bins in the PIT histogram.
If not given, will use the square root of n}

\item{n_replicates}{the number of tests to perform,
each time re-randomising the PIT}
}
\value{
list
}
\description{
Performs an Anderson-Darling test for uniformity for a
randomised PIT histogram using predictive Monte-Carlo samples

See Eqs. 1 and 2 in Czado, Geniting and Held (2009),
Predictive Model Assessment for Count Data,
Biometrics Vol. 65, No. 4 (Dec., 2009), pp. 1254-1261
}
