% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.forecast_counts}
\alias{plot.forecast_counts}
\title{Visualise Where Forecasts Are Available}
\usage{
\method{plot}{forecast_counts}(
  x,
  yvar = "model",
  xvar = "forecast_date",
  make_xvar_factor = TRUE,
  show_numbers = TRUE,
  ...
)
}
\arguments{
\item{x}{an S3 object of class "forecast_counts"
as produced by \code{\link[=get_forecast_counts]{get_forecast_counts()}}}

\item{yvar}{character vector of length one that denotes the name of the column
to appear on the y-axis of the plot. Default is "model".}

\item{xvar}{character vector of length one that denotes the name of the column
to appear on the x-axis of the plot. Default is "forecast_date".}

\item{make_xvar_factor}{logical (default is TRUE). Whether or not to convert
the variable on the x-axis to a factor. This has an effect e.g. if dates
are shown on the x-axis.}

\item{show_numbers}{logical (default is \code{TRUE}) that indicates whether
or not to show the actual count numbers on the plot}

\item{...}{additional arguments (not used here)}
}
\value{
ggplot object with a plot of interval coverage
}
\description{
Visualise Where Forecasts Are Available
}
\examples{
library(ggplot2)
available_forecasts <- get_forecast_counts(
  example_quantile, by = c("model", "target_type", "target_end_date")
)
plot(
 available_forecasts, xvar = "target_end_date", show_numbers = FALSE
) +
 facet_wrap("target_type")
}
