% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.scoringutils_available_forecasts}
\alias{plot.scoringutils_available_forecasts}
\title{Visualise Where Forecasts Are Available}
\usage{
\method{plot}{scoringutils_available_forecasts}(
  x,
  yvar = "model",
  xvar = "forecast_date",
  make_xvar_factor = TRUE,
  show_numbers = TRUE,
  ...
)
}
\arguments{
\item{x}{an S3 object of class "scoringutils_available_forecasts"
as produced by \code{\link[=available_forecasts]{available_forecasts()}}}

\item{yvar}{character vector of length one that denotes the name of the column
to appear on the y-axis of the plot. Default is "model".}

\item{xvar}{character vector of length one that denotes the name of the column
to appear on the x-axis of the plot. Default is "forecast_date".}

\item{make_xvar_factor}{logical (default is TRUE). Whether or not to convert
the variable on the x-axis to a factor. This has an effect e.g. if dates
are shown on the x-axis.}

\item{show_numbers}{logical (default is \code{TRUE}) that indicates whether
or not to show the actual count numbers on the plot}

\item{...}{additional arguments (not used here)}
}
\value{
ggplot object with a plot of interval coverage
}
\description{
Visualise Where Forecasts Are Available
}
\examples{
library(ggplot2)
available_forecasts <- available_forecasts(
  example_quantile, by = c("model", "target_type", "target_end_date")
)
plot(
 available_forecasts, xvar = "target_end_date", show_numbers = FALSE
) +
 facet_wrap("target_type")
}
\seealso{
Other plotting functions: 
\code{\link{plot.scoringutils_correlation}()},
\code{\link{plot.scoringutils_pairwise}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_interval_coverage}()},
\code{\link{plot_pit}()},
\code{\link{plot_predictions}()},
\code{\link{plot_quantile_coverage}()},
\code{\link{plot_ranges}()},
\code{\link{plot_score_table}()},
\code{\link{plot_wis}()}
}
\concept{plotting functions}
