% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_avail_forecasts}
\alias{plot_avail_forecasts}
\title{Visualise Where Forecasts Are Available \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
plot_avail_forecasts(
  available_forecasts,
  y = "model",
  x = "forecast_date",
  make_x_factor = TRUE,
  show_numbers = TRUE
)
}
\arguments{
\item{available_forecasts}{an S3 object of class "scoringutils_available_forecasts"
as produced by \code{\link[=available_forecasts]{available_forecasts()}}}

\item{y}{character vector of length one that denotes the name of the column
to appear on the y-axis of the plot. Default is "model".}

\item{x}{character vector of length one that denotes the name of the column
to appear on the x-axis of the plot. Default is "forecast_date".}

\item{make_x_factor}{logical (default is TRUE). Whether or not to convert
the variable on the x-axis to a factor. This has an effect e.g. if dates
are shown on the x-axis.}

\item{show_numbers}{logical (default is \code{TRUE}) that indicates whether
or not to show the actual count numbers on the plot}
}
\description{
Old version of \code{\link[=plot.scoringutils_available_forecasts]{plot.scoringutils_available_forecasts()}} for compatibility.
}
