% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_correlation}
\alias{plot_correlation}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Plot Correlation Between Metrics}
\usage{
plot_correlation(correlations)
}
\arguments{
\item{correlations}{A data.table with correalations as produced by
\code{\link[=correlation]{correlation()}}}
}
\description{
Deprecated version of \code{\link[=plot.scoringutils_correlation]{plot.scoringutils_correlation()}} for compatibility.
}
