% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_interval_ranges}
\alias{plot_interval_ranges}
\title{Plot Metrics by Range of the Prediction Interval}
\usage{
plot_interval_ranges(scores, y = "wis", x = "model", colour = "interval_range")
}
\arguments{
\item{scores}{A data.frame of scores based on quantile forecasts as
produced by \code{\link[=score]{score()}} or \code{\link[=summarise_scores]{summarise_scores()}}. Note that "interval_range"
must be included in the \code{by} argument when running \code{\link[=summarise_scores]{summarise_scores()}}}

\item{y}{The variable from the scores you want to show on the y-Axis.
This could be something like "wis" (the default) or "dispersion"}

\item{x}{The variable from the scores you want to show on the x-Axis.
Usually this will be "model"}

\item{colour}{Character vector of length one used to determine a variable
for colouring dots. The Default is "interval_range".}
}
\value{
A ggplot2 object showing a contributions from the three components of
the weighted interval score
}
\description{
Visualise the metrics by range, e.g. if you are interested how different
interval interval ranges contribute to the overall interval score, or how
sharpness / dispersion changes by interval range.
}
\examples{
library(ggplot2)
ex <- example_quantile
ex$range <- scoringutils:::get_range_from_quantile(ex$quantile)
scores <- score(ex, metrics = list("wis" = wis))
scores$interval_range <- scores$range
summarised <- summarise_scores(
  scores,
  by = c("model", "target_type", "interval_range")
)
plot_interval_ranges(summarised, x = "model") +
  facet_wrap(~target_type, scales = "free")
}
