% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pairwise_comparison}
\alias{plot_pairwise_comparison}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Plot Heatmap of Pairwise Comparisons}
\usage{
plot_pairwise_comparison(
  comparison_result,
  type = c("mean_scores_ratio", "pval")
)
}
\arguments{
\item{comparison_result}{A data.frame as produced by
\code{\link[=pairwise_comparison]{pairwise_comparison()}}}

\item{type}{character vector of length one that is either
"mean_scores_ratio" or "pval". This denotes whether to
visualise the ratio or the p-value of the pairwise comparison.
Default is "mean_scores_ratio".}
}
\value{
A data.table with pairwise comparisons. In addition, the output is
of class \code{scoringutils_pairwise} and can be visualised using \code{\link[=plot]{plot()}}.
}
\description{
Deprecated version of \code{\link[=plot.scoringutils_pairwise]{plot.scoringutils_pairwise()}}
}
\examples{
library(ggplot2)
scores <- score(example_quantile)
pairwise <- pairwise_comparison(scores, by = "target_type")
plot(pairwise, type = "mean_scores_ratio") +
  facet_wrap(~target_type)
}
\seealso{
Other plotting functions: 
\code{\link{plot.scoringutils_available_forecasts}()},
\code{\link{plot.scoringutils_correlation}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_interval_coverage}()},
\code{\link{plot_pit}()},
\code{\link{plot_predictions}()},
\code{\link{plot_quantile_coverage}()},
\code{\link{plot_ranges}()},
\code{\link{plot_score_table}()},
\code{\link{plot_wis}()}
}
