% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.forecast_binary}
\alias{print.forecast_binary}
\title{Print information about a \code{forecast_binary} object.}
\usage{
\method{print}{forecast_binary}(x, ...)
}
\description{
Print information about a \code{forecast_binary} object.
}
