% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.forecast_integer}
\alias{print.forecast_integer}
\title{Print information about a \code{forecast_integer} object.}
\usage{
\method{print}{forecast_integer}(x, ...)
}
\description{
Print information about a \code{forecast_integer} object.
}
