% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.forecast_point}
\alias{print.forecast_point}
\title{Print information about a \code{forecast_point} object.}
\usage{
\method{print}{forecast_point}(x, ...)
}
\description{
Print information about a \code{forecast_point} object.
}
