% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.forecast_quantile}
\alias{print.forecast_quantile}
\title{Print information about a \code{forecast_quantile} object.}
\usage{
\method{print}{forecast_quantile}(x, ...)
}
\description{
Print information about a \code{forecast_quantile} object.
}
