% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.forecast_sample}
\alias{print.forecast_sample}
\title{Print information about a \code{forecast_sample} object.}
\usage{
\method{print}{forecast_sample}(x, ...)
}
\description{
Print information about a \code{forecast_sample} object.
}
