% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-forecast-multivariate-sample.R
\name{print.forecast_sample_multivariate}
\alias{print.forecast_sample_multivariate}
\title{Print information about a multivariate forecast object}
\usage{
\method{print}{forecast_sample_multivariate}(x, ...)
}
\arguments{
\item{x}{A forecast object of class \code{forecast_sample_multivariate}.}

\item{...}{Additional arguments for \code{\link[=print]{print()}}.}
}
\value{
Returns \code{x} invisibly.
}
\description{
This function prints information about a multivariate forecast object,
including "Forecast type", "Forecast unit", and "Joint across" columns.
}
\keyword{gain-insights}
