% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-scores.R
\name{print.scores}
\alias{print.scores}
\title{Print a scores object}
\usage{
\method{print}{scores}(x, ...)
}
\arguments{
\item{x}{A \code{scores} object}

\item{...}{Additional arguments for \code{\link[=print]{print()}}.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Prints a \code{scores} object. Suppresses autoprinting during data.table
\verb{:=} operations by checking data.table's internal \code{shouldPrint()} flag.
}
\keyword{gain-insights}
