% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_forecast_info}
\alias{print_forecast_info}
\title{Print information about a \code{forecast} object.}
\usage{
print_forecast_info(dat, ...)
}
\description{
This function prints information about a forecast object,
including "Forecast type", "Protected columns", "Score columns",
"Forecast unit".
}
\examples{
print(example_quantile \%>\% as_forecast)
}
\keyword{print-forecasts}
