% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quantile_example_data}
\alias{quantile_example_data}
\title{Quantile Example Data}
\format{
A data frame with 5,152 rows and 10 columns:
\describe{
  \item{value_date}{the date for which a prediction was made}
  \item{value_type}{the target to be predicted (short form)}
  \item{geography}{the region for which a prediction was made}
  \item{value_desc}{long form description of the prediction target}
  \item{true_value}{true observed values}
  \item{model}{name of the model that generated the forecasts}
  \item{creation_date}{date on which the forecast was made}
  \item{quantile}{quantile of the corresponding prediction}
  \item{prediction}{quantile predictions}
  \item{horizon}{forecast horizon in days}

}
}
\usage{
quantile_example_data
}
\description{
A data set with predictions for different quantities relevant in the
2020 UK Covid-19 epidemic.
}
\keyword{datasets}
