% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quantile_example_data_plain}
\alias{quantile_example_data_plain}
\title{Quantile Example Data - Plain Quantile Format}
\format{
A data frame with 600 rows and 6 variables:
\describe{
  \item{true_value}{true observed values}
  \item{id}{unique identifier for true observed values}
  \item{horizon}{forecast horzin (e.g. 1 day ahead forecast)}
  \item{model}{name of the model that generated the forecasts}
  \item{prediction}{quantile predictions}
  \item{quantile}{quantile of the corresponding prediction}
}
}
\usage{
quantile_example_data_plain
}
\description{
A toy dataset for quantile forecasts of an outcome variable in a format
that uses plain quantiles instead of interval ranges
}
\keyword{datasets}
