% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-quantile.R
\name{quantile_score}
\alias{quantile_score}
\title{Quantile score}
\usage{
quantile_score(observed, predicted, quantile_level, weigh = TRUE)
}
\arguments{
\item{observed}{Numeric vector of size n with the observed values.}

\item{predicted}{Numeric nxN matrix of predictive
quantiles, n (number of rows) being the number of forecasts (corresponding
to the number of observed values) and N
(number of columns) the number of quantiles per forecast.
If \code{observed} is just a single number, then predicted can just be a
vector of size N.}

\item{quantile_level}{Vector of of size N with the quantile levels
for which predictions were made.}

\item{weigh}{Logical. If \code{TRUE} (the default), weigh the score by
\eqn{\alpha / 2}, so it can be averaged into an interval score that, in
the limit (for an increasing number of equally spaced quantiles/prediction
intervals), corresponds
to the CRPS. \eqn{\alpha} is the value that corresponds to the
(\eqn{\alpha/2}) or (\eqn{1 - \alpha/2}), i.e. it is the decimal
value that represents how much is outside a central prediction interval
(E.g. for a 90 percent central prediction interval, alpha is 0.1).}
}
\description{
Proper Scoring Rule to score quantile predictions. Smaller values are better.
The quantile score is closely related to the interval score (see \code{\link[=wis]{wis()}}) and
is the quantile equivalent that works with single quantiles instead of
central prediction intervals.
}
\examples{
observed <- rnorm(10, mean = 1:10)
alpha <- 0.5

lower <- qnorm(alpha / 2, rnorm(10, mean = 1:10))
upper <- qnorm((1 - alpha / 2), rnorm(10, mean = 1:10))

qs_lower <- quantile_score(observed,
  predicted = lower,
  quantile_level = alpha / 2
)
qs_upper <- quantile_score(observed,
  predicted = upper,
  quantile_level = 1 - alpha / 2
)
interval_score <- (qs_lower + qs_upper) / 2
}
\references{
Strictly Proper Scoring Rules, Prediction,and Estimation,
Tilmann Gneiting and Adrian E. Raftery, 2007, Journal of the American
Statistical Association, Volume 102, 2007 - Issue 477

Evaluating epidemic forecasts in an interval format,
Johannes Bracher, Evan L. Ray, Tilmann Gneiting and Nicholas G. Reich,
\url{https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1008618}
}
\keyword{metric}
