% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_handling.R
\name{quantile_to_interval}
\alias{quantile_to_interval}
\alias{quantile_to_interval.data.frame}
\alias{quantile_to_interval.numeric}
\title{Transform From a Quantile Format to an Interval Format}
\usage{
quantile_to_interval(...)

\method{quantile_to_interval}{data.frame}(dt, format = "long", keep_quantile_col = FALSE, ...)

\method{quantile_to_interval}{numeric}(observed, predicted, quantile_level, ...)
}
\arguments{
\item{...}{method arguments}

\item{dt}{a data.table with columns \code{quantile_level} and \code{predicted}}

\item{format}{the format of the output. Either "long" or "wide". If "long"
(the default), there will be a column \code{boundary} (with values either "upper"
or "lower" and a column \code{range} that contains the range of the interval.
If "wide", there will be a column \code{range} and two columns
\code{lower} and \code{upper} that contain the lower and upper bounds of the
prediction interval, respectively.}

\item{keep_quantile_col}{keep the \code{quantile_level} column in the final
output after transformation (default is FALSE). This only works if
\code{format = "long"}. If \code{format = "wide"}, the \code{quantile_level} column will
always be dropped.}

\item{observed}{a numeric vector of observed values of size n}

\item{predicted}{a numeric vector of predicted values of size n x N. If
\code{observed} is a single number, then \code{predicted} can be a vector of length N}

\item{quantile_level}{a numeric vector of quantile levels of size N}
}
\value{
\emph{quantile_to_interval.data.frame}:
a data.frame in an interval format (either "long" or "wide"), with or
without a \code{quantile_level} column. Rows will not be reordered.

\emph{quantile_to_interval.numeric}:
a data.frame in a wide interval format with columns \code{forecast_id},
\code{observed}, \code{lower}, \code{upper}, and \code{range}. The \code{forecast_id} column is a
unique identifier for each forecast. Rows will be reordered according to
\code{forecast_id} and \code{range}.
}
\description{
\strong{Quantile format}
In a quantile format, a prediction is characterised by one or multiple
predicted values and the corresponding quantile levels. For example, a
prediction in a quantile format could be represented by the 0.05, 0.25, 0.5,
0.75 and 0.95 quantiles of the predictive distribution.

\strong{Interval format}
In the interval format, two quantiles are assumed to form a prediction
interval. Prediction intervals need to be symmetric around the median and
are characterised by a lower and an upper bound. The lower bound is defined
by the lower quantile and the upper bound is defined by the upper quantile.
A 90\% prediction interval, for example, covers 90\% of the probability mass
and is defined by the 5\% and 95\% quantiles. A forecast could therefore
be characterised by one or multiple prediction intervals, e.g. the lower
and upper bounds of the 50\% and 90\% prediction intervals (corresponding to
the 0.25 and 0.75 as well as the 0.05 and 0.095 quantiles).
}
\keyword{data-handling}
