% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{range_example_data_long}
\alias{range_example_data_long}
\title{Range Forecast Example Data (Long Format)}
\format{
A data frame with 5,419 rows and 12 columns:
\describe{
  \item{value_date}{the date for which a prediction was made}
  \item{value_type}{the target to be predicted (short form)}
  \item{geography}{the region for which a prediction was made}
  \item{value_desc}{long form description of the prediction target}
  \item{true_value}{true observed values}
  \item{model}{name of the model that generated the forecasts}
  \item{creation_date}{date on which the forecast was made}
  \item{prediction}{value for the lower or upper bound of the given prediction interval}
  \item{horizon}{forecast horizon in days}
  \item{boundary}{indicate lower or upper bound of prediction interval}
  \item{range}{range of the corresponding prediction interval}
}
}
\usage{
range_example_data_long
}
\description{
A data set with predictions with different interval ranges relevant in the
2020 UK Covid-19 epidemic.
}
\keyword{datasets}
