% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{range_example_data_wide}
\alias{range_example_data_wide}
\title{Range Forecast Example Data (Wide Format)}
\format{
A data frame with 346 rows and 28 columns:
\describe{
  \item{value_date}{the date for which a prediction was made}
  \item{value_type}{the target to be predicted (short form)}
  \item{geography}{the region for which a prediction was made}
  \item{value_desc}{long form description of the prediction target}
  \item{true_value}{true observed values}
  \item{model}{name of the model that generated the forecasts}
  \item{creation_date}{date on which the forecast was made}
  \item{horizon}{forecast horizon in days}
  \item{lower_0}{prediction for the lower bound of the 0\% interval range (median)}
  \item{lower_10}{prediction for the lower bound of the 10\% interval range}
  \item{lower_20}{prediction for the lower bound of the 20\% interval range}
  \item{lower_30}{prediction for the lower bound of the 30\% interval range}
  \item{lower_40}{prediction for the lower bound of the 40\% interval range}
  \item{lower_50}{prediction for the lower bound of the 50\% interval range}
  \item{lower_60}{prediction for the lower bound of the 60\% interval range}
  \item{lower_70}{prediction for the lower bound of the 70\% interval range}
  \item{lower_80}{prediction for the lower bound of the 80\% interval range}
  \item{lower_90}{prediction for the lower bound of the 90\% interval range}
  \item{upper_0}{prediction for the upper bound of the 0\% interval range}
  \item{upper_10}{prediction for the upper bound of the 1\% interval range}
  \item{upper_20}{prediction for the upper bound of the 20\% interval range}
  \item{upper_30}{prediction for the upper bound of the 30\% interval range}
  \item{upper_40}{prediction for the upper bound of the 40\% interval range}
  \item{upper_50}{prediction for the upper bound of the 50\% interval range}
  \item{upper_60}{prediction for the upper bound of the 60\% interval range}
  \item{upper_70}{prediction for the upper bound of the 70\% interval range}
  \item{upper_80}{prediction for the upper bound of the 80\% interval range}
  \item{upper_90}{prediction for the upper bound of the 90\% interval range}
}
}
\usage{
range_example_data_wide
}
\description{
A data set with predictions with different interval ranges relevant in the
2020 UK Covid-19 epidemic.
}
\keyword{datasets}
