% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_handling.R
\name{range_long_to_quantile}
\alias{range_long_to_quantile}
\title{Change Data from a Range Format to a Quantile Format}
\usage{
range_long_to_quantile(data, keep_range_col = FALSE)
}
\arguments{
\item{data}{a data.frame following the specifications from
\code{\link[=score]{score()}}) for quantile forecasts.}

\item{keep_range_col}{keep the range and boundary columns after
transformation (default is FALSE)}
}
\value{
a data.frame in a plain quantile format
}
\description{
Transform data from a format that uses interval ranges to denote quantiles
to a format that uses quantiles only.
}
\keyword{internal}
