% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_handling.R
\name{range_long_to_wide}
\alias{range_long_to_wide}
\title{Pivot Range Format Forecasts From Long to Wide Format}
\usage{
range_long_to_wide(data)
}
\arguments{
\item{data}{a data.frame following the specifications from
\code{\link[=score]{score()}}) for quantile forecasts.}
}
\value{
a data.frame in wide format
}
\description{
Pivot range data from long to wide
}
\keyword{internal}
