% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_handling.R
\name{range_to_quantile}
\alias{range_to_quantile}
\title{Pivot Change Data from a Range Format to a Quantile Format}
\usage{
range_to_quantile(data, keep_range_col = FALSE)
}
\arguments{
\item{data}{a data.frame following the specifications from
\code{\link{eval_forecasts}}) for quantile forecasts. For an example, see
\code{\link{range_example_data_long}})}

\item{keep_range_col}{keep the range and boundary columns after
transformation (default is FALSE)}
}
\value{
a data.frame in long format
}
\description{
Legacy function that will not be supported in future updates.
}
