% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_handling.R
\name{range_wide_to_long}
\alias{range_wide_to_long}
\title{Pivot Range Format Forecasts From Wide to Long Format}
\usage{
range_wide_to_long(data)
}
\arguments{
\item{data}{a data.frame following the specifications from
\code{\link[=score]{score()}}) for quantile forecasts.}
}
\value{
a data.frame in long format
}
\description{
Pivot wide range format to long range format.
}
\keyword{internal}
