% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{remake_forecast}
\alias{remake_forecast}
\title{Remakes a forecast object with a subclass}
\usage{
remake_forecast(forecast, old_classname, new_classname, verbose = TRUE)
}
\arguments{
\item{forecast}{The forecast object to be modified}

\item{old_classname}{The name of the class to be removed from the forecast
object}

\item{new_classname}{The name of the class to be added to the forecast object}

\item{verbose}{A logical value indicating whether to print verbose output}
}
\value{
The modified forecast object
}
\description{
This function takes a forecast object and modifies its class name by removing
the old class name and adding a new class name. It also performs an assertion
check on the modified forecast object.
}
