% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_scoringutils_class}
\alias{remove_scoringutils_class}
\title{Remove scoringutils_ Class and Attributes}
\usage{
remove_scoringutils_class(object)
}
\arguments{
\item{object}{An object to remove scoringutils classes and attributes from}
}
\value{
The object with scoringutils classes and attributes removed
}
\description{
This function removes all classes that start with
"scoringutils_" and all attributes associated with scoringutils.
}
\keyword{internal}
