% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_scores.R
\name{round_scores}
\alias{round_scores}
\title{Round scores}
\usage{
round_scores(scores, digits = 2, signif = TRUE)
}
\arguments{
\item{scores}{A data.table of scores as produced by \code{\link[=score]{score()}}.}

\item{digits}{the number of significant digits (if \code{signif = TRUE}) or
decimal places (if \code{signif = FALSE}) to be used (default is 2).}

\item{signif}{Whether to round using significant digits. For 2 significant
digits, for example, 36340 would be rounded to 36000, and 0.36340 to 0.36.}
}
\value{
a data.table with rounded scores
}
\description{
Round scores as returned by \code{\link[=score]{score()}} or \code{\link[=summarise_scores]{summarise_scores()}}.
}
\details{
The function rounds all metrics to a desired number of significant digits.
}
\examples{
library(magrittr) # pipe operator

score(example_quantile) \%>\%
  summarise_scores(by = c("model", "target_type")) \%>\%
  round_scores()
}
\keyword{scoring}
