% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rules}
\alias{rules}
\title{Summary information for selected scoring rules}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 22 rows and 8 columns.
}
\usage{
rules
}
\description{
A data set with summary information on selected scoring rules implemented in
\pkg{scoringutils}
}
\details{
The data was created using the script create-rule-tables.R in the inst/
folder (or the top level folder in a compiled package).
}
\keyword{info}
