% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-scoring-rules.R
\name{rules_quantile}
\alias{rules_quantile}
\title{Scoring Rules for Quantile-Based Forecasts}
\usage{
rules_quantile(select = NULL, exclude = NULL)
}
\arguments{
\item{select}{A character vector of scoring rules to select from the list.
If \code{select} is \code{NULL} (the default), all possible scoring rules are returned.}

\item{exclude}{A character vector of scoring rules to exclude from the list.
If \code{select} is not \code{NULL}, this argument is ignored.}
}
\value{
A list of scoring rules.
}
\description{
Helper function that returns a named list of default
scoring rules suitable for forecasts in a quantile-based format

The default scoring rules are:
\itemize{
\item "wis" = \link{wis}
\item "overprediction" = \code{\link[=overprediction]{overprediction()}}
\item "underprediction" = \code{\link[=underprediction]{underprediction()}}
\item "dispersion" = \code{\link[=dispersion]{dispersion()}}
\item "bias" = \code{\link[=bias_quantile]{bias_quantile()}}
\item "coverage_50" = \code{\link[=interval_coverage_quantile]{interval_coverage_quantile()}}
\item "coverage_90" = (...) \{
run_safely(..., range = 90, fun = \link{interval_coverage_quantile})
\}
\item "coverage_deviation" = \code{\link[=interval_coverage_dev_quantile]{interval_coverage_dev_quantile()}},
\item "ae_median" = \code{\link[=ae_median_quantile]{ae_median_quantile()}}
}
}
\examples{
rules_quantile()
rules_quantile(select = "wis")
}
\keyword{info}
