% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_handling.R
\name{sample_to_interval_long}
\alias{sample_to_interval_long}
\title{Change Data from a Sample Based Format to a Long Interval Range Format}
\usage{
sample_to_interval_long(
  data,
  interval_range = c(0, 50, 90),
  type = 7,
  keep_quantile_col = TRUE
)
}
\arguments{
\item{data}{a data.frame with samples}

\item{interval_range}{a numeric vector of interval ranges to extract
(e.g. \code{c(0, 50, 90)})}

\item{type}{type argument passed down to the quantile function. For more
information, see \code{\link[=quantile]{quantile()}}}

\item{keep_quantile_col}{keep quantile_level column, default is TRUE}
}
\value{
a data.table in a long interval interval range format
}
\description{
Transform data from a format that is based on predictive samples to a format
based on interval ranges
}
\keyword{internal}
