% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_handling.R
\name{sample_to_quantile}
\alias{sample_to_quantile}
\title{Change Data from a Sample Based Format to a Quantile Format}
\usage{
sample_to_quantile(data, quantiles = c(0.05, 0.25, 0.5, 0.75, 0.95), ...)
}
\arguments{
\item{data}{a data.frame with samples}

\item{quantiles}{a numeric vector of quantiles to extract}

\item{...}{Additional arguments that can be passed down to \code{\link[=quantile]{quantile()}}.
In contrast to the default behaviour of \code{\link[=quantile]{quantile()}}, this function by
default has \code{na.rm = TRUE}}
}
\value{
a data.frame in a long interval range format
}
\description{
Transform data from a format that is based on predictive samples to a format
based on plain quantiles.
}
\examples{
sample_to_quantile(example_integer)
}
\keyword{data-handling}
