% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_binary.R
\name{score.scoringutils_binary}
\alias{score.scoringutils_binary}
\title{Evaluate forecasts in a Binary Format}
\usage{
\method{score}{scoringutils_binary}(x, metrics = NULL, ...)
}
\arguments{
\item{x}{PLACEHOLDER}

\item{metrics}{the metrics you want to have in the output. If \code{NULL} (the
default), all available metrics will be computed. For a list of available
metrics see \code{\link[=available_metrics]{available_metrics()}}, or  check the \link{metrics} data set.}

\item{...}{additional parameters passed down to \code{\link[=score.scoringutils_quantile]{score.scoringutils_quantile()}} (internal
function used for scoring forecasts in a quantile-based format).}
}
\value{
A data.table with appropriate scores. For more information see
\code{\link[=score]{score()}}.
}
\description{
Evaluate forecasts in a Binary Format
}
\author{
Nikos Bosse \email{nikosbosse@gmail.com}
}
\keyword{scoring}
