% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_continuous_integer.R
\name{score.scoringutils_sample}
\alias{score.scoringutils_sample}
\title{Evaluate forecasts in a Sample-Based Format (Integer or Continuous)}
\usage{
\method{score}{scoringutils_sample}(x, metrics = NULL, ...)
}
\arguments{
\item{x}{PLACEHOLDER}

\item{metrics}{the metrics you want to have in the output. If \code{NULL} (the
default), all available metrics will be computed. For a list of available
metrics see \code{\link[=available_metrics]{available_metrics()}}, or  check the \link{metrics} data set.}

\item{...}{additional parameters passed down to \code{\link[=score.scoringutils_quantile]{score.scoringutils_quantile()}} (internal
function used for scoring forecasts in a quantile-based format).}
}
\value{
A data.table with appropriate scores. For more information see
\code{\link[=score]{score()}}
}
\description{
Evaluate forecasts in a Sample-Based Format (Integer or Continuous)
}
\references{
Funk S, Camacho A, Kucharski AJ, Lowe R, Eggo RM, Edmunds WJ
(2019) Assessing the performance of real-time epidemic forecasts: A
case study of Ebola in the Western Area region of Sierra Leone, 2014-15.
PLoS Comput Biol 15(2): e1006785. \doi{10.1371/journal.pcbi.1006785}
}
\author{
Nikos Bosse \email{nikosbosse@gmail.com}
}
\keyword{scoring}
