% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_continuous_integer.R
\name{score_sample}
\alias{score_sample}
\title{Evaluate forecasts in a Sample-Based Format (Integer or Continuous)}
\usage{
score_sample(data, forecast_unit, metrics, prediction_type)
}
\arguments{
\item{data}{A data.frame or data.table with the predictions and observations.
Note: it is easiest to have a look at the example files provided in the
package and in the examples below.
The following columns need to be present:
\itemize{
\item \code{true_value} - the true observed values
\item \code{prediction} - predictions or predictive samples for one
true value. (You only don't need to provide a prediction column if
you want to score quantile forecasts in a wide range format.)}
For integer and continuous forecasts a \code{sample} column is needed:
\itemize{
\item \code{sample} - an index to identify the predictive samples in the
prediction column generated by one model for one true value. Only
necessary for continuous and integer forecasts, not for
binary predictions.}
For a quantile-format forecast you should provide:
\itemize{
\item \code{prediction}: prediction to the corresponding quantile
\item \code{quantile}: quantile to which the prediction corresponds
}}

\item{forecast_unit}{A character vector with the column names that define
the unit of a single forecast, i.e. a forecast was made for a combination
of the values in \code{forecast_unit}}

\item{metrics}{the metrics you want to have in the output. If \code{NULL} (the
default), all available metrics will be computed. For a list of available
metrics see \code{\link[=available_metrics]{available_metrics()}}}

\item{prediction_type}{character, should be either "continuous" or "integer"}
}
\value{
A data.table with appropriate scores. For more information see
\code{\link[=score]{score()}}
}
\description{
Evaluate forecasts in a Sample-Based Format (Integer or Continuous)
}
\references{
Funk S, Camacho A, Kucharski AJ, Lowe R, Eggo RM, Edmunds WJ
(2019) Assessing the performance of real-time epidemic forecasts: A
case study of Ebola in the Western Area region of Sierra Leone, 2014-15.
PLoS Comput Biol 15(2): e1006785. \url{doi:10.1371/journal.pcbi.1006785}
}
\author{
Nikos Bosse \email{nikosbosse@gmail.com}
}
