% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{score_table}
\alias{score_table}
\title{Plot Coloured Score Table}
\usage{
score_table(summarised_scores, select_metrics = NULL)
}
\arguments{
\item{summarised_scores}{A data.frame of summarised scores as produced by
\code{\link{eval_forecasts}}}

\item{select_metrics}{A character vector with the metrics to show. If set to
\code{NULL} (default), all metrics present in \code{summarised_scores} will
be shown}
}
\value{
A ggplot2 object with a coloured table of summarised scores
}
\description{
Plots a coloured table of summarised scores obtained using
\code{\link{eval_forecasts}}
}
\examples{
scores <- scoringutils::eval_forecasts(scoringutils::quantile_example_data_wide,
                                       by = c("model", "id", "horizon"),
                                       summarise_by = "model")
scoringutils::score_table(scores)
}
