% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoringRules_wrappers.R
\name{scoringRules_checks}
\alias{scoringRules_checks}
\title{Check Inputs For scoringRules Wrapper Functions}
\usage{
scoringRules_checks(predictions, true_values)
}
\arguments{
\item{predictions}{nxN matrix of predictive samples, n (number of rows) being
the number of data points and N (number of columns) the
number of Monte Carlo samples}

\item{true_values}{A vector with the true observed values of size n}
}
\description{
Helper fucntion to check inputs for \code{scoringRules} functions
}
\keyword{internal}
