% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-sample.R
\name{se_mean_sample}
\alias{se_mean_sample}
\title{Squared Error of the Mean (Sample-based Version)}
\usage{
se_mean_sample(observed, predicted)
}
\arguments{
\item{observed}{A vector with observed values of size n}

\item{predicted}{nxN matrix of predictive samples, n (number of rows) being
the number of data points and N (number of columns) the number of Monte
Carlo samples. Alternatively, \code{predicted} can just be a vector of size n.}
}
\value{
vector with the scoring values
}
\description{
Squared error of the mean calculated as

\deqn{
  \textrm{mean}(\textrm{observed} - \textrm{prediction})^2
}{
  mean(observed - mean_prediction)^2
}
}
\examples{
observed <- rnorm(30, mean = 1:30)
predicted_values <- rnorm(30, mean = 1:30)
se_mean_sample(observed, predicted_values)
}
\seealso{
\code{\link[=squared_error]{squared_error()}}
}
\keyword{metric}
