% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience-functions.R
\name{set_forecast_unit}
\alias{set_forecast_unit}
\title{Set unit of a single forecast manually}
\usage{
set_forecast_unit(data, forecast_unit)
}
\arguments{
\item{data}{A data.frame or data.table with the following columns:
\itemize{
\item \code{observed} - the observed values
\item \code{predicted} - predictions, predictive samples or predictive quantiles
\item \code{model} - name of the model or forecaster who made a prediction
}

Depending on the forecast type, one of the following columns may be required:
\itemize{
\item \code{sample_id} - index for the predictive samples in the 'predicted' column
\item \code{quantile}: quantile-level of the corresponding value in \code{predicted}
}

For more information see the vignettes and the example data
(\link{example_quantile}, \link{example_continuous},
\link{example_integer}, \code{\link[=example_point]{example_point()}}, and \link{example_binary}).}

\item{forecast_unit}{Character vector with the names of the columns that
uniquely identify a single forecast.}
}
\value{
A data.table with only those columns kept that are relevant to
scoring or denote the unit of a single forecast as specified by the user.
}
\description{
Helper function to set the unit of a single forecast (i.e. the
combination of columns that uniquely define a single forecast) manually.
This simple function keeps the columns specified in \code{forecast_unit} (plus
additional protected columns, e.g. for observed values, predictions or
quantile levels) and removes duplicate rows.
If not done manually, \code{scoringutils} attempts to determine the unit
of a single forecast automatically by simply assuming that all column names
are relevant to determine the forecast unit. This may lead to unexpected
behaviour, so setting the forecast unit explicitly can help make the code
easier to debug and easier to read. When used as part of a workflow,
\code{set_forecast_unit()} can be directly piped into \code{validate()} to
check everything is in order.
}
\examples{
set_forecast_unit(
  example_quantile,
  c("location", "target_end_date", "target_type", "horizon", "model")
)
}
\keyword{data-handling}
