% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-forecast-multivariate-sample.R
\name{set_grouping}
\alias{set_grouping}
\title{Set grouping}
\usage{
set_grouping(data, joint_across)
}
\arguments{
\item{data}{A data.frame (or similar) with predicted and observed values.
See the details section of for additional information
on the required input format.}

\item{joint_across}{Character vector with columns names that define the
variables which are forecasted jointly. Conceptually, several univariate
forecasts are pooled together to form a single multivariate forecasts.
For example, if you have a column \code{country} and want to define
a multivariate forecast for several countries at once, you could set
\code{joint_across = "country"}.}
}
\value{
A data.table with an additional column \code{.mv_group_id} that
contains the group id for each row.
}
\description{
Helper function to set the grouping of a forecast.
}
\keyword{internal}
