% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-point.R
\name{squared_error}
\alias{squared_error}
\title{Squared Error}
\usage{
squared_error(observed, predicted)
}
\arguments{
\item{observed}{A vector with the observed values of size n}

\item{predicted}{A vector with predicted values of size n}
}
\value{
vector with the scoring values
}
\description{
Squared Error SE calculated as

\deqn{
  (\textrm{observed} - \textrm{predicted})^2
}{
  (observed - predicted)^2
}
}
\examples{
observed <- rnorm(30, mean = 1:30)
predicted_values <- rnorm(30, mean = 1:30)
squared_error(observed, predicted_values)
}
\keyword{metric}
