% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip_attributes}
\alias{strip_attributes}
\title{Strip attributes from an object}
\usage{
strip_attributes(object, attributes)
}
\arguments{
\item{object}{An object to remove attributes from}

\item{attributes}{A character vector of attribute names to remove from the
object}
}
\value{
The object with attributes removed
}
\description{
This function removes all attributes from an object that are
specified in the \code{attributes} argument.
}
\keyword{internal}
