% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-forecast-type-score().R
\name{test_columns_not_present}
\alias{test_columns_not_present}
\title{Test whether column names are NOT present in a data.frame}
\usage{
test_columns_not_present(data, columns)
}
\arguments{
\item{data}{A data.frame or similar to be checked}

\item{columns}{names of columns to be checked}
}
\value{
Returns TRUE if none of the columns are present and FALSE otherwise
}
\description{
Test whether column names are NOT present in a data.frame
}
\keyword{internal}
