% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_-functions.R
\name{test_forecast_type_is_nominal}
\alias{test_forecast_type_is_nominal}
\title{Test whether data could be a nominal forecast.}
\usage{
test_forecast_type_is_nominal(data)
}
\arguments{
\item{data}{A data.frame or similar to be checked}
}
\value{
Returns TRUE if basic requirements are satisfied and FALSE otherwise
}
\description{
Checks type of the necessary columns.
}
\keyword{internal_input_check}
