% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-input-helpers.R
\name{test_has_attribute}
\alias{test_has_attribute}
\title{Test whether an attribute is present}
\usage{
test_has_attribute(object, attribute)
}
\arguments{
\item{data}{An object to be checked}
}
\value{
Returns TRUE if attribute is there and FALSE otherwise
a string otherwise
}
\description{
Tests whether an object has an attribute
}
\keyword{check-inputs}
