% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-package-interfaces.R
\name{to_yardstick_binary_class_prob}
\alias{to_yardstick_binary_class_prob}
\title{Convert to yardstick format for class probability predictions}
\usage{
to_yardstick_binary_class_prob(binary_predictions)
}
\arguments{
\item{binary_predictions}{A data frame of binary predictions following the
same format used for \code{\link[=score]{score()}}. \code{to_yardstick_binary_class} can also be
called on the output of \code{\link[=score]{score()}}.}
}
\value{
A data.table that conforms to the formatting requirements of
\code{yardstick}.
}
\description{
A function to convert from the format for binary forecasts that
\code{scoringutils} uses to the one used by \code{yardstick} for class
probability predictions.

The format \code{yardstick} uses for (binary) class probability predictions is
very similar to the one used by \code{scoringutils}. The only difference is that
\code{yardstick} expects the outcome to be a factor. The function merely
converts "true_value" to a factor.

See \link{example_binary} and
\url{https://yardstick.tidymodels.org/articles/metric-types.html} for
more details on the formats used.
}
\examples{


\dontrun{
library(yardstick)
library(dplyr)
ex <- to_yardstick_binary_class_prob(example_binary)

ex |>
  group_by(model) |>
  filter(!is.na(prediction)) |>
  average_precision(truth = true_value, prediction, event_level = "first")

}
}
\keyword{data-handling}
