% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_list}
\alias{update_list}
\title{Update a List}
\usage{
update_list(defaults = list(), optional = list())
}
\arguments{
\item{defaults}{A list of default settings}

\item{optional}{A list of optional settings to override defaults}
}
\value{
A list
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Used to handle updating settings in a list. For example when making
changes to \code{interval_score_arguments} in \code{eval_forecasts()}
}
\keyword{internal}
