% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{validate_forecast}
\alias{validate_forecast}
\title{Re-validate an existing forecast object}
\usage{
validate_forecast(forecast, forecast_type = NULL, verbose = TRUE)
}
\arguments{
\item{forecast}{A forecast object (a validated data.table with predicted and
observed values, see \code{\link[=as_forecast]{as_forecast()}})}

\item{forecast_type}{(optional) The forecast type you expect the forecasts
to have. If the forecast type as determined by \code{scoringutils} based on the
input does not match this, an error will be thrown. If \code{NULL} (the
default), the forecast type will be inferred from the data.}

\item{verbose}{Logical. If \code{FALSE} (default is \code{TRUE}), no messages and
warnings will be created.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
The function re-validates an existing forecast object. It is similar to
\code{\link[=assert_forecast]{assert_forecast()}}, but returns the input data instead of an invisible
\code{NULL}. See \code{\link[=as_forecast]{as_forecast()}} for details on the expected input formats.
}
\examples{
forecast <- as_forecast_binary(example_binary)
assert_forecast(forecast)
}
\keyword{check-forecasts}
