% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{validate_general}
\alias{validate_general}
\title{Validation common to all forecast types}
\usage{
validate_general(data, silent = FALSE)
}
\arguments{
\item{data}{A data.table with forecasts and observed values that should
be validated.}
}
\value{
returns the input
}
\description{
The function runs input checks that apply to all input data, regardless of
forecast type. The function
\itemize{
\item asserts that the forecast is a data.table which has columns \code{observed} and
\code{predicted}, as well as a column called \code{model}.
\item checks the forecast type and forecast unit
\item checks there are no duplicate forecasts
\item if appropriate, checks the number of samples / quantiles is the same
for all forecasts.
}
}
\keyword{internal_input_check}
