% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{validate_metrics}
\alias{validate_metrics}
\title{Validate metrics}
\usage{
validate_metrics(metrics)
}
\arguments{
\item{metrics}{A named list with metrics. Every element should be a scoring
function to be applied to the data.}
}
\value{
A named list of metrics, with those filtered out that are not
valid functions
}
\description{
This function validates whether the list of metrics is a list
of valid functions.

The function is used in \code{\link[=score]{score()}} to make sure that all metrics are valid
functions.
}
\keyword{internal_input_check}
