% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_rules}
\alias{validate_rules}
\title{Validate scoring rules}
\usage{
validate_rules(rules)
}
\arguments{
\item{rules}{A named list with scoring rules. Every element should be a
scoring function to be applied to the data.}
}
\value{
A named list of rules, with those filtered out that are not
valid functions
}
\description{
This function validates whether the list of scoring rules is a
list of valid functions.

The function is used in \code{\link[=score]{score()}} to make sure that all scoring rules are
valid functions
}
\keyword{internal_input_check}
