% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{wis_components}
\alias{wis_components}
\title{Plot Contributions to the Weighted Interval Score}
\usage{
wis_components(
  scores,
  x = "model",
  group = NULL,
  relative_contributions = FALSE,
  x_text_angle = 90,
  xlab = x,
  ylab = "WIS contributions"
)
}
\arguments{
\item{scores}{A data.frame of scores based on quantile forecasts as
produced by \code{\link[=score]{score()}}}

\item{x}{The variable from the scores you want to show on the x-Axis.
Usually this will be "model"}

\item{group}{Choose a grouping variable for the plot that gets directly
passed down to ggplot. Default is \code{NULL}}

\item{relative_contributions}{show relative contributions instead of absolute
contributions. Default is FALSE and this functionality is not available yet.}

\item{x_text_angle}{Angle for the text on the x-axis. Default is 90}

\item{xlab}{Label for the x-axis. Default is the variable name on the x-axis}

\item{ylab}{Label for the y-axis. Default is "WIS contributions"}
}
\value{
A ggplot2 object showing a contributions from the three components of
the weighted interval score
}
\description{
Visualise the components of the weighted interval score: penalties for
over-prediction, under-prediction and for high dispersion (lack of sharpness)
}
\examples{
library("scoringutils")
library(ggplot2)
scores <- score(example_quantile)
scores <- summarise_scores(scores, by = c("model", "target_type"))

wis_components(scores, x = "model",
              relative_contributions = TRUE) +
  facet_wrap(~ target_type)
wis_components(scores, x = "model",
              relative_contributions = FALSE) +
  facet_wrap(~ target_type)
}
\references{
Bracher J, Ray E, Gneiting T, Reich, N (2020) Evaluating epidemic forecasts
in an interval format. \url{https://arxiv.org/abs/2005.12881}
}
