% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-point.R
\name{abs_error}
\alias{abs_error}
\title{Absolute Error}
\usage{
abs_error(observed, predicted)
}
\arguments{
\item{observed}{numeric vector of size n with the observed values}

\item{predicted}{numeric nxN matrix of predictive
quantiles, n (number of rows) being the number of forecasts (corresponding
to the number of observed values) and N
(number of columns) the number of quantiles per forecast.
If \code{observed} is just a single number, then predicted can just be a
vector of size N.}
}
\value{
vector with the absolute error
}
\description{
Calculate absolute error as

\deqn{
  \textrm{abs}(\textrm{observed} - \textrm{median\_prediction})
}{
  abs(observed - prediction)
}
}
\examples{
observed <- rnorm(30, mean = 1:30)
predicted_values <- rnorm(30, mean = 1:30)
abs_error(observed, predicted_values)
}
\seealso{
\code{\link[=ae_median_sample]{ae_median_sample()}}, \code{\link[=ae_median_quantile]{ae_median_quantile()}}
}
\keyword{metric}
