% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_forecasts_helper.R
\name{add_quantiles}
\alias{add_quantiles}
\title{Add Quantiles to Predictions When Summarising}
\usage{
add_quantiles(dt, varnames, quantiles, by)
}
\arguments{
\item{dt}{the data.table operated on}

\item{varnames}{names of the variables for which to calculate quantiles}

\item{quantiles}{the desired quantiles}

\item{by}{grouping variable in `eval_forecasts()}
}
\value{
`data.table` with quantiles added
}
\description{
Helper function used within eval_forecasts
}
