% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise-comparisons.R
\name{add_relative_skill}
\alias{add_relative_skill}
\title{Add pairwise comparisons}
\usage{
add_relative_skill(
  scores,
  by = "model",
  metric = intersect(c("wis", "crps", "brier_score"), names(scores)),
  baseline = NULL
)
}
\description{
Adds a columns with relative skills computed by running
pairwise comparisons on the scores.
For more information on
the computation of relative skill, see \code{\link[=get_pairwise_comparisons]{get_pairwise_comparisons()}}.
Relative skill will be calculated for the aggregation level specified in
\code{by}.
}
\keyword{keyword}
\keyword{scoring}
