% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_forecasts_helper.R
\name{add_sd}
\alias{add_sd}
\title{Add Standard Deviation to Predictions When Summarising}
\usage{
add_sd(dt, varnames, by)
}
\arguments{
\item{dt}{the data.table operated on}

\item{varnames}{names of the variables for which to calculate the sd}

\item{by}{grouping variable in `eval_forecasts()}
}
\value{
`data.table` with sd added
}
\description{
Helper function used within eval_forecasts
}
