% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-quantile.R
\name{ae_median_quantile}
\alias{ae_median_quantile}
\title{Absolute Error of the Median (Quantile-based Version)}
\usage{
ae_median_quantile(observed, predicted, quantile)
}
\arguments{
\item{observed}{numeric vector of size n with the observed values}

\item{predicted}{numeric nxN matrix of predictive
quantiles, n (number of rows) being the number of forecasts (corresponding
to the number of observed values) and N
(number of columns) the number of quantiles per forecast.
If \code{observed} is just a single number, then predicted can just be a
vector of size N.}

\item{quantile}{vector with quantile levels of size N}
}
\value{
numeric vector of length N with the absolute error of the median
}
\description{
Compute the absolute error of the median calculated as
\deqn{
  \textrm{abs}(\textrm{observed} - \textrm{median prediction})
}{
  abs(observed - median_prediction)
}
The median prediction is the predicted value for which quantile == 0.5,
the function therefore requires 0.5 to be among the quantile levels in
\code{quantile}.
}
\examples{
observed <- rnorm(30, mean = 1:30)
predicted_values <- matrix(rnorm(30, mean = 1:30))
ae_median_quantile(observed, predicted_values, quantile = 0.5)
}
\seealso{
\code{\link[=ae_median_sample]{ae_median_sample()}}
}
\keyword{rule}
