% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{apply_metrics}
\alias{apply_metrics}
\title{Apply a list of functions to a data table of forecasts}
\usage{
apply_metrics(forecast, metrics, ...)
}
\arguments{
\item{forecast}{A forecast object (a validated data.table with predicted and
observed values, see \code{\link[=as_forecast]{as_forecast()}})}

\item{metrics}{A named list of scoring functions. Names will be used as
column names in the output. See \code{\link[=metrics_point]{metrics_point()}}, \code{\link[=metrics_binary]{metrics_binary()}},
\code{\link[=metrics_quantile]{metrics_quantile()}}, and \code{\link[=metrics_sample]{metrics_sample()}} for more information on the
default metrics used.}

\item{...}{Additional arguments to be passed to the scoring rules.}
}
\value{
A data table with the forecasts and the calculated metrics.
}
\description{
This helper function applies scoring rules (stored as a list of
functions) to a data table of forecasts. \code{apply_metrics} is used within
\code{score()} to apply all scoring rules to the data.
Scoring rules are wrapped in \code{\link[=run_safely]{run_safely()}} to catch errors and to make
sure that only arguments are passed to the scoring rule that are actually
accepted by it.
}
\keyword{internal}
