% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{apply_metrics}
\alias{apply_metrics}
\title{Apply A List Of Functions To A Data Table Of Forecasts}
\usage{
apply_metrics(data, metrics, ...)
}
\arguments{
\item{data}{A data.frame or data.table with predicted and observed values.}

\item{metrics}{A named list of scoring functions. Names will be used as
column names in the output. See \code{\link[=metrics_point]{metrics_point()}}, \code{\link[=metrics_binary]{metrics_binary()}},
\code{metrics_quantile()}, and \code{\link[=metrics_sample]{metrics_sample()}} for more information on the
default metrics used.}

\item{...}{additional arguments}
}
\value{
A data table with the forecasts and the calculated metrics
}
\description{
This helper function applies scoring rules (stored as a list of
functions) to a data table of. \code{apply_metrics} is used within \code{score()} to
apply all scoring rules to the data.
Scoring rules are wrapped in \code{\link[=run_safely]{run_safely()}} to catch errors and to make
sure that only arguments are passed to the scoring rule that are actually
accepted by it.
}
