% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{as_forecast}
\alias{as_forecast}
\title{Create a \code{forecast} Object}
\usage{
as_forecast(data)
}
\arguments{
\item{data}{A data.frame or data.table with predicted and observed values.}
}
\value{
Depending on the forecast type, an object of class
\code{scoringutils_binary}, \code{scoringutils_point}, \code{scoringutils_sample} or
\code{scoringutils_quantile}.
}
\description{
Convert a data.frame or similar of forecasts into an object of
class \verb{forecast_*} and validate it.

\code{as_forecast()} determines the forecast type (binary, point, sample-based or
quantile-based) from the input data (using the function
\code{\link[=get_forecast_type]{get_forecast_type()}}. It then constructs an object of the
appropriate class (\code{forecast_binary}, \code{forecast_point}, \code{forecast_sample}, or
\code{forecast_quantile}, using the function \code{\link[=new_scoringutils]{new_scoringutils()}}).
Lastly, it calls \code{\link[=as_forecast]{as_forecast()}} on the object to make sure it conforms with
the required input formats.
}
\section{Forecast types and input format}{
Various different forecast types / forecast formats are supported. At the
moment, those are
\itemize{
\item point forecasts
\item binary forecasts ("soft binary classification")
\item Probabilistic forecasts in a quantile-based format (a forecast is
represented as a set of predictive quantiles)
\item Probabilistic forecasts in a sample-based format (a forecast is represented
as a set of predictive samples)
}

Forecast types are determined based on the columns present in the input data.

\emph{Point forecasts} require a column \code{observed} of type numeric and a column
\code{predicted} of type numeric.

\emph{Binary forecasts} require a column \code{observed} of type factor with exactly
two levels and a column \code{predicted} of type numeric with probabilities,
corresponding to the probability that \code{observed} is equal to the second
factor level. See details \link[=brier_score]{here} for more information.

\emph{Quantile-based forecasts} require a column \code{observed} of type numeric,
a column \code{predicted} of type numeric, and a column \code{quantile} of type numeric
with quantile-levels (between 0 and 1).

\emph{Sample-based forecasts} require a column \code{observed} of type numeric,
a column \code{predicted} of type numeric, and a column \code{sample_id} of type
numeric with sample indices.

For more information see the vignettes and the example data
(\link{example_quantile}, \link{example_continuous}, \link{example_integer},
\code{\link[=example_point]{example_point()}}, and \link{example_binary}).
}

\examples{
as_forecast(example_binary)
as_forecast(example_quantile)
}
\keyword{check-forecasts}
