% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{as_forecast_binary}
\alias{as_forecast_binary}
\title{Create a \code{forecast} object for binary forecasts}
\usage{
as_forecast_binary(
  data,
  forecast_unit = NULL,
  observed = NULL,
  predicted = NULL,
  model = NULL
)
}
\description{
Create a \code{forecast} object for binary forecasts
}
