% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{as_forecast_nominal}
\alias{as_forecast_nominal}
\title{Create a \code{forecast} object for nominal forecasts}
\usage{
as_forecast_nominal(
  data,
  forecast_unit = NULL,
  observed = NULL,
  predicted = NULL,
  predicted_label = NULL
)
}
\arguments{
\item{data}{A data.frame (or similar) with predicted and observed values.
See the details section of \code{\link[=as_forecast]{as_forecast()}} for additional information
on required input formats.}

\item{forecast_unit}{(optional) Name of the columns in \code{data} (after
any renaming of columns) that denote the unit of a
single forecast. See \code{\link[=get_forecast_unit]{get_forecast_unit()}} for details.
If \code{NULL} (the default), all columns that are not required columns are
assumed to form the unit of a single forecast. If specified, all columns
that are not part of the forecast unit (or required columns) will be removed.}

\item{observed}{(optional) Name of the column in \code{data} that contains the
observed values. This column will be renamed to "observed".}

\item{predicted}{(optional) Name of the column in \code{data} that contains the
predicted values. This column will be renamed to "predicted".}

\item{predicted_label}{(optional) Name of the column in \code{data} that denotes
the outcome to which a predicted probability corresponds to.
This column will be renamed to "predicted_label". Only applicable to
nominal forecasts.}
}
\description{
Nominal forecasts are a form of categorical forecasts where the possible
outcomes that the observed values can assume are not ordered. In that sense,
Nominal forecasts represent a generalisation of binary forecasts.
}
\seealso{
Other functions to create forecast objects: 
\code{\link{as_forecast}},
\code{\link{as_forecast_binary}()},
\code{\link{as_forecast_point}()},
\code{\link{as_forecast_quantile}()},
\code{\link{as_forecast_sample}()}
}
\concept{functions to create forecast objects}
\keyword{as_forecast}
