% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-forecast-point.R
\name{as_forecast_point}
\alias{as_forecast_point}
\alias{as_forecast_point.default}
\title{Create a \code{forecast} object for point forecasts}
\usage{
as_forecast_point(data, ...)

\method{as_forecast_point}{default}(
  data,
  forecast_unit = NULL,
  observed = NULL,
  predicted = NULL,
  ...
)
}
\arguments{
\item{data}{A data.frame (or similar) with predicted and observed values.
See the details section of for additional information
on the required input format.}

\item{...}{Unused}

\item{forecast_unit}{(optional) Name of the columns in \code{data} (after
any renaming of columns) that denote the unit of a
single forecast. See \code{\link[=get_forecast_unit]{get_forecast_unit()}} for details.
If \code{NULL} (the default), all columns that are not required columns are
assumed to form the unit of a single forecast. If specified, all columns
that are not part of the forecast unit (or required columns) will be removed.}

\item{observed}{(optional) Name of the column in \code{data} that contains the
observed values. This column will be renamed to "observed".}

\item{predicted}{(optional) Name of the column in \code{data} that contains the
predicted values. This column will be renamed to "predicted".}
}
\value{
A \code{forecast} object of class \code{forecast_point}
}
\description{
Process and validate a data.frame (or similar) or similar with forecasts
and observations. If the input passes all input checks, those functions will
be converted to a \code{forecast} object. A forecast object is a \code{data.table} with
a class \code{forecast} and an additional class that depends on the forecast type.

The arguments \code{observed}, \code{predicted}, etc. make it possible to rename
existing columns of the input data to match the required columns for a
forecast object. Using the argument \code{forecast_unit}, you can specify
the columns that uniquely identify a single forecast (and thereby removing
other, unneeded columns. See section "Forecast Unit" below for details).
}
\section{Target format}{
The input for all further scoring needs to be a data.frame or similar with
the following columns:
\itemize{
\item \code{observed}: Column of type \code{numeric} with observed values.
\item \code{predicted}: Column of type \code{numeric} with predicted values.
}

For convenience, we recommend an additional column \code{model} holding the name
of the forecaster or model that produced a prediction, but this is not
strictly necessary.

See the \link{example_point} data set for an example.
}

\seealso{
Other functions to create forecast objects: 
\code{\link{as_forecast_binary}()},
\code{\link{as_forecast_multivariate_sample}()},
\code{\link{as_forecast_nominal}()},
\code{\link{as_forecast_ordinal}()},
\code{\link{as_forecast_point.forecast_quantile}()},
\code{\link{as_forecast_quantile}()},
\code{\link{as_forecast_quantile.forecast_sample}()},
\code{\link{as_forecast_sample}()}
}
\concept{functions to create forecast objects}
\keyword{as_forecast}
\keyword{transform}
