% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{as_forecast_quantile}
\alias{as_forecast_quantile}
\alias{as_forecast_quantile.default}
\alias{as_forecast_quantile.forecast_sample}
\title{Create a \code{forecast} object for quantile-based forecasts}
\usage{
as_forecast_quantile(data, ...)

\method{as_forecast_quantile}{default}(
  data,
  forecast_unit = NULL,
  observed = NULL,
  predicted = NULL,
  model = NULL,
  quantile_level = NULL
)

\method{as_forecast_quantile}{forecast_sample}(
  data,
  probs = c(0.05, 0.25, 0.5, 0.75, 0.95),
  type = 7,
  ...
)
}
\arguments{
\item{quantile_level}{(optional) Name of the column in \code{data} that contains
the quantile level of the predicted values. This column will be renamed to
"quantile_level". Only applicable to quantile-based forecasts.}

\item{probs}{A numeric vector of quantile levels for which
quantiles will be computed. Corresponds to the \code{probs} argument in
\code{\link[=quantile]{quantile()}}.}

\item{type}{Type argument passed down to the quantile function. For more
information, see \code{\link[=quantile]{quantile()}}.}
}
\description{
Create a \code{forecast} object for quantile-based forecasts. See more information
on forecast types and expected input formats by calling \verb{?}\code{\link[=as_forecast]{as_forecast()}}.
}
\seealso{
\code{\link[=as_forecast_point]{as_forecast_point()}}, \code{\link[=as_forecast_binary]{as_forecast_binary()}},
\code{\link[=as_forecast_sample]{as_forecast_sample()}}, \code{\link[=as_forecast_quantile]{as_forecast_quantile()}}
}
