% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{as_forecast_quantile}
\alias{as_forecast_quantile}
\alias{as_forecast_quantile.default}
\alias{as_forecast_quantile.forecast_sample}
\title{Create a \code{forecast} object for quantile-based forecasts}
\usage{
as_forecast_quantile(data, ...)

\method{as_forecast_quantile}{default}(
  data,
  forecast_unit = NULL,
  observed = NULL,
  predicted = NULL,
  model = NULL,
  quantile_level = NULL,
  ...
)

\method{as_forecast_quantile}{forecast_sample}(
  data,
  probs = c(0.05, 0.25, 0.5, 0.75, 0.95),
  type = 7,
  ...
)
}
\arguments{
\item{data}{A data.frame (or similar) with predicted and observed values.
See the details section of \code{\link[=as_forecast]{as_forecast()}} for additional information
on required input formats.}

\item{...}{Unused}

\item{forecast_unit}{(optional) Name of the columns in \code{data} (after
any renaming of columns) that denote the unit of a
single forecast. See \code{\link[=get_forecast_unit]{get_forecast_unit()}} for details.
If \code{NULL} (the default), all columns that are not required columns are
assumed to form the unit of a single forecast. If specified, all columns
that are not part of the forecast unit (or required columns) will be removed.}

\item{observed}{(optional) Name of the column in \code{data} that contains the
observed values. This column will be renamed to "observed".}

\item{predicted}{(optional) Name of the column in \code{data} that contains the
predicted values. This column will be renamed to "predicted".}

\item{model}{(optional) Name of the column in \code{data} that contains the names
of the models/forecasters that generated the predicted values.
This column will be renamed to "model".}

\item{quantile_level}{(optional) Name of the column in \code{data} that contains
the quantile level of the predicted values. This column will be renamed to
"quantile_level". Only applicable to quantile-based forecasts.}

\item{probs}{A numeric vector of quantile levels for which
quantiles will be computed. Corresponds to the \code{probs} argument in
\code{\link[=quantile]{quantile()}}.}

\item{type}{Type argument passed down to the quantile function. For more
information, see \code{\link[=quantile]{quantile()}}.}
}
\description{
Create a \code{forecast} object for quantile-based forecasts. See more information
on forecast types and expected input formats by calling \verb{?}\code{\link[=as_forecast]{as_forecast()}}.

When creating a \code{forecast_quantile} object from a \code{forecast_sample} object,
the quantiles are estimated by computing empircal quantiles from the samples
via \code{\link[=quantile]{quantile()}}. Note that empirical quantiles are a biased estimator for
the true quantiles in particular in the tails of the distribution and
when the number of available samples is low.
}
\seealso{
\code{\link[=as_forecast]{as_forecast()}}, \code{\link[=as_forecast_point]{as_forecast_point()}}, \code{\link[=as_forecast_binary]{as_forecast_binary()}},
\code{\link[=as_forecast_sample]{as_forecast_sample()}}, \code{\link[=as_forecast_quantile]{as_forecast_quantile()}}
}
