% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{as_forecast_sample}
\alias{as_forecast_sample}
\title{Create a \code{forecast} object for sample-based forecasts}
\usage{
as_forecast_sample(
  data,
  forecast_unit = NULL,
  observed = NULL,
  predicted = NULL,
  model = NULL,
  sample_id = NULL
)
}
\arguments{
\item{sample_id}{(optional) Name of the column in \code{data} that contains the
sample id. This column will be renamed to "sample_id". Only applicable to
sample-based forecasts.}
}
\description{
Create a \code{forecast} object for sample-based forecasts
}
\seealso{
\code{\link[=as_forecast_point]{as_forecast_point()}}, \code{\link[=as_forecast_binary]{as_forecast_binary()}},
\code{\link[=as_forecast_sample]{as_forecast_sample()}}, \code{\link[=as_forecast_quantile]{as_forecast_quantile()}}
}
