% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{as_forecast_sample}
\alias{as_forecast_sample}
\title{Create a \code{forecast} object for sample-based forecasts}
\usage{
as_forecast_sample(
  data,
  forecast_unit = NULL,
  observed = NULL,
  predicted = NULL,
  model = NULL,
  sample_id = NULL
)
}
\arguments{
\item{data}{A data.frame (or similar) with predicted and observed values.
See the details section of \code{\link[=as_forecast]{as_forecast()}} for additional information
on required input formats.}

\item{forecast_unit}{(optional) Name of the columns in \code{data} (after
any renaming of columns) that denote the unit of a
single forecast. See \code{\link[=get_forecast_unit]{get_forecast_unit()}} for details.
If \code{NULL} (the default), all columns that are not required columns are
assumed to form the unit of a single forecast. If specified, all columns
that are not part of the forecast unit (or required columns) will be removed.}

\item{observed}{(optional) Name of the column in \code{data} that contains the
observed values. This column will be renamed to "observed".}

\item{predicted}{(optional) Name of the column in \code{data} that contains the
predicted values. This column will be renamed to "predicted".}

\item{sample_id}{(optional) Name of the column in \code{data} that contains the
sample id. This column will be renamed to "sample_id". Only applicable to
sample-based forecasts.}
}
\description{
Create a \code{forecast} object for sample-based forecasts
}
\seealso{
Other functions to create forecast objects: 
\code{\link{as_forecast}},
\code{\link{as_forecast_binary}()},
\code{\link{as_forecast_nominal}()},
\code{\link{as_forecast_point}()},
\code{\link{as_forecast_quantile}()}
}
\concept{functions to create forecast objects}
\keyword{as_forecast}
