% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_quantile.R
\name{as_quantile.forecast_sample}
\alias{as_quantile.forecast_sample}
\title{Convert a sample forecast to a quantile forecast}
\usage{
\method{as_quantile}{forecast_sample}(
  forecast,
  quantile_levels = seq(from = 0.01, to = 0.99, by = 0.01),
  ...
)
}
\arguments{
\item{forecast}{The \verb{forecast_sample`` object to convert to a }forecast_quantile`.}

\item{quantile_levels}{A vector of quantile levels. Defaults to
0.01 to 0.99 by 0.01.}

\item{...}{Additional arguments passing inherited from the default method but
unused.}
}
\value{
A \code{forecast_quantile} object.
}
\description{
This function takes a sample forecast and converts it to a quantile forecast
sample.
}
\examples{
as_quantile(as_forecast(example_sample_continuous))
}
\keyword{check-forecasts}
