% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-forecast-type.R
\name{assert_forecast_type}
\alias{assert_forecast_type}
\title{Assert that forecast type is as expected}
\usage{
assert_forecast_type(data, actual = get_forecast_type(data), desired = NULL)
}
\arguments{
\item{data}{A forecast object.}

\item{actual}{The actual forecast type of the data}

\item{desired}{The desired forecast type of the data}
}
\value{
Returns NULL invisibly if the assertion was successful and throws an
error otherwise.
}
\description{
Assert that forecast type is as expected
}
\keyword{internal_input_check}
