% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-inputs-scoring-functions.R
\name{assert_input_nominal}
\alias{assert_input_nominal}
\title{Assert that inputs are correct for nominal forecasts}
\usage{
assert_input_nominal(observed, predicted, predicted_label)
}
\arguments{
\item{observed}{XXX}

\item{predicted}{XXX}

\item{predicted_label}{XXX}
}
\value{
Returns NULL invisibly if the assertion was successful and throws an
error otherwise.
}
\description{
Function assesses whether the inputs correspond to the
requirements for scoring nominal forecasts.
}
\keyword{internal_input_check}
