% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-input-helpers.R
\name{assure_model_column}
\alias{assure_model_column}
\title{Assure that Data Has a \code{model} Column}
\usage{
assure_model_column(data)
}
\arguments{
\item{data}{A data.frame (or similar) with predicted and observed values.
See \code{\link[=as_forecast]{as_forecast()}} for additional information on input formats.}
}
\value{
The data.table with a column called \code{model}
}
\description{
Check whether the data.table has a column called \code{model}.
If not, a column called \code{model} is added with the value \verb{Unspecified model}.
}
\keyword{internal_input_check}
