% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_forecasts.R
\name{check_clean_data}
\alias{check_clean_data}
\title{Clean forecast data}
\usage{
check_clean_data(data, verbose = TRUE)
}
\arguments{
\item{data}{A data.frame or similar as it gets passed to \code{\link[=eval_forecasts]{eval_forecasts()}}.}

\item{verbose}{Boolean (default is \code{TRUE}), whether or not to print warnings}
}
\value{
A data.table with NA values in \code{true_value} or \code{prediction} removed.
}
\description{
Helper function to check that the input is in fact a data.frame
or similar and remove rows with no value for \code{prediction} or \code{true_value}
}
\keyword{internal}
