% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_forecasts.R
\name{check_columns_present}
\alias{check_columns_present}
\title{Check column names are present in a data.frame}
\usage{
check_columns_present(data, ...)
}
\arguments{
\item{data}{A data.frame or similar to be checked}

\item{...}{names of columns to be checked}
}
\value{
Returns an string with an error message if any of the column names
is not in data, otherwise returns TRUE
}
\description{
Check column names are present in a data.frame
}
\keyword{internal}
