% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-input-helpers.R
\name{check_columns_present}
\alias{check_columns_present}
\title{Check column names are present in a data.frame}
\usage{
check_columns_present(data, columns)
}
\arguments{
\item{data}{A data.frame or similar to be checked}

\item{columns}{A character vector of column names to check}
}
\value{
Returns TRUE if the check was successful and a string with an
error message otherwise.
}
\description{
The functions loops over the column names and checks whether they are
present. If an issue is encountered, the function immediately stops
and returns a message with the first issue encountered.
}
\keyword{internal_input_check}
