% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-input-helpers.R
\name{check_data_columns}
\alias{check_data_columns}
\title{Check whether data is data.frame with correct columns}
\usage{
check_data_columns(data)
}
\arguments{
\item{data}{A data.frame or similar to be checked}
}
\value{
Returns TRUE if the check was successful and a string with an
error message otherwise
}
\description{
Checks whether data is a data.frame, whether columns
"observed" and "predicted" are present, and checks that only one of
"quantile_level" and "sample_id" is present.
}
\keyword{internal_input_check}
