% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-duplicate-forecasts.R
\name{check_duplicates}
\alias{check_duplicates}
\title{Check that there are no duplicate forecasts}
\usage{
check_duplicates(data)
}
\arguments{
\item{data}{A data.frame (or similar) with predicted and observed values.
See the details section of \code{\link[=as_forecast]{as_forecast()}} for additional information
on required input formats.}
}
\value{
Returns TRUE if the check was successful and a string with an
error message otherwise.
}
\description{
Runs \code{\link[=get_duplicate_forecasts]{get_duplicate_forecasts()}} and returns a message if an issue is
encountered
}
\keyword{internal_input_check}
