% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-input-helpers.R
\name{check_duplicates}
\alias{check_duplicates}
\title{Check that there are no duplicate forecasts}
\usage{
check_duplicates(data, forecast_unit = NULL)
}
\arguments{
\item{data}{A data.frame as used for \code{\link[=score]{score()}}}

\item{forecast_unit}{A character vector with the column names that define
the unit of a single forecast. If \code{NULL} (the default) the function tries
to infer the unit of a single forecast.}
}
\value{
Returns an string with an error message if an issue is found,
otherwise returns TRUE
}
\description{
Runs \code{\link[=find_duplicates]{find_duplicates()}} and returns a message if an issue is encountered
}
\keyword{internal}
