% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-input-helpers.R
\name{check_duplicates}
\alias{check_duplicates}
\title{Check that there are no duplicate forecasts}
\usage{
check_duplicates(data, forecast_unit = get_forecast_unit(data))
}
\arguments{
\item{data}{A data.frame as used for \code{\link[=score]{score()}}}

\item{forecast_unit}{A character vector with the column names that define
the unit of a single forecast. By default the forecast unit will be
automatically inferred from the data (see \code{\link[=get_forecast_unit]{get_forecast_unit()}})}
}
\value{
Returns TRUE if the check was successful and a string with an
error message otherwise.
}
\description{
Runs \code{\link[=get_duplicate_forecasts]{get_duplicate_forecasts()}} and returns a message if an issue is
encountered
}
\keyword{internal_input_check}
