% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_forecasts.R
\name{check_forecasts}
\alias{check_forecasts}
\title{Check forecasts}
\usage{
check_forecasts(data)
}
\arguments{
\item{data}{A data.frame or similiar as would be used as input to \code{\link[=eval_forecasts]{eval_forecasts()}}}
}
\value{
some output that tells you what scoringutils thinks you want to do.
}
\description{
missing
}
\details{
missing
}
\examples{

## Probability Forecast for Binary Target
binary_example <- data.table::setDT(scoringutils::binary_example_data)
eval <- scoringutils::eval_forecasts(binary_example,
                                     summarise_by = c("model"),
                                     quantiles = c(0.5), sd = TRUE,
                                     verbose = FALSE)

}
\author{
Nikos Bosse \email{nikosbosse@gmail.com}
}
