% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-input-helpers.R
\name{check_no_NA_present}
\alias{check_no_NA_present}
\title{Check columns in data.frame don't have NA values}
\usage{
check_no_NA_present(data, columns)
}
\arguments{
\item{data}{A data.frame or similar to be checked}

\item{columns}{A character vector of column names to check}
}
\value{
Returns TRUE if the check was successful and a string with an
error message otherwise
}
\description{
Function checks whether any of the columns in a data.frame,
as specified in \code{columns}, have NA values. If so, it returns a string with
an error message, otherwise it returns TRUE.
}
\keyword{internal_input_check}
