% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-forecast.R
\name{check_number_per_forecast}
\alias{check_number_per_forecast}
\title{Check that all forecasts have the same number of rows}
\usage{
check_number_per_forecast(data, forecast_unit)
}
\arguments{
\item{data}{A data.frame or similar to be checked}

\item{forecast_unit}{Character vector denoting the unit of a single forecast.}
}
\value{
Returns TRUE if the check was successful and a string with an
error message otherwise.
}
\description{
Helper function that checks the number of rows (corresponding e.g to
quantiles or samples) per forecast.
If the number of quantiles or samples is the same for all forecasts, it
returns TRUE and a string with an error message otherwise.
}
\keyword{internal_input_check}
