% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-check-helpers.R
\name{check_numeric_vector}
\alias{check_numeric_vector}
\title{Check whether an input is an atomic vector of mode 'numeric'}
\usage{
check_numeric_vector(x, ...)
}
\arguments{
\item{x}{additional arguments to pass to \code{check_numeric()}}
}
\value{
Either TRUE if the test is successful or a string with an error
message
}
\description{
Helper function
}
\keyword{internal}
