% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_helper.R
\name{check_score_params}
\alias{check_score_params}
\title{Check input parameters for \code{\link[=score]{score()}}}
\usage{
check_score_params(
  data,
  forecast_unit,
  metrics,
  summarise_by,
  compute_relative_skill,
  baseline,
  rel_skill_metric
)
}
\arguments{
\item{data}{A data.frame or data.table with the predictions and observations.
Note: it is easiest to have a look at the example files provided in the
package and in the examples below.
The following columns need to be present:
\itemize{
\item \code{true_value} - the true observed values
\item \code{prediction} - predictions or predictive samples for one
true value. (You only don't need to provide a prediction column if
you want to score quantile forecasts in a wide range format.)}
For integer and continuous forecasts a \code{sample} column is needed:
\itemize{
\item \code{sample} - an index to identify the predictive samples in the
prediction column generated by one model for one true value. Only
necessary for continuous and integer forecasts, not for
binary predictions.}
For quantile forecasts the data can be provided in variety of formats. You
can either use a range-based format or a quantile-based format. (You can
convert between formats using \code{\link[=quantile_to_range_long]{quantile_to_range_long()}},
\code{\link[=range_long_to_quantile]{range_long_to_quantile()}},
\code{\link[=sample_to_range_long]{sample_to_range_long()}},
\code{\link[=sample_to_quantile]{sample_to_quantile()}})
For a quantile-format forecast you should provide:
\itemize{
\item \code{prediction}: prediction to the corresponding quantile
\item \code{quantile}: quantile to which the prediction corresponds
For a range format (long) forecast you need
\item \code{prediction}: the quantile forecasts
\item \code{boundary}: values should be either "lower" or "upper", depending
on whether the prediction is for the lower or upper bound of a given range
\item \code{range} the range for which a forecast was made. For a 50\% interval
the range should be 50. The forecast for the 25\% quantile should have
the value in the \code{prediction} column, the value of \code{range}
should be 50 and the value of \code{boundary} should be "lower".
If you want to score the median (i.e. \code{range = 0}), you still
need to include a lower and an upper estimate, so the median has to
appear twice.
Alternatively you can also provide the format in a wide range format.
This format needs:
\item pairs of columns called something like 'upper_90' and 'lower_90',
or 'upper_50' and 'lower_50', where the number denotes the interval range.
For the median, you need to provide columns called 'upper_0' and 'lower_0'
}}

\item{metrics}{the metrics you want to have in the output. If \code{NULL} (the
default), all available metrics will be computed. For a list of available
metrics see \code{\link[=available_metrics]{available_metrics()}}}

\item{summarise_by}{character vector of columns to group the summary by. By
default, this is equal to \code{by} and no summary takes place.
But sometimes you may want to to summarise
over categories different from the scoring.
\code{summarise_by} is also the grouping level used to compute
(and possibly plot) the probability integral transform(pit). Sometimes you
may want to include 'range', 'quantile' or 'sample', to summarise by
range, quantile or sample.}

\item{compute_relative_skill}{logical, whether or not to compute relative
performance between models. If \code{TRUE} (default is \code{FALSE}), then a column called
'model' must be present in the input data. For more information on
the computation of relative skill, see \code{\link[=pairwise_comparison]{pairwise_comparison()}}.
Relative skill will be calculated for the aggregation level specified in
\code{summarise_by}.}

\item{baseline}{character string with the name of a model. If a baseline is
given, then a scaled relative skill with respect to the baseline will be
returned. By default (\code{NULL}), relative skill will not be scaled with
respect to a baseline model.}

\item{rel_skill_metric}{character string with the name of the metric for which
a relative skill shall be computed. If equal to 'auto' (the default), then
one of interval score, crps or brier score will be used where appropriate}
}
\description{
A helper function to check the input parameters for
\code{\link[=score]{score()}}.
}
\keyword{internal}
