% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-input-helpers.R
\name{check_try}
\alias{check_try}
\title{Helper function to convert assert statements into checks}
\usage{
check_try(expr)
}
\arguments{
\item{expr}{an expression to be evaluated}
}
\value{
Returns TRUE if the check was successful and a string with an
error message otherwise.
}
\description{
Tries to execute an expression. Internally, this is used to
see whether assertions fail when checking inputs (i.e. to convert an
\verb{assert_*()} statement into a check). If the expression fails, the error
message is returned. If the expression succeeds, \code{TRUE} is returned.
}
\keyword{internal_input_check}
