% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_forecasts.R
\name{collapse_messages}
\alias{collapse_messages}
\title{Collapse several messages to one}
\usage{
collapse_messages(type = "messages", messages)
}
\arguments{
\item{type}{character, should be either "messages", "warnings" or "errors"}

\item{messages}{the messages or warnings to collapse}
}
\value{
string with the message or warning
}
\description{
Internal helper function to facilitate generating messages
and warnings in \code{\link[=check_forecasts]{check_forecasts()}}
}
\keyword{internal}
