% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{continuous_example_data}
\alias{continuous_example_data}
\title{Continuous Example Data}
\format{
A data frame with 6000 rows and 6 variables:
\describe{
  \item{id}{unique identifier for true observed values}
  \item{model}{name of the model that generated the forecasts}
  \item{horizon}{forecast horzin (e.g. 1 day ahead forecast)}
  \item{true_value}{true observed values}
  \item{sample}{number that identifies the predictive sample generated by
  a specific model for a specific observed value}
  \item{prediction}{predictive sample for the corresponding true
  value}
}
}
\usage{
continuous_example_data
}
\description{
A toy dataset for a probabilistic forecast of a continuous outcome variable
}
\keyword{datasets}
