% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{correlation}
\alias{correlation}
\title{Correlation Between Scoring rules}
\usage{
correlation(scores, rules = NULL, digits = NULL)
}
\arguments{
\item{scores}{A data.table of scores as produced by \code{\link[=score]{score()}}.}

\item{rules}{A character vector with the rules to show. If set to
\code{NULL} (default), all rules present in \code{scores} will
be shown}

\item{digits}{A number indicating how many decimal places the result should
be rounded to. By default (\code{digits = NULL}) no rounding takes place.}
}
\value{
A data.table with correlations for the different rules
}
\description{
Calculate the correlation between different rules for a data.frame of
scores as produced by \code{\link[=score]{score()}}.
}
\examples{
scores <- score(example_quantile)
correlation(scores, digits = 2)
}
\keyword{scoring}
