% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{correlation}
\alias{correlation}
\title{Correlation Between Metrics}
\usage{
correlation(scores, metrics = NULL)
}
\arguments{
\item{scores}{A data.table of scores as produced by \code{\link[=score]{score()}}.}

\item{metrics}{A character vector with the metrics to show. If set to
\code{NULL} (default), all metrics present in \code{scores} will
be shown}
}
\value{
A data.table with correlations for the different metrics
}
\description{
Calculate the correlation between different metrics for a data.frame of
scores as produced by \code{\link[=score]{score()}}.
}
\examples{
scores <- score(example_quantile)
correlation(scores)
}
\keyword{scoring}
