% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-templates.R
\name{document_check_functions}
\alias{document_check_functions}
\title{Documentation template for check functions}
\arguments{
\item{data}{A data.frame or similar to be checked}

\item{observed}{Input to be checked. Should be a numeric vector with the
observed values of size n.}

\item{columns}{A character vector of column names to check}
}
\value{
Returns TRUE if the check was successful and a string with an
error message otherwise.
}
\description{
Documentation template for check functions
}
\keyword{internal}
