% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-templates.R
\name{document_score_data}
\alias{document_score_data}
\title{Documentation template for scoring input data}
\arguments{
\item{data}{A data frame (or similar) of forecasts following the
specifications detailed in \code{\link[=score]{score()}}.}

\item{scores}{A data.table of scores as produced by \code{\link[=score]{score()}}.}
}
\description{
Documentation template for scoring input data
}
