% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-sample-multivariate.R
\name{energy_score_multivariate}
\alias{energy_score_multivariate}
\title{Energy score for multivariate forecasts

Compute the multivariate energy score (see )}
\usage{
energy_score_multivariate(observed, predicted, grouping_id, w = NULL)
}
\arguments{
\item{observed}{A vector with observed values of size n}

\item{predicted}{nxN matrix of predictive samples, n (number of rows) being
the number of data points and N (number of columns) the number of Monte
Carlo samples. Alternatively, if n = 1, \code{predicted} can just be a vector
of size n.}

\item{w}{numeric vector of weights for forecast draws (length equal to number of columns of \code{dat})}
}
\description{
Energy score for multivariate forecasts

Compute the multivariate energy score (see )
}
\keyword{internal_input_check}
