% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-input-helpers.R
\name{ensure_data.table}
\alias{ensure_data.table}
\title{Ensure that an object is a \code{data.table}}
\usage{
ensure_data.table(data)
}
\arguments{
\item{data}{An object to ensure is a data table.}
}
\value{
A data.table/a copy of an existing data.table.
}
\description{
This function ensures that an object is a \verb{data table}.
If the object is not a data table, it is converted to one. If the object
is a data table, a copy of the object is returned.
}
\keyword{internal}
