% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_forecasts.R
\name{eval_forecasts_prob_int}
\alias{eval_forecasts_prob_int}
\title{Applies forecast assessments and scores to probabilistic
forecasts of integers}
\usage{
eval_forecasts_prob_int(true_values, predictions, metrics = c(), output = "df")
}
\arguments{
\item{true_values}{A vector with the true observed values of size n}

\item{predictions}{a list of nxN matrices of predictive samples,
n (number of rows) being
the number of data points and N (number of columns) the
number of Monte Carlo samples}

\item{metrics}{what metrics to include. currently not used}

\item{output}{"df" returns a data.frame, everything else returns a list.}
}
\value{
either a data.frame or list of data frames with the forecast scores
}
\description{
Applies forecast assessments and scores to probabilistic
forecasts of integers
}
\examples{
## Example: Probabilistic predictions for integer values
true_values <- rpois(100, lambda = 1:100)
predictions <- replicate(5000, rpois(n = 100, lambda = 1:100))
predictions <- list(dat1 = replicate(5000, rpois(n = 100, lambda = 1:100)),
                    dat2 = replicate(5000, rpois(n = 100, lambda = 1:100)))
eval_forecasts(true_values = true_values, predictions = predictions)

}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
