% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_quantile_forecasts_only}
\alias{example_quantile_forecasts_only}
\title{Quantile Example Data - Forecasts only}
\format{
A data frame with 7,581 rows and 9 columns:
\describe{
\item{location}{the country for which a prediction was made}
\item{target_end_date}{the date for which a prediction was made}
\item{target_type}{the target to be predicted (cases or deaths)}
\item{forecast_date}{the date on which a prediction was made}
\item{quantile}{quantile of the corresponding prediction}
\item{prediction}{predicted value}
\item{model}{name of the model that generated the forecasts}
\item{horizon}{forecast horizon in weeks}
}
}
\source{
\url{https://github.com/european-modelling-hubs/covid19-forecast-hub-europe/commit/a42867b1ea152c57e25b04f9faa26cfd4bfd8fa6/} # nolint
}
\usage{
example_quantile_forecasts_only
}
\description{
A data set with quantile predictions for COVID-19 cases and deaths
submitted to the European Forecast Hub.
}
\details{
The data was created using the script create-example-data.R in the inst/
folder (or the top level folder in a compiled package).
}
\keyword{datasets}
