% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise-comparisons.R
\name{geom_mean}
\alias{geom_mean}
\title{Calculate Geometric Mean

Used in \code{\link[=pairwise_comparisons]{pairwise_comparisons()}}.}
\usage{
geom_mean(x)
}
\arguments{
\item{x}{numeric vector of values for which to calculate the geometric mean}
}
\value{
the geometric mean of the values in \code{x}
}
\description{
Calculate Geometric Mean

Used in \code{\link[=pairwise_comparisons]{pairwise_comparisons()}}.
}
\keyword{internal}
