% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coverage.R
\name{get_coverage}
\alias{get_coverage}
\title{Get Quantile And Interval Coverage Values For Quantile-Based Forecasts}
\usage{
get_coverage(data, by = "model")
}
\arguments{
\item{data}{A forecast object (a validated data.table with predicted and
observed values, see \code{\link[=as_forecast]{as_forecast()}})}

\item{by}{character vector that denotes the level of grouping for which the
coverage values should be computed. By default (\code{"model"}), one coverage
value per model will be returned.}
}
\value{
a data.table with columns "interval_coverage",
"interval_coverage_deviation", "quantile_coverage",
"quantile_coverage_deviation" and the columns specified in \code{by}.
}
\description{
Compute interval coverage of central prediction intervals,
quantile coverage for predictive quantiles, as well as the deviation between
desired and actual coverage to a data.table. Forecasts should be in a
quantile format (following the input requirements of \code{score()}).

\strong{Interval coverage}

Interval coverage for a given interval range is defined as the proportion of
observations that fall within the corresponding central prediction intervals.
Central prediction intervals are symmetric around the median and formed
by two quantiles that denote the lower and upper bound. For example, the 50\%
central prediction interval is the interval between the 0.25 and 0.75
quantiles of the predictive distribution.

The function \code{get_coverage()} computes the coverage per central prediction
interval. This means that if you set \code{by} to the unit of a single forecast,
interval coverage will always be either \code{TRUE}
(observed value falls within the interval) or \code{FALSE} (observed value falls
outside the interval) and analogously for quantile coverage.
Coverage values become meaningful by summarising them across different
dimensions, as specified in the \code{by} argument (thereby returning the
proportion of values covered by all prediction intervals/quantiles).

\strong{Quantile coverage}

Quantile coverage for a given quantile is defined as the proportion of
observed values that are smaller than the corresponding predictive quantile.
For example, the 0.5 quantile coverage is the proportion of observed values
that are smaller than the 0.5 quantile of the predictive distribution.
Just as above, for a single observation and the quantile of a single
predictive distribution, the value will either be \code{TRUE} or \code{FALSE}.

\strong{Coverage deviation}

The coverage deviation is the difference between the desired coverage
(can be either interval or quantile coverage) and the
actual coverage. For example, if the desired coverage is 90\% and the actual
coverage is 80\%, the coverage deviation is -0.1.
}
\examples{
library(magrittr) # pipe operator
example_quantile \%>\%
  as_forecast() \%>\%
  get_coverage(by = "model")
}
\keyword{scoring}
