% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_-functions.R
\name{get_duplicate_forecasts}
\alias{get_duplicate_forecasts}
\title{Find duplicate forecasts}
\usage{
get_duplicate_forecasts(data, forecast_unit = get_forecast_unit(data))
}
\arguments{
\item{data}{A data.frame as used for \code{\link[=score]{score()}}}

\item{forecast_unit}{A character vector with the column names that define
the unit of a single forecast. By default the forecast unit will be
automatically inferred from the data (see \code{\link[=get_forecast_unit]{get_forecast_unit()}})}
}
\value{
A data.frame with all rows for which a duplicate forecast was found
}
\description{
Helper function to identify duplicate forecasts, i.e.
instances where there is more than one forecast for the same prediction
target.
}
\examples{
example <- rbind(example_quantile, example_quantile[1000:1010])
get_duplicate_forecasts(example)
}
\keyword{check-forecasts}
