% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_-functions.R
\name{get_forecast_type}
\alias{get_forecast_type}
\title{Infer the type of a forecast based on a data.frame}
\usage{
get_forecast_type(data)
}
\arguments{
\item{data}{A data.frame or data.table with predicted and observed values.}
}
\value{
Character vector of length one with either "binary", "quantile",
"sample" or "point".
}
\description{
Internal helper function to get the type of the forecast.
Options are "sample-based", "quantile-based", "binary" or "point" forecast.
The function runs additional checks to make sure the data satisfies
requirements and throws an informative error if any issues are found.
}
\keyword{internal}
