% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-forecast-type.R
\name{get_forecast_type}
\alias{get_forecast_type}
\title{Get forecast type from forecast object}
\usage{
get_forecast_type(forecast)
}
\arguments{
\item{forecast}{A forecast object (a validated data.table with predicted and
observed values).}
}
\value{
Character vector of length one with the forecast type.
}
\description{
Get forecast type from forecast object
}
\keyword{internal_input_check}
