% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_-functions.R
\name{get_forecast_unit}
\alias{get_forecast_unit}
\title{Get unit of a single forecast}
\usage{
get_forecast_unit(data)
}
\arguments{
\item{data}{A data.frame or data.table with predicted and observed values.}
}
\value{
A character vector with the column names that define the unit of
a single forecast
}
\description{
Helper function to get the unit of a single forecast, i.e.
the column names that define where a single forecast was made for.
This just takes all columns that are available in the data and subtracts
the columns that are protected, i.e. those returned by
\code{\link[=get_protected_columns]{get_protected_columns()}} as well as the names of the metrics that were
specified during scoring, if any.
}
\keyword{check-forecasts}
