% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_-functions.R
\name{get_forecast_unit}
\alias{get_forecast_unit}
\title{Get unit of a single forecast}
\usage{
get_forecast_unit(data)
}
\arguments{
\item{data}{A data.frame or data.table with the following columns:
\itemize{
\item \code{observed} - observed values
\item \code{predicted} - predictions, predictive samples or predictive quantiles
\item \code{model} - name of the model or forecaster who made a prediction
}

Depending on the forecast type, one of the following columns may be required:
\itemize{
\item \code{sample_id} - index for the predictive samples in the 'predicted' column
\item \code{quantile}: quantile-level of the corresponding value in \code{predicted}
}

For more information see the vignettes and the example data
(\link{example_quantile}, \link{example_continuous},
\link{example_integer}, \code{\link[=example_point]{example_point()}}, and \link{example_binary}).}
}
\value{
A character vector with the column names that define the unit of
a single forecast
}
\description{
Helper function to get the unit of a single forecast, i.e.
the column names that define where a single forecast was made for.
This just takes all columns that are available in the data and subtracts
the columns that are protected, i.e. those returned by
\code{\link[=get_protected_columns]{get_protected_columns()}} as well as the names of the metrics that were
specified during scoring, if any.
}
\keyword{internal}
