% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_forecast_unit}
\alias{get_forecast_unit}
\title{Get unit of a single forecast}
\usage{
get_forecast_unit(data)
}
\arguments{
\item{data}{data.frame with predictions in the same format required for
\code{\link[=score]{score()}}.}
}
\value{
A character vector with the column names that define the unit of
a single forecast
}
\description{
Helper function to get the unit of a single forecast, i.e.
the column names that define where a single forecast was made for
}
\keyword{internal}
