% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_metrics}
\alias{get_metrics}
\title{Get metrics}
\usage{
get_metrics(x, ...)
}
\arguments{
\item{x}{A \code{forecast} or \code{scores} object.}

\item{...}{Additional arguments passed to the method.}
}
\description{
Generic function to to obtain default metrics availble for scoring or metrics
that were used for scoring.
\itemize{
\item If called on \code{forecast} object it returns a list of functions that can be
used for scoring.
\item If called on a \code{scores} object (see \code{\link[=score]{score()}}), it returns a character vector
with the names of the metrics that were used for scoring.
}

See the documentation for the actual methods in the \verb{See Also} section below
for more details. Alternatively call \verb{?get_metrics.<forecast_type>} or
\code{?get_metrics.scores}.
}
\details{
See \code{\link[=as_forecast]{as_forecast()}} for more information on \code{forecast} objects and \code{\link[=score]{score()}}
for more information on \code{scores} objects.
}
\seealso{
Other \code{get_metrics} functions: 
\code{\link{get_metrics.forecast_binary}()},
\code{\link{get_metrics.forecast_nominal}()},
\code{\link{get_metrics.forecast_point}()},
\code{\link{get_metrics.forecast_quantile}()},
\code{\link{get_metrics.forecast_sample}()},
\code{\link{get_metrics.scores}()}
}
\concept{\code{get_metrics} functions}
\keyword{handle-metrics}
