% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_-functions.R
\name{get_metrics}
\alias{get_metrics}
\title{Get metrics that were used for scoring}
\usage{
get_metrics(scores)
}
\arguments{
\item{score}{A data.table with an attribute \code{metric_names}}
}
\value{
Character vector with the metrics that were used for scoring.
}
\description{
Internal helper function to get the metrics that were used
to score forecasts.
}
\keyword{internal_input_check}
