% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pit.R
\name{get_pit}
\alias{get_pit}
\title{Probability Integral Transformation (data.frame Format)}
\usage{
get_pit(data, by, n_replicates = 100)
}
\arguments{
\item{data}{a data.frame with the following columns: \code{observed},
\code{predicted}, \code{sample_id}.}

\item{by}{Character vector with the columns according to which the
PIT values shall be grouped. If you e.g. have the columns 'model' and
'location' in the data and want to have a PIT histogram for
every model and location, specify \code{by = c("model", "location")}.}

\item{n_replicates}{the number of draws for the randomised PIT for
integer predictions.}
}
\value{
a data.table with PIT values according to the grouping specified in
\code{by}
}
\description{
Compute the Probability Integral Transformation (PIT) for
validated forecast objects.
}
\examples{
result <- get_pit(as_forecast(example_continuous), by = "model")
plot_pit(result)

# example with quantile data
result <- get_pit(as_forecast(example_quantile), by = "model")
plot_pit(result)
}
\references{
Sebastian Funk, Anton Camacho, Adam J. Kucharski, Rachel Lowe,
Rosalind M. Eggo, W. John Edmunds (2019) Assessing the performance of
real-time epidemic forecasts: A case study of Ebola in the Western Area
region of Sierra Leone, 2014-15, \doi{10.1371/journal.pcbi.1006785}
}
\keyword{scoring}
