% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-protected-columns.R
\name{get_protected_columns}
\alias{get_protected_columns}
\title{Get protected columns from data}
\usage{
get_protected_columns(data = NULL)
}
\arguments{
\item{data}{A data.frame (or similar) with predicted and observed values.
See the details section of for additional information
on the required input format.}
}
\value{
A character vector with the names of protected columns in the data.
If data is \code{NULL} (default) then it returns a list of all columns that are
protected in scoringutils.
}
\description{
Helper function to get the names of all columns in a data frame
that are protected columns.
}
\keyword{internal}
