% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_-functions.R
\name{get_target_type}
\alias{get_target_type}
\title{Get type of the target true values of a forecast}
\usage{
get_target_type(data)
}
\arguments{
\item{data}{A data.frame or data.table with the following columns:
\itemize{
\item \code{observed} - observed values
\item \code{predicted} - predictions, predictive samples or predictive quantiles
\item \code{model} - name of the model or forecaster who made a prediction
}

Depending on the forecast type, one of the following columns may be required:
\itemize{
\item \code{sample_id} - index for the predictive samples in the 'predicted' column
\item \code{quantile}: quantile-level of the corresponding value in \code{predicted}
}

For more information see the vignettes and the example data
(\link{example_quantile}, \link{example_continuous},
\link{example_integer}, \code{\link[=example_point]{example_point()}}, and \link{example_binary}).}
}
\value{
Character vector of length one with either "binary", "integer", or
"continuous"
}
\description{
Internal helper function to get the type of the target
true values of a forecast. That is inferred based on the type and the
content of the \code{observed} column.
}
\keyword{internal}
