% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_target_type}
\alias{get_target_type}
\title{Get type of the target true values of a forecast}
\usage{
get_target_type(data)
}
\arguments{
\item{data}{data.frame with predictions in the same format required for
\code{\link[=score]{score()}}.}
}
\value{
Character vector of length one with either "binary", "integer", or
"continuous"
}
\description{
Internal helper function to get the type of the target
true values of a forecast. That is inferred based on the which columns
are present in the data.
}
\keyword{internal}
