% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-missing-scores.R
\name{impute_mean_score}
\alias{impute_mean_score}
\title{Impute with mean observed score}
\usage{
impute_mean_score()
}
\value{
A function suitable for use as the \code{strategy}
argument in \code{\link[=impute_missing_scores]{impute_missing_scores()}}.
}
\description{
Creates an imputation strategy that fills each missing
metric with the mean observed value for that metric within
the same target combination across all elements of
\code{compare}.
}
\examples{
\dontshow{
  data.table::setDTthreads(2)
}
scores <- example_quantile |>
  as_forecast_quantile() |>
  score()

impute_missing_scores(scores, strategy = impute_mean_score())
}
\keyword{postprocess-scores}
