% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-missing-scores.R
\name{impute_worst_score}
\alias{impute_worst_score}
\title{Impute with worst (maximum) observed score}
\usage{
impute_worst_score()
}
\value{
A strategy function for \code{\link[=impute_missing_scores]{impute_missing_scores()}}.
}
\description{
Strategy for \code{\link[=impute_missing_scores]{impute_missing_scores()}} that fills each
missing metric with the worst (maximum) observed value for
that metric within the same target combination across all
values of the \code{compare} column. Target combinations with no
non-NA observations are filled with \code{NA_real_}.
}
\examples{
\dontshow{
  data.table::setDTthreads(2)
}
scores <- example_quantile |>
  as_forecast_quantile() |>
  score()

impute_missing_scores(scores, strategy = impute_worst_score())
}
\seealso{
\code{\link[=impute_missing_scores]{impute_missing_scores()}}, \code{\link[=impute_mean_score]{impute_mean_score()}}
}
\keyword{postprocess-scores}
