% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise-comparisons.R
\name{infer_rel_skill_metric}
\alias{infer_rel_skill_metric}
\title{Infer metric for pairwise comparisons}
\usage{
infer_rel_skill_metric(scores)
}
\arguments{
\item{scores}{A data.table of scores as produced by \code{\link[=score]{score()}}.}
}
\description{
Helper function to infer the metric for which pairwise comparisons shall
be made. The function simply checks the names of the available columns and
chooses the most widely used metric.
}
\keyword{internal}
