% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{integer_example_data}
\alias{integer_example_data}
\title{Integer Forecast Example Data}
\format{
A data frame with 13,429 rows and 10 columns:
\describe{
  \item{value_date}{the date for which a prediction was made}
  \item{value_type}{the target to be predicted (short form)}
  \item{geography}{the region for which a prediction was made}
  \item{value_desc}{long form description of the prediction target}
  \item{model}{name of the model that generated the forecasts}
  \item{creation_date}{date on which the forecast was made}
  \item{horizon}{forecast horizon in days}
  \item{prediction}{prediction value for the corresponding sample}
  \item{sample}{id for the corresponding sample}
  \item{true_value}{true observed values}
}
}
\usage{
integer_example_data
}
\description{
A data set with integer predictions in a sample-based format relevant in the
2020 UK Covid-19 epidemic.
}
\keyword{datasets}
