% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{integer_example_data}
\alias{integer_example_data}
\title{Integer Example Data}
\format{
A data frame with 3000 rows and 5 variables:
\describe{
  \item{id}{unique identifier for true observed values}
  \item{model}{name of the model that generated the forecasts}
  \item{true_values}{true observed values}
  \item{sample}{number that identifies the predictive sample generated by
  a specific model for a specific observed value}
  \item{predictions}{predictive sample for the corresponding true
  value}
}
}
\usage{
integer_example_data
}
\description{
A toy dataset for a probabilistic forecast of an integer outcome variable
}
\keyword{datasets}
