% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{is_forecast}
\alias{is_forecast}
\title{Test Whether An Object Is Of Class \verb{forecast_*}}
\usage{
is_forecast(x, ...)
}
\arguments{
\item{x}{An R object.}
}
\value{
\code{TRUE} if the object is of class \verb{forecast_*}, \code{FALSE} otherwise.
}
\description{
Generic function to test whether an object is of class \verb{forecast_*}.
}
\examples{
forecast_binary <- as_forecast(example_binary)
is_forecast(forecast_binary)
}
\keyword{check-forecasts}
