% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{is_forecast}
\alias{is_forecast}
\alias{is_forecast_sample}
\alias{is_forecast_binary}
\alias{is_forecast_point}
\alias{is_forecast_quantile}
\alias{is_forecast_nominal}
\title{Test whether an object is a forecast object}
\usage{
is_forecast(x)

is_forecast_sample(x)

is_forecast_binary(x)

is_forecast_point(x)

is_forecast_quantile(x)

is_forecast_nominal(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
\emph{\code{is_forecast}}: \code{TRUE} if the object is of class \code{forecast},
\code{FALSE} otherwise.

\emph{\verb{is_forecast_<type>*}}: \code{TRUE} if the object is of class \verb{forecast_*} in addition
to class \code{forecast}, \code{FALSE} otherwise.
}
\description{
Test whether an object is a forecast object (see \code{\link[=as_forecast]{as_forecast()}} for more
information).

You can test for a specific \verb{forecast_*} class using the appropriate
\verb{is_forecast_*} function.
}
\examples{
forecast_binary <- as_forecast_binary(example_binary)
is_forecast(forecast_binary)
}
\keyword{validate-forecast-object}
