% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-scoring-rules.R
\name{metrics_binary}
\alias{metrics_binary}
\title{Default Metrics And Scoring Rules for Binary Forecasts}
\usage{
metrics_binary(select = NULL, exclude = NULL)
}
\arguments{
\item{select}{A character vector of scoring rules to select from the list.
If \code{select} is \code{NULL} (the default), all possible scoring rules are returned.}

\item{exclude}{A character vector of scoring rules to exclude from the list.
If \code{select} is not \code{NULL}, this argument is ignored.}
}
\value{
A list of scoring rules.
}
\description{
Helper function that returns a named list of default
scoring rules suitable for binary forecasts.

The default scoring rules are:
\itemize{
\item "brier_score" = \code{\link[=brier_score]{brier_score()}}
\item "log_score" = \code{\link[=logs_binary]{logs_binary()}}
}
}
\examples{
metrics_binary()
metrics_binary(select = "brier_score")
metrics_binary(exclude = "log_score")
}
\keyword{metric}
