% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-scoring-rules.R
\name{metrics_sample}
\alias{metrics_sample}
\title{Default metrics and scoring rules sample-based forecasts}
\usage{
metrics_sample(select = NULL, exclude = NULL)
}
\arguments{
\item{select}{A character vector of scoring rules to select from the list. If
\code{select} is \code{NULL} (the default), all possible scoring rules are returned.}

\item{exclude}{A character vector of scoring rules to exclude from the list.
If \code{select} is not \code{NULL}, this argument is ignored.}
}
\value{
A list of scoring rules.
}
\description{
Helper function that returns a named list of default
scoring rules suitable for forecasts in a sample-based format.

The default scoring rules are:
\itemize{
\item "crps" = \code{\link[=crps_sample]{crps_sample()}}
\item "overprediction" = \code{\link[=overprediction_sample]{overprediction_sample()}}
\item "underprediction" = \code{\link[=underprediction_sample]{underprediction_sample()}}
\item "dispersion" = \code{\link[=dispersion_sample]{dispersion_sample()}}
\item "log_score" = \code{\link[=logs_sample]{logs_sample()}}
\item "dss" = \code{\link[=dss_sample]{dss_sample()}}
\item "mad" = \code{\link[=mad_sample]{mad_sample()}}
\item "bias" = \code{\link[=bias_sample]{bias_sample()}}
\item "ae_median" = \code{\link[=ae_median_sample]{ae_median_sample()}}
\item "se_mean" = \code{\link[=se_mean_sample]{se_mean_sample()}}
}
}
\examples{
metrics_sample()
metrics_sample(select = "mad")
}
\keyword{metric}
