% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metrics_summary}
\alias{metrics_summary}
\title{Summary information for selected metrics}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 18 rows and 8 columns.
}
\usage{
metrics_summary
}
\description{
A data set with summary information on selected metrics implemented in
\pkg{scoringutils}
}
\keyword{info}
