% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-forecast.R
\name{new_forecast}
\alias{new_forecast}
\title{Class constructor for \code{forecast} objects}
\usage{
new_forecast(data, classname)
}
\arguments{
\item{data}{A data.frame (or similar) with predicted and observed values.
See the details section of for additional information
on the required input format.}

\item{classname}{name of the class to be created}
}
\value{
An object of the class indicated by \code{classname}
}
\description{
Construct a class based on a data.frame or similar. The constructor
\itemize{
\item coerces the data into a data.table
\item assigns a class
}
}
\keyword{internal}
