% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{new_scoringutils}
\alias{new_scoringutils}
\title{Class constructor for scoringutils objects}
\usage{
new_scoringutils(data, classname)
}
\arguments{
\item{data}{A data.frame or data.table with the following columns:
\itemize{
\item \code{observed} - observed values
\item \code{predicted} - predictions, predictive samples or predictive quantiles
\item \code{model} - name of the model or forecaster who made a prediction
}

Depending on the forecast type, one of the following columns may be required:
\itemize{
\item \code{sample_id} - index for the predictive samples in the 'predicted' column
\item \code{quantile}: quantile-level of the corresponding value in \code{predicted}
}

For more information see the vignettes and the example data
(\link{example_quantile}, \link{example_continuous},
\link{example_integer}, \code{\link[=example_point]{example_point()}}, and \link{example_binary}).}

\item{classname}{name of the class to be created}
}
\value{
An object of the class indicated by \code{classname}
}
\description{
Construct a class based on a data.frame or similar. The constructor
\itemize{
\item coerces the data into a data.table
\item makes sure that a column called \code{model} exists and if not creates one
\item assigns a class
}
}
\keyword{validate}
