% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise-comparisons.R
\name{pairwise_comparison}
\alias{pairwise_comparison}
\title{Do pairwise Comparisons of Scores}
\usage{
pairwise_comparison(
  scores,
  var = "model",
  metric = "interval_score",
  permutation_test = FALSE,
  baseline = NULL
)
}
\arguments{
\item{scores}{A data.frame of unsummarised scores as produced by
\code{\link{eval_forecasts}}}

\item{var}{string that denotes the column name with the models or forecasters
to compare. Will probably be changed in the future such that a column
named "model" must be present.}

\item{metric}{A character vector of length one with the metric to do
the comparison on.}

\item{permutation_test}{logical. If TRUE, a permutation test is done
alongside the comparison. Default is FALSE}

\item{baseline}{the baseline model against which to compare other models.}
}
\value{
A ggplot2 object with a coloured table of summarised scores
}
\description{
Make pairwise comparisons between models. The code for the pairwise
comparisons is inspired by an implementation by Johannes Bracher.

The implementation of the permutation test follows the function
\link[surveillance]{permutationTest} from the `surveillance` package
 by Michael Höhle,
Andrea Riebler and Michaela Paul.
}
\examples{
df <- data.frame(model = rep(c("model1", "model2", "model3"), each = 10),
                 id = rep(1:10),
                 interval_score = (abs(rnorm(30))),
                 aem = (abs(rnorm(30))))

res <- scoringutils::pairwise_comparison(df, permutation_test = TRUE,
                                         baseline = "model1")
scoringutils::plot_pairwise_comparison(res, type = "ratio")

}
\author{
Johannes Bracher, https://jbracher.github.io/

Nikos Bosse
}
