% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.scoringutils_correlation}
\alias{plot.scoringutils_correlation}
\title{Plot Correlation Between Metrics}
\usage{
\method{plot}{scoringutils_correlation}(x, ...)
}
\arguments{
\item{x}{An S3 object with correlations between scores as produced by
\code{\link[=correlation]{correlation()}}.}

\item{...}{additional arguments (not used here)}
}
\value{
A ggplot2 object showing a coloured matrix of correlations
between metrics
}
\description{
Plots a heatmap of correlations between different metrics
}
\examples{
scores <- score(example_quantile)
correlations <- correlation(
 summarise_scores(scores)
)
plot(correlations)
}
