% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Plot Correlation Between Metrics}
\usage{
plot_correlation(correlations)
}
\arguments{
\item{correlations}{A data.table of correlations between scores as produced
by \code{\link[=correlation]{correlation()}}.}
}
\value{
A ggplot2 object showing a coloured matrix of correlations
between metrics
}
\description{
Plots a heatmap of correlations between different metrics
}
\examples{
scores <- score(example_quantile)
correlations <- correlation(
 summarise_scores(scores)
)
plot_correlation(correlations)
}
