% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_interval_coverage}
\alias{plot_interval_coverage}
\title{Plot Interval Coverage}
\usage{
plot_interval_coverage(coverage, colour = "model")
}
\arguments{
\item{coverage}{A data frame of coverage values as produced by
\code{get_coverage()}}

\item{colour}{According to which variable shall the graphs be coloured?
Default is "model".}
}
\value{
ggplot object with a plot of interval coverage
}
\description{
Plot interval coverage
}
\examples{
\dontshow{
  data.table::setDTthreads(2) # restricts number of cores used on CRAN
}
coverage <- get_coverage(example_quantile, by = c("model"))
plot_interval_coverage(coverage)
}
