% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pairwise_comparison}
\alias{plot_pairwise_comparison}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Plot Heatmap of Pairwise Comparisons}
\usage{
plot_pairwise_comparison(
  comparison_result,
  type = c("mean_scores_ratio", "pval")
)
}
\arguments{
\item{comparison_result}{A data.frame as produced by
\code{\link[=pairwise_comparison]{pairwise_comparison()}}}

\item{type}{character vector of length one that is either
"mean_scores_ratio" or "pval". This denotes whether to
visualise the ratio or the p-value of the pairwise comparison.
Default is "mean_scores_ratio".}
}
\value{
ggplot object with a plot of interval coverage

A ggplot2 object showing a coloured matrix of correlations
between metrics
}
\description{
Deprecated version of \code{\link[=plot.scoringutils_pairwise]{plot.scoringutils_pairwise()}}
}
\examples{
library(ggplot2)
scores <- score(example_quantile)
pairwise <- pairwise_comparison(scores, by = "target_type")
plot(pairwise, type = "mean_scores_ratio") +
  facet_wrap(~target_type)
library(ggplot2)
available_forecasts <- available_forecasts(
  example_quantile, by = c("model", "target_type", "target_end_date")
)
plot(
 available_forecasts, xvar = "target_end_date", show_numbers = FALSE
) +
 facet_wrap("target_type")
scores <- score(example_quantile)
correlations <- correlation(
 summarise_scores(scores)
)
plot(correlations)
}
