% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise-comparisons.R
\name{plot_pairwise_comparison}
\alias{plot_pairwise_comparison}
\title{Plot Heatmap of Pairwise Comparisons}
\usage{
plot_pairwise_comparison(
  comparison_result,
  type = c("together", "ratio", "pval"),
  smaller_is_good = TRUE
)
}
\arguments{
\item{comparison_result}{A data.frame as produced by
\code{\link{pairwise_comparison}}}

\item{type}{character vector of length one that is either "ratio" or "pval".
This denotes whether to visualise the ratio or the p-value of the
pairwise comparison. Default is "ratio"}
}
\description{
Creates a heatmap of the ratios or pvalues from a pairwise comparison
between models
}
\examples{
df <- data.frame(model = rep(c("model1", "model2", "model3"), each = 10),
                 id = rep(1:10),
                 interval_score = abs(rnorm(30, mean = rep(c(1, 1.3, 2), each = 10))),
                 aem = (abs(rnorm(30))))

res <- scoringutils::pairwise_comparison(df, permutation_test = TRUE, baseline = "model1")
plot_pairwise_comparison(res, smaller_is_good = T)
plot_pairwise_comparison(res, smaller_is_good = T, type = "pval")
}
