% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise-comparisons.R
\name{plot_pairwise_comparison}
\alias{plot_pairwise_comparison}
\title{Plot Heatmap of Pairwise Comparisons}
\usage{
plot_pairwise_comparison(
  comparison_result,
  type = c("mean_scores_ratio", "pval", "together"),
  smaller_is_good = TRUE
)
}
\arguments{
\item{comparison_result}{A data.frame as produced by
\code{\link{pairwise_comparison}}}

\item{type}{character vector of length one that is either "mean_scores_ratio" or "pval".
This denotes whether to visualise the ratio or the p-value of the
pairwise comparison. Default is "mean_scores_ratio"}

\item{smaller_is_good}{logical (default is TRUE) that indicates whether
smaller or larger values are to be interpreted as 'good' (as you could just
invert the mean scores ratio)}
}
\description{
Creates a heatmap of the ratios or pvalues from a pairwise comparison
between models
}
\examples{
df <- data.frame(model = rep(c("model1", "model2", "model3"), each = 10),
                 id = rep(1:10),
                 interval_score = abs(rnorm(30, mean = rep(c(1, 1.3, 2), each = 10))),
                 aem = (abs(rnorm(30))))

res <- scoringutils::pairwise_comparison(df, baseline = "model1")
plot_pairwise_comparison(res, smaller_is_good = TRUE)

res <- pairwise_comparison(df)
scoringutils::plot_pairwise_comparison(res, smaller_is_good = TRUE)
scoringutils::plot_pairwise_comparison(res, smaller_is_good = TRUE, type = "pval")
}
