% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_quantile_coverage}
\alias{plot_quantile_coverage}
\title{Plot Quantile Coverage}
\usage{
plot_quantile_coverage(coverage, colour = "model")
}
\arguments{
\item{coverage}{A data frame of coverage values as produced by
\code{\link[=get_coverage]{get_coverage()}}}

\item{colour}{According to which variable shall the graphs be coloured?
Default is "model".}
}
\value{
ggplot object with a plot of interval coverage
}
\description{
Plot quantile coverage
}
\examples{
coverage <- get_coverage(as_forecast(example_quantile), by = "model")
plot_quantile_coverage(coverage)
}
