% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_quantile_coverage}
\alias{plot_quantile_coverage}
\title{Plot Quantile Coverage}
\usage{
plot_quantile_coverage(data, colour = "model")
}
\arguments{
\item{data}{A data frame of coverage values as produced by
\code{get_coverage()}}

\item{colour}{According to which variable shall the graphs be coloured?
Default is "model".}
}
\value{
ggplot object with a plot of interval coverage
}
\description{
Plot quantile coverage
}
\examples{
data_coverage <- add_coverage(example_quantile)
summarised <- summarise_scores(data_coverage, by = c("model", "quantile_level"))
plot_quantile_coverage(summarised)
}
