% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-coverage.R
\name{plot_quantile_coverage}
\alias{plot_quantile_coverage}
\title{Plot quantile coverage}
\usage{
plot_quantile_coverage(coverage, colour = "model")
}
\arguments{
\item{coverage}{A data frame of coverage values as produced by
\code{\link[=get_coverage]{get_coverage()}}.}

\item{colour}{String, according to which variable shall the graphs be
coloured? Default is "model".}
}
\value{
A ggplot object with a plot of interval coverage
}
\description{
Plot quantile coverage values (see \code{\link[=get_coverage]{get_coverage()}} for more information).
}
\examples{
example <- as_forecast_quantile(example_quantile)
coverage <- get_coverage(example, by = "model")
plot_quantile_coverage(coverage)
}
