% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-forecast.R
\name{print.forecast}
\alias{print.forecast}
\title{Print information about a forecast object}
\usage{
\method{print}{forecast}(x, ...)
}
\arguments{
\item{x}{A forecast object}

\item{...}{Additional arguments for \code{\link[=print]{print()}}.}
}
\value{
Returns \code{x} invisibly.
}
\description{
This function prints information about a forecast object,
including "Forecast type", "Score columns",
"Forecast unit".
}
\examples{
dat <- as_forecast_quantile(example_quantile)
print(dat)
}
\keyword{gain-insights}
