% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.forecast_binary}
\alias{print.forecast_binary}
\alias{print.forecast_quantile}
\alias{print.forecast_point}
\alias{print.forecast_sample}
\title{Print Information About A Forecast Object}
\usage{
\method{print}{forecast_binary}(x, ...)

\method{print}{forecast_quantile}(x, ...)

\method{print}{forecast_point}(x, ...)

\method{print}{forecast_sample}(x, ...)
}
\arguments{
\item{x}{An object of class 'forecast_*' object as produced by
\code{as_forecast()}}

\item{...}{additional arguments for \code{\link[=print]{print()}}}
}
\value{
returns x invisibly
}
\description{
This function prints information about a forecast object,
including "Forecast type", "Score columns",
"Forecast unit".
}
\examples{
dat <- as_forecast(example_quantile)
print(dat)
}
\keyword{check-forecasts}
