% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-scores.R
\name{print.scores}
\alias{print.scores}
\title{Print a scores object}
\usage{
\method{print}{scores}(x, ...)
}
\arguments{
\item{x}{A \code{scores} object as produced by \code{\link[=score]{score()}}.}

\item{...}{Additional arguments passed to the \code{print} method of the
underlying data.table.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Prints a summary of the scoring metrics used and a link to the
scoring rules vignette, followed by the underlying data.
}
\examples{
\dontrun{
scores <- score(example_quantile)
print(scores)
}
}
\keyword{gain-insights}
