% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_forecasts.R
\name{print.scoringutils_check}
\alias{print.scoringutils_check}
\title{Print output from \code{check_forecasts()}}
\usage{
\method{print}{scoringutils_check}(check)
}
\arguments{
\item{data}{A list of class 'scoringutils_check'.}
}
\description{
Helper function that prints the output generated by
\code{\link[=check_forecasts]{check_forecasts()}}
}
\keyword{internal}
