% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{quantile_coverage}
\alias{quantile_coverage}
\title{Plot Quantile Coverage}
\usage{
quantile_coverage(
  summarised_scores,
  colour = "model",
  facet_formula = NULL,
  facet_wrap_or_grid = "facet_wrap",
  scales = "free_y"
)
}
\arguments{
\item{colour}{According to which variable shall the graphs be coloured?
Default is "model".}

\item{facet_formula}{formula for facetting in ggplot. If this is \code{NULL}
(the default), no facetting will take place}

\item{facet_wrap_or_grid}{Use ggplot2's \code{facet_wrap} or
\code{facet_grid}? Anything other than "facet_wrap" will be interpreted as
\code{facet_grid}. This only takes effect if \code{facet_formula} is not
\code{NULL}}

\item{scales}{scales argument that gets passed down to ggplot. Only necessary
if you make use of facetting. Default is "free_y"}

\item{data}{a data.frame that follows the same specifications outlined in
\code{\link{eval_forecasts}}. The data.frame needs to have columns called
"true_value", "prediction" and then either a column called sample, or one
called "quantile" or two columns called "range" and "boundary".}

\item{summarise_by}{grouping used to determine the percentage of true values
below a predictive quantile. Default is \code{c("model", "quantile")}.
"quantile" must always be present in \code{summarise_by}.}
}
\value{
ggplot object with a plot of interval coverage
}
\description{
Plot quantile coverage
}
\examples{
example1 <- scoringutils::quantile_example_data_long
scores <- scoringutils::eval_forecasts(example1,
                                       summarise_by = c("model", "quantile"))
quantile_coverage(scores)
}
