% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quantile_example_data_long}
\alias{quantile_example_data_long}
\title{Quantile Example Data - Long Format}
\format{
A data frame with 720 rows and 7 variables:
\describe{
  \item{true_value}{true observed values}
  \item{id}{unique identifier for true observed values}
  \item{horizon}{forecast horizon (e.g. 1 day ahead forecast)}
  \item{model}{name of the model that generated the forecasts}
  \item{prediction}{quantile predictions}
  \item{boundary}{lower or upper bound of an interval range}
  \item{range}{interval range for which the quantile forecast was made}
}
}
\usage{
quantile_example_data_long
}
\description{
A toy dataset for quantile forecasts of an outcome variable
}
\keyword{datasets}
