% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_handling.R
\name{quantile_to_long}
\alias{quantile_to_long}
\title{Pivot Range Format Forecasts From Wide to Long Format}
\usage{
quantile_to_long(data)
}
\arguments{
\item{data}{a data.frame following the specifications from
\code{\link{eval_forecasts}}) for quantile forecasts. For an example, see
\code{\link{range_example_data_long}})}
}
\value{
a data.frame in long format
}
\description{
Legacy function that will not be supported in future updates.
}
