% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_handling.R
\name{quantile_to_range}
\alias{quantile_to_range}
\title{Change Data from a Plain Quantile Format to a Long Range Format}
\usage{
quantile_to_range(data, keep_quantile_col = FALSE)
}
\arguments{
\item{data}{a data.frame following the specifications shown in the example
\code{\link{range_example_data_long}})}

\item{keep_quantile_col}{keep the quantile column in the final
output after transformation (default is FALSE)}
}
\value{
a data.frame in long format
}
\description{
Legacy function that will not be supported in future updates.
}
