% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-scoring-report.R
\name{render_scoring_report}
\alias{render_scoring_report}
\title{Render a Scoring Report}
\usage{
render_scoring_report(
  format = "html_document",
  truth_data,
  prediction_data,
  params = list(locations = "all", forecast_dates = "all", horizons = "all"),
  interactive = FALSE,
  save_dir = NULL,
  filename = NULL
)
}
\arguments{
\item{format}{Character string, defaults to `"html_document"`. The format to render the report to.
See `?rmarkdown::render` for details.}

\item{interactive}{Logical, defaults to `FALSE`. When the format allows should graphs be interactive.}

\item{save_dir}{Character string, defaults to `NULL`.
If not given then the report is rendered to a temporary directory (although only if `filename` is also not given).}

\item{filename}{Character string defaults `NULL`. Name to save the report under, defaults to `"scoring_report"`.}
}
\value{
Renders a scoring report
}
\description{
This function renders a country level TB report.
}
\examples{




}
