% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-scoring-rules.R
\name{rules_quantile}
\alias{rules_quantile}
\title{Scoring Rules for Quantile-Based Forecasts}
\usage{
rules_quantile(select = NULL, exclude = NULL)
}
\arguments{
\item{select}{A character vector of scoring rules to select from the list.
If \code{select} is \code{NULL} (the default), all possible scoring rules are returned.}

\item{exclude}{A character vector of scoring rules to exclude from the list.
If \code{select} is not \code{NULL}, this argument is ignored.}
}
\value{
A list of scoring rules.
}
\description{
Helper function that returns a named list of default
scoring rules suitable for forecasts in a quantile-based format

The default scoring rules are:
\itemize{
\item "wis" = \link{wis}
\item "overprediction" = \code{\link[=overprediction]{overprediction()}}
\item "underprediction" = \code{\link[=underprediction]{underprediction()}}
\item "dispersion" = \code{\link[=dispersion]{dispersion()}}
\item "bias" = \code{\link[=bias_quantile]{bias_quantile()}}
\item "interval_coverage_50" = \code{\link[=interval_coverage]{interval_coverage()}}
\item "interval_coverage_90" = function(...) \{
run_safely(..., range = 90, fun = \link{interval_coverage})
\}
\item "interval_coverage_deviation" = \code{\link[=interval_coverage_dev_quantile]{interval_coverage_dev_quantile()}},
\item "ae_median" = \code{\link[=ae_median_quantile]{ae_median_quantile()}}
}

Note: The \code{coverage_90} scoring rule is created as a wrapper around
\code{\link[=interval_coverage]{interval_coverage()}}, making use of the function \code{\link[=run_safely]{run_safely()}}.
This construct allows the function to deal with arbitrary arguments in \code{...},
while making sure that only those that \code{\link[=interval_coverage]{interval_coverage()}} can
accept get passed on to it. \code{range = 90} is set in the function definition,
as passing an argument \code{range = 90} to \code{\link[=score]{score()}} would mean it would also
get passed to \code{coverage_50}.
}
\examples{
rules_quantile()
rules_quantile(select = "wis")
}
\keyword{metric}
