% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-quantile-interval-range.R
\name{sample_to_interval_long}
\alias{sample_to_interval_long}
\title{Change data from a sample-based format to a long interval range format}
\usage{
sample_to_interval_long(
  data,
  interval_range = c(0, 50, 90),
  type = 7,
  keep_quantile_col = TRUE
)
}
\arguments{
\item{data}{A data.frame (or similar) with predicted and observed values.
See the details section of for additional information
on the required input format.}

\item{type}{Type argument passed down to the quantile function. For more
information, see \code{\link[=quantile]{quantile()}}.}

\item{keep_quantile_col}{keep quantile_level column, default is TRUE}
}
\value{
A data.table in a long interval interval range format
}
\description{
Transform data from a format that is based on predictive samples to a format
based on interval ranges.
}
\keyword{internal}
