% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_handling.R
\name{sample_to_quantile}
\alias{sample_to_quantile}
\title{Change data from a sample based format to a quantile format}
\usage{
sample_to_quantile(
  data,
  quantile_level = c(0.05, 0.25, 0.5, 0.75, 0.95),
  type = 7
)
}
\arguments{
\item{data}{A \code{forecast} object of class \code{forecast_sample} (a validated
data.table with predicted and observed values, see \code{\link[=as_forecast]{as_forecast()}}).}

\item{quantile_level}{A numeric vector of quantile levels for which
quantiles will be computed.}

\item{type}{Type argument passed down to the quantile function. For more
information, see \code{\link[=quantile]{quantile()}}.}
}
\value{
a data.table in a long interval range format
}
\description{
Transform data from a format that is based on predictive samples to a format
based on plain quantiles.
}
\examples{
sample_to_quantile(as_forecast(example_integer))
}
\keyword{data-handling}
