% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_handling.R
\name{sample_to_range}
\alias{sample_to_range}
\title{Change Data from a Sample Based Format to a Long Interval Range Format}
\usage{
sample_to_range(data, range = c(0, 50, 90), type = 7, keep_quantile_col = TRUE)
}
\arguments{
\item{data}{a data.frame with samples}

\item{range}{a numeric vector of interval ranges to extract
(e.g. \code{c(0, 50, 90)})}

\item{type}{type argument passed down to the quantile function. For more
information, see \code{\link{quantile}}}

\item{keep_quantile_col}{keep quantile column, default is TRUE}
}
\value{
a data.frame in long format
}
\description{
Legacy function that will not be supported in future updates.
}
