% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoringutils.R
\docType{package}
\name{scoringutils-package}
\alias{scoringutils}
\alias{scoringutils-package}
\title{scoringutils: Utilities for Scoring and Assessing Predictions}
\description{
Combines a collection of metrics and proper scoring rules (Tilmann Gneiting & Adrian E Raftery (2007) <doi:10.1198/016214506000001437>) with an easy to use wrapper that can be used to automatically evaluate predictions. Apart from proper scoring rules functions are provided to assess bias, sharpness and calibration (Sebastian Funk, Anton Camacho, Adam J. Kucharski, Rachel Lowe, Rosalind M. Eggo, W. John Edmunds (2019) <doi:10.1371/journal.pcbi.1006785>) of forecasts. Several types of predictions can be evaluated: probabilistic forecasts (generally predictive samples generated by Markov Chain Monte Carlo procedures), quantile forecasts or point forecasts. Observed values and predictions can be either continuous, integer, or binary. Users can either choose to apply these rules separately in a vector / matrix format that can be flexibly used within other packages, or they can choose to do an automatic evaluation of their forecasts. This is implemented with 'data.table' and provides a consistent and very efficient framework for evaluating various types of predictions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epiforecasts/scoringutils}
  \item \url{https://epiforecasts.io/scoringutils/}
  \item Report bugs at \url{https://github.com/epiforecasts/scoringutils/issues}
}

}
\author{
\strong{Maintainer}: Nikos Bosse \email{nikosbosse@gmail.com} (\href{https://orcid.org/0000-0002-7750-5280}{ORCID})

Authors:
\itemize{
  \item Sam Abbott \email{contact@samabbott.co.uk} (\href{https://orcid.org/0000-0001-8057-8037}{ORCID})
  \item Hugo Gruson \email{hugo.gruson@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID})
}

Other contributors:
\itemize{
  \item Johannes Bracher \email{johannes.bracher@kit.edu} (\href{https://orcid.org/0000-0002-3777-1410}{ORCID}) [contributor]
  \item Sebastian Funk \email{sebastian.funk@lshtm.ac.uk} [contributor]
}

}
\keyword{internal}
