% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoringutils.R
\docType{package}
\name{scoringutils-package}
\alias{scoringutils}
\alias{scoringutils-package}
\title{scoringutils: Utilities for Scoring and Assessing Predictions}
\description{
Facilitate the evaluation of forecasts in a convenient framework based on data.table. It allows user to to check their forecasts and diagnose issues, to visualise forecasts and missing data, to transform data before scoring, to handle missing forecasts, to aggregate scores, and to visualise the results of the evaluation. The package mostly focuses on the evaluation of probabilistic forecasts and allows evaluating several different forecast types and input formats. Find more information about the package in the Vignettes as well as in the accompanying paper, \doi{10.48550/arXiv.2205.07090}.
}
\seealso{
Useful links:
\itemize{
  \item \doi{10.48550/arXiv.2205.07090}
  \item \url{https://epiforecasts.io/scoringutils/}
  \item \url{https://github.com/epiforecasts/scoringutils}
  \item Report bugs at \url{https://github.com/epiforecasts/scoringutils/issues}
}

}
\author{
\strong{Maintainer}: Nikos Bosse \email{nikosbosse@gmail.com} (\href{https://orcid.org/0000-0002-7750-5280}{ORCID})

Authors:
\itemize{
  \item Sam Abbott \email{contact@samabbott.co.uk} (\href{https://orcid.org/0000-0001-8057-8037}{ORCID})
  \item Hugo Gruson \email{hugo.gruson+R@normalesup.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID})
  \item Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}

Other contributors:
\itemize{
  \item Johannes Bracher \email{johannes.bracher@kit.edu} (\href{https://orcid.org/0000-0002-3777-1410}{ORCID}) [contributor]
  \item Toshiaki Asakura \email{toshiaki.asa9ra@gmail.com} (\href{https://orcid.org/0000-0001-8838-785X}{ORCID}) [contributor]
  \item James Mba Azam \email{james.azam@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5782-7330}{ORCID}) [contributor]
  \item Michael Chirico \email{michaelchirico4@gmail.com} (\href{https://orcid.org/0000-0003-0787-087X}{ORCID}) [contributor]
}

}
\keyword{internal}
