% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-scoring-rules.R
\name{select_rules}
\alias{select_rules}
\title{Select Scoring Rules From A List of Possibilities}
\usage{
select_rules(possibilities, select, exclude = NULL)
}
\arguments{
\item{possibilities}{A list of scoring rules.}

\item{select}{A character vector of scoring rules to select from the list.
If \code{select = all} (the default), all possible scoring rules are returned.}

\item{exclude}{A character vector of scoring rules to exclude from the list.
If \code{select != "all"}, this argument is ignored.}
}
\value{
A list of scoring rules.
}
\description{
Helper function to return only the scoring rules selected by
the user from a list of possible scoring rules.
}
\examples{
scoringutils:::select_rules(
  possibilities = rules_binary(),
  select = "brier_score"
)
}
\keyword{internal}
