% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-forecast-sample-multivariate.R
\name{set_grouping}
\alias{set_grouping}
\title{Set grouping}
\usage{
set_grouping(data, grouping)
}
\arguments{
\item{data}{A data.frame (or similar) with predicted and observed values.
See the details section of for additional information
on the required input format.}

\item{grouping}{Character vector with the names of the columns that
define the grouping.}
}
\value{
A data.table with an additional column \code{.scoringutils_group_id} that
contains the group id for each row.
}
\description{
Helper function to set the grouping of a forecast.
}
\keyword{internal}
