% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-forecast-sample-multivariate.R
\name{set_grouping}
\alias{set_grouping}
\title{Set grouping}
\usage{
set_grouping(data, by)
}
\arguments{
\item{data}{A data.frame (or similar) with predicted and observed values.
See the details section of for additional information
on the required input format.}

\item{by}{Character vector with the names of the columns that define the
grouping. The grouping variable is set analogously to the forecast unit: by
listing all columns that, together, uniquely define a single group - meaning
that a unique combination of values in those columns is associated with a
single group. You may, however, find it more intuitive to group "across"
certain columns. You can do this using \code{\link[=setdiff]{setdiff()}} and \code{\link[=get_forecast_unit]{get_forecast_unit()}}.
For example, if you want to make all forecasts for the same location part of
the same group, you could do something like
\code{by = setdiff(get_forecast_unit(data, "location"))}.}
}
\value{
A data.table with an additional column \code{.scoringutils_group_id} that
contains the group id for each row.
}
\description{
Helper function to set the grouping of a forecast.
}
\keyword{internal}
