% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_probabilstic_forecasts_integer.R
\name{sharpness}
\alias{sharpness}
\title{Determines sharpness of an incidence forecast as the width of the prediction interval}
\usage{
sharpness(samples)
}
\arguments{
\item{samples}{nxN matrix of predictive samples, n (number of rows) being
the number of data points and N (number of columns) the
number of Monte Carlo samples}
}
\value{
data frame with sharpness for each interval by date
}
\description{
explanation missing
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
export
}
