% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_scores.R
\name{summarise_scores}
\alias{summarise_scores}
\alias{summarize_scores}
\title{Summarise scores as produced by \code{\link[=score]{score()}}}
\usage{
summarise_scores(scores, by = "model", across = NULL, fun = mean, ...)

summarize_scores(scores, by = "model", across = NULL, fun = mean, ...)
}
\arguments{
\item{scores}{A data.table of scores as produced by \code{\link[=score]{score()}}.}

\item{by}{character vector with column names to summarise scores by. Default
is \code{model}, meaning that there will be one score per model in the output.
The \emph{unit of a single forecast} is determined by the columns present in the
input data that do not correspond to a metric produced by \code{\link[=score]{score()}}, which
indicate indicate a grouping of forecasts (for example there may be one
forecast per day, location and model). Adding additional, unrelated, columns
may alter results in an unpredictable way.}

\item{across}{character vector with column names from the vector of variables
that define the \emph{unit of a single forecast} (see above) to summarise scores
across (meaning that the specified columns will be dropped). This is an
alternative to specifying \code{by} directly. If \code{NULL} (default), then \code{by} will
be used or inferred internally if also not specified. Only  one of \code{across}
and \code{by}  may be used at a time.}

\item{fun}{a function used for summarising scores. Default is \code{mean}.}

\item{...}{additional parameters that can be passed to the summary function
provided to \code{fun}. For more information see the documentation of the
respective function.}
}
\value{
a data.table with summarised scores. Scores are summarised according
to the names of the columns of the original data specified in \code{by} or
\code{across} using the \code{fun} passed to \code{summarise_scores()}.
}
\description{
Summarise scores as produced by \code{\link[=score]{score()}}
}
\examples{
\dontshow{
  data.table::setDTthreads(2) # restricts number of cores used on CRAN
}
library(magrittr) # pipe operator
\dontrun{
scores <- score(example_continuous)
}

# get scores by model
summarise_scores(scores,by = "model")

# get scores by model and target type
summarise_scores(scores, by = c("model", "target_type"))

# Get scores summarised across horizon, forecast date, and target end date
summarise_scores(
 scores, across = c("horizon", "forecast_date", "target_end_date")
)

# get standard deviation
summarise_scores(scores, by = "model", fun = sd)

# round digits
summarise_scores(scores,by = "model") \%>\%
  summarise_scores(fun = signif, digits = 2)
}
\keyword{scoring}
