% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_forecasts.R
\name{summarise_scores}
\alias{summarise_scores}
\title{Summarise scores as produced by \code{\link[=score]{score()}}}
\usage{
summarise_scores(scores, summarise_by, quantiles = c(), sd = FALSE)
}
\arguments{
\item{scores}{a data.table of unsummarised scores as produced by
\code{\link[=score]{score()}}, when running it with the option \code{summarise_by = NULL}}

\item{summarise_by}{character vector of columns to group the summary by. By
default, this is equal to \code{by} and no summary takes place.
But sometimes you may want to to summarise
over categories different from the scoring.
\code{summarise_by} is also the grouping level used to compute
(and possibly plot) the probability integral transform(pit). Sometimes you
may want to include 'range', 'quantile' or 'sample', to summarise by
range, quantile or sample.}

\item{quantiles}{numeric vector of quantiles to be returned when summarising.
Instead of just returning a mean, quantiles will be returned for the
groups specified through \code{summarise_by}. By default, no quantiles are
returned.}

\item{sd}{if \code{TRUE} (the default is \code{FALSE}) the standard deviation of all
metrics will be returned when summarising.}
}
\description{
Summarise scores as produced by \code{\link[=score]{score()}}-
}
\examples{
library(scoringutils)
data <- example_quantile
scores <- score(data)
summarise_scores(scores,
                 summarise_by = c("model"))

}
