% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-input-helpers.R
\name{test_columns_not_present}
\alias{test_columns_not_present}
\title{Test whether column names are NOT present in a data.frame}
\usage{
test_columns_not_present(data, columns)
}
\arguments{
\item{data}{A data.frame or similar to be checked}

\item{columns}{A character vector of column names to check}
}
\value{
Returns TRUE if none of the columns are present and FALSE otherwise
}
\description{
The function checks whether all column names are NOT present.
If none of the columns are present, the function returns TRUE. If one or
more columns are present, the function returns FALSE.
}
\keyword{internal_input_check}
