% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{theme_scoringutils}
\alias{theme_scoringutils}
\title{Scoringutils ggplot2 theme}
\usage{
theme_scoringutils()
}
\value{
A ggplot2 theme
}
\description{
A theme for ggplot2 plots used in scoringutils
}
\keyword{plotting}
