% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience-functions.R
\name{transform_forecasts}
\alias{transform_forecasts}
\title{Add transformations}
\usage{
transform_forecasts(
  data,
  transform = c("log", "sqrt"),
  fun = NULL,
  add = TRUE,
  label = transform,
  ...
)
}
\arguments{
\item{data}{A data.frame or data.table with the predictions and observations.
For scoring using \code{\link[=score]{score()}}, the following columns need to be present:
\itemize{
\item \code{true_value} - the true observed values
\item \code{prediction} - predictions or predictive samples for one
true value. (You only don't need to provide a prediction column if
you want to score quantile forecasts in a wide range format.)}
For scoring integer and continuous forecasts a \code{sample} column is needed:
\itemize{
\item \code{sample} - an index to identify the predictive samples in the
prediction column generated by one model for one true value. Only
necessary for continuous and integer forecasts, not for
binary predictions.}
For scoring predictions in a quantile-format forecast you should provide
a column called \code{quantile}:
\itemize{
\item \code{quantile}: quantile to which the prediction corresponds
}

In addition a \code{model} column is suggested and if not present this will be
flagged and added to the input data with all forecasts assigned as an
"unspecified model").

You can check the format of your data using \code{\link[=check_forecasts]{check_forecasts()}} and there
are examples for each format (\link{example_quantile}, \link{example_continuous},
\link{example_integer}, and \link{example_binary}).}

\item{fun}{A function used to transform both true values and predictions}

\item{label}{A string for the newly created 'scale' column to denote the
newly transformed values.}

\item{...}{additional parameters passed down to \code{\link[=score_quantile]{score_quantile()}} (internal
function used for scoring forecasts in a quantile-based format).}
}
\value{
A data.table that includes the original data as well as a
transformation of the original data. There will be one additional column,
'scale', present which will be set to "natural" for the untransformed
forecasts.
}
\description{
Add transformations of the forecasts and observations for
later scoring. For more information on why this might be desirable, check
out the linked reference.
}
\examples{
add_transformation(example_quantile)
}
\references{
Transformation of forecasts for evaluating predictive
performance in an epidemiological context
Nikos I. Bosse, Sam Abbott, Anne Cori, Edwin van Leeuwen, Johannes Bracher,
Sebastian Funk
medRxiv 2023.01.23.23284722
\doi{https://doi.org/10.1101/2023.01.23.23284722}
\url{https://www.medrxiv.org/content/10.1101/2023.01.23.23284722v1} # nolint
}
\author{
Nikos Bosse \email{nikosbosse@gmail.com}
}
\keyword{check-forecasts}
