% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\alias{validate.default}
\alias{validate.scoringutils_binary}
\alias{validate.scoringutils_point}
\alias{validate.scoringutils_quantile}
\alias{validate.scoringutils_sample}
\title{Validate input data}
\usage{
validate(data, ...)

\method{validate}{default}(data, ...)

\method{validate}{scoringutils_binary}(data, ...)

\method{validate}{scoringutils_point}(data, ...)

\method{validate}{scoringutils_quantile}(data, ...)

\method{validate}{scoringutils_sample}(data, ...)
}
\arguments{
\item{data}{A data.frame or data.table with the following columns:
\itemize{
\item \code{observed} - observed values
\item \code{predicted} - predictions, predictive samples or predictive quantiles
\item \code{model} - name of the model or forecaster who made a prediction
}

Depending on the forecast type, one of the following columns may be required:
\itemize{
\item \code{sample_id} - index for the predictive samples in the 'predicted' column
\item \code{quantile}: quantile-level of the corresponding value in \code{predicted}
}

For more information see the vignettes and the example data
(\link{example_quantile}, \link{example_continuous},
\link{example_integer}, \code{\link[=example_point]{example_point()}}, and \link{example_binary}).}

\item{...}{additional parameters passed down to other functions.}
}
\value{
Depending on the forecast type, an object of class
\code{scoringutils_binary}, \code{scoringutils_point}, \code{scoringutils_sample} or
\code{scoringutils_quantile}.
}
\description{
The default method, \code{\link[=validate.default]{validate.default()}} only runs basic input checks.
It's main purpose is to determine the forecast type (binary, point,
sample-based or quantile-based) based on the input data. It then constructs
an appropriate class and calls \code{\link[=validate]{validate()}} again which dispatches to the
appropriate method.

Methods for the different classes run \code{\link[=validate_general]{validate_general()}}, which performs
checks that are the same for all forecast types and then perform specific
checks for the specific forecast type.

You can find more information about input formats in the vignette.
To summarise, the data should come in one of four different formats:
\itemize{
\item A format for binary predictions (see \link{example_binary})
\item A format for point predictions (see \link{example_point})
\item A sample-based format for discrete or continuous predictions
(see \link{example_continuous} and \link{example_integer})
\item A quantile-based format (see \link{example_quantile})
}
}
\keyword{validate}
