% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_forecast}
\alias{validate_forecast}
\alias{validate_forecast.forecast_quantile}
\alias{validate_forecast.forecast_sample}
\title{Validate input data}
\usage{
validate_forecast(data, ...)

\method{validate_forecast}{forecast_quantile}(data, ...)

\method{validate_forecast}{forecast_sample}(data, ...)
}
\arguments{
\item{data}{A data.frame or data.table with predicted and observed values.}

\item{...}{additional arguments}
}
\value{
Depending on the forecast type, an object of class
\code{forecast_binary}, \code{forecast_point}, \code{forecast_sample} or
\code{forecast_quantile}.
}
\description{
Methods for the different classes run \code{\link[=validate_general]{validate_general()}}, which performs
checks that are the same for all forecast types and then perform specific
checks for the specific forecast type.
}
\section{Forecast types and input format}{
Various different forecast types / forecast formats are supported. At the
moment, those are
\itemize{
\item point forecasts
\item binary forecasts ("soft binary classification")
\item Probabilistic forecasts in a quantile-based format (a forecast is
represented as a set of predictive quantiles)
\item Probabilistic forecasts in a sample-based format (a forecast is represented
as a set of predictive samples)
}

Forecast types are determined based on the columns present in the input data.

\emph{Point forecasts} require a column \code{observed} of type numeric and a column
\code{predicted} of type numeric.

\emph{Binary forecasts} require a column \code{observed} of type factor with exactly
two levels and a column \code{predicted} of type numeric with probabilities,
corresponding to the probability that \code{observed} is equal to the second
factor level. See details \link[=brier_score]{here} for more information.

\emph{Quantile-based forecasts} require a column \code{observed} of type numeric,
a column \code{predicted} of type numeric, and a column \code{quantile_level} of type
numeric with quantile-levels (between 0 and 1).

\emph{Sample-based forecasts} require a column \code{observed} of type numeric,
a column \code{predicted} of type numeric, and a column \code{sample_id} of type
numeric with sample indices.

For more information see the vignettes and the example data
(\link{example_quantile}, \link{example_continuous}, \link{example_integer},
\code{\link[=example_point]{example_point()}}, and \link{example_binary}).
}

\examples{
forecast <- as_forecast(example_binary)
validate_forecast(forecast)
}
\keyword{check-forecasts}
