% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_general}
\alias{validate_general}
\title{Apply scoringutls input checks that are the same across forecast types}
\usage{
validate_general(data)
}
\arguments{
\item{data}{A data.frame or data.table with the following columns:
\itemize{
\item \code{observed} - observed values
\item \code{predicted} - predictions, predictive samples or predictive quantiles
\item \code{model} - name of the model or forecaster who made a prediction
}

Depending on the forecast type, one of the following columns may be required:
\itemize{
\item \code{sample_id} - index for the predictive samples in the 'predicted' column
\item \code{quantile}: quantile-level of the corresponding value in \code{predicted}
}

For more information see the vignettes and the example data
(\link{example_quantile}, \link{example_continuous},
\link{example_integer}, \code{\link[=example_point]{example_point()}}, and \link{example_binary}).}
}
\value{
returns the input, with a few new attributes that hold additional
information, messages and warnings
}
\description{
The function runs input checks that apply to all input data, regardless of
forecast type. The function
\itemize{
\item asserts that the data is a data.table which has columns \code{observed} and
\code{predicted}, as well as a column called \code{model}.
\item checks the forecast type and forecast unit
\item checks there are no duplicate forecasts
\item if appropriate, checks the number of samples / quantiles is the same
for all forecasts
}
}
\keyword{validate}
