% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-multivariate-point.R
\name{variogram_score_multivariate_point}
\alias{variogram_score_multivariate_point}
\title{Variogram score for multivariate point forecasts}
\usage{
variogram_score_multivariate_point(
  observed,
  predicted,
  mv_group_id,
  w_vs = NULL,
  p = 0.5
)
}
\arguments{
\item{observed}{A vector with observed values of size n}

\item{predicted}{nxN matrix of predictive samples, n (number of rows) being
the number of data points and N (number of columns) the number of Monte
Carlo samples. Alternatively, if n = 1, \code{predicted} can just be a vector
of size n.}

\item{mv_group_id}{Numeric vector of length n with ids indicating the
grouping of predicted values. Conceptually, each row of the \code{predicted}
matrix \emph{could} be seen as a separate (univariate) forecast.
The grouping id then groups several of those forecasts together, treating
them as a single multivariate forecast.}

\item{w_vs}{Numeric matrix of weights for the variogram score.
See \code{\link[scoringRules:scores_sample_multiv]{scoringRules::vs_sample()}} for details.}

\item{p}{Numeric, order of the variogram score.
Defaults to 0.5. See \code{\link[scoringRules:scores_sample_multiv]{scoringRules::vs_sample()}} for details.}
}
\description{
Compute the variogram score
(see \link[scoringRules:vs_sample]{scoringRules::vs_sample})
for each group defined by \code{mv_group_id}, treating each point
forecast as a single-sample ensemble.
}
\keyword{metric}
