% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-quantile.R
\name{wis}
\alias{wis}
\title{Weighted Interval Score}
\usage{
wis(
  observed,
  predicted,
  quantile,
  separate_results = FALSE,
  weigh = TRUE,
  count_median_twice = FALSE,
  na.rm = TRUE
)
}
\description{
Weighted Interval Score
}
